### 2025-09-05 dfb644a
fix minor_radius
