### 2025-09-09 2102c34
fix style inconsistencies in quantities
