+++ python/gvec/quantities.py
@@ -62,14 +62,14 @@ def xyz(ds: xr.Dataset):
 # === profiles ========================================================================= #
 
 
-def _profile(var, evalvar, deriv, long_name, symbol):
+def _profile_factory(var, evalvar, deriv, long_name, symbol):
     """Factory function for profile quantities."""
 
     @register(
         quantities=var,
         attrs=dict(long_name=long_name, symbol=symbol),
     )
-    def profile(ds: xr.Dataset, state: State):
+    def _profile(ds: xr.Dataset, state: State):
         if "rho" not in ds:
             raise KeyError("Evaluation of profiles requires the radial coordinate 'rho'.")
         if ds.rho.dims == ("rad",):
@@ -79,7 +79,7 @@ def _profile(var, evalvar, deriv, long_name, symbol):
             output = state.evaluate_profile(evalvar, rho, deriv=deriv)
             ds[var] = (ds.rho.dims, output.reshape(ds.rho.shape))
 
-    return profile
+    return _profile
 
 
 for var, name, symbol in [
@@ -88,11 +88,11 @@ for var, name, symbol in [
     ("chi", "poloidal magnetic flux", r"\chi"),
     ("Phi", "toroidal magnetic flux", r"\Phi"),
 ]:
-    globals()[var] = _profile(var, var, 0, name, symbol)
-    globals()[f"d{var}_dr"] = _profile(
+    globals()[var] = _profile_factory(var, var, 0, name, symbol)
+    globals()[f"d{var}_dr"] = _profile_factory(
         f"d{var}_dr", var, 1, f"{name} gradient", f"\\frac{{d{symbol}}}{{d\\rho}}"
     )
-    globals()[f"d{var}_drr"] = _profile(
+    globals()[f"d{var}_drr"] = _profile_factory(
         f"d{var}_drr", var, 2, f"{name} curvature", f"\\frac{{d^2{symbol}}}{{d\\rho^2}}"
     )
 
@@ -107,7 +107,7 @@ def Phi_edge(ds: xr.Dataset, state: State):
 # === base ============================================================================= #
 
 
-def _base(var, long_name, symbol):
+def _base_factory(var, long_name, symbol):
     """Factory function for base quantities."""
 
     @register(
@@ -121,7 +121,7 @@ def _base(var, long_name, symbol):
             for i in ("r", "t", "z", "rr", "rt", "rz", "tt", "tz", "zz")
         },
     )
-    def base(ds: xr.Dataset, state: State):
+    def _base(ds: xr.Dataset, state: State):
         if "rho" not in ds or "theta" not in ds or "zeta" not in ds:
             raise KeyError(
                 "Evaluation of base variables requires 'rho', 'theta', 'zeta' to be defined."
@@ -130,7 +130,7 @@ def _base(var, long_name, symbol):
         # mesh in logical coordinates (rho, theta, zeta) -> rho(rad), theta(pol), zeta(tor)
         if ds.rho.dims == ("rad",) and ds.theta.dims == ("pol",) and ds.zeta.dims == ("tor",):
             outputs = state.evaluate_base_tens_all(var, ds.rho, ds.theta, ds.zeta)
[…]
