### 2025-09-10 #8909 @fhindenlang
Regarding the historical joss papers found by the editorial bot, only "SIMSOPT: A flexible framework for stellarator optimization" is related to this submission and is already cited.
