### 2025-09-10 4b65473
add initialisation stage to gvec.run
* the Run class now always creates an initialisation stage which can contain init_LA, boundary_perturb and VMEC readin. These parameters are then disabled in the global fallback parameters.
* this allows using boundary perturbations with a prescribed current profile and removes any special handling of init_LA
* new _run_stage_target_force method (used if picard_current="off")
* new util.compute_boundary_perturbation method (wip)
