+++ python/gvec/core/run.py
@@ -24,7 +24,7 @@ from gvec.lib import modgvec_py_binding as _binding
 
 def run(
     parameters: Mapping | Path | str,
-    restartstate: Path | None = None,
+    restartstate: Path | State | None = None,
     runpath: Path | str | None = None,
     redirect_gvec_stdout: bool = True,
     quiet: bool = False,
@@ -38,7 +38,7 @@ def run(
     ----------
     parameters : Mapping | Path | str
         GVEC parameter dictionary or parameter file.
-    restartstate : Path | None, optional
+    restartstate : Path | State | None, optional
         Path to restart file or restart State object, by default None
     runpath : Path | str | None, optional
         Path to the directory where GVEC is executed. Overwrites existing directories or cerates a new directory if it does not exist.
@@ -139,14 +139,12 @@ class Run:
                 f"Neither valid 'params' nor a valid 'state' are provided! params: {params}, state: {state}"
             )
 
+        # params: user provided parameters
+        # self.parameters: global/fallback parameters
         # _state_parameters reflect the state of the Run, that is they change during each stage and/or restart
-        # after a stage they are reset to parameters.They are NOT state.parameters, as those reflect the state of a
-        # finished restart.
-        self._state_parameters = (
-            self.parameters.copy()
-        )  # parameters that will be changed during each restart
-        if "ProjectName" not in self._state_parameters:
-            self._state_parameters["ProjectName"] = "GVEC"
+        #   after a stage they are reset to parameters. They are NOT state.parameters, as those reflect the state of a
+        #   finished restart.
+        if "ProjectName" not in self.parameters:
             self.parameters["ProjectName"] = "GVEC"
         if "picard_current" not in self.parameters:
             self.parameters["picard_current"] = "off"
@@ -154,7 +152,7 @@ class Run:
         self.logger = logging.getLogger("gvec.run")
         self.filetype = parameter_format
 
-        project_dir = Path(f"{self._state_parameters['ProjectName']}_gvec_stages")
+        project_dir = Path(f"{self.parameters['ProjectName']}_gvec_stages")
         if project_dir.exists():
             self.logger.debug(f"Removing existing run directory {project_dir}")
             shutil.rmtree(project_dir)
@@ -170,38 +168,58 @@ class Run:
         self.GVEC_iter_used = 0
         self.redirect_gvec_stdout = redirect_gvec_stdout
 
-        self.totaliter = params.get("totaliter", int(1e5))
-        if "maxIter" not in params:
-            self._state_parameters["maxIter"] = self.totaliter
+        self.totaliter = self.parameters.get("totaliter", int(1e5))
+        if "maxIter" not in self.parameters:
+            self.parameters["maxIter"] = self.totaliter
 
-        self.has_Itor = "I_tor" in params
-
-        picard_current = params.get("picard_current", "off")
-        if self.has_Itor and picard_current == "off":
+        picard_current = self.parameters.get("picard_current", "off")
+        if "I_tor" in self.parameters and picard_current != "off":
+            self.curr_constraint = True
+        elif picard_current == "off":
             raise KeyError(
                 "'I_tor' is provided but 'picard_current' is set to 'off' or not provided. Please provide a valid 'picard_current', e.g. 'auto'."
             )
-        if not self.has_Itor and picard_current != "off":
+        elif "I_tor" not in self.parameters:
             raise KeyError(
                 "Expected 'I_tor' in the parameters since 'picard_current' is not 'off'."
                 + " Please set 'picard_current' to 'off' if you want to use a fixed 'iota' profile or provide 'I_tor'."
             )
+        else:
[…]
