### 2025-09-11 2160838
cas3d: quantities docstrings
