### 2025-09-11 0b729db
add cas3d reparam test
