### 2025-09-11 2a54514
fixed initialization of initial stage
