+++ test-CI/unit-pygvec/test_scripts.py
@@ -201,12 +201,13 @@ def test_stages_without_current():
         {"minimize_tol": 1e-3, "sgrid": {"nelems": 3}},
     ]
     run_with_stages = gvec.run(parameters, keep_intermediates="all")
+    final_rundir = run_with_stages.rundir
     assert Path(f"{ProjectName}_State_final.dat").exists()
     assert Path(f"parameter_{ProjectName}_final.ini").exists()
     assert Path(f"{ProjectName}_gvec_stages").is_dir()
-    assert Path(f"{ProjectName}_gvec_stages/0-00").is_dir()
-    assert Path(f"{ProjectName}_gvec_stages/0-00/stdout.txt").exists()
-    with open(f"{ProjectName}_gvec_stages/0-00/stdout.txt", "r") as file:
+    assert final_rundir.is_dir()
+    assert (final_rundir / "stdout.txt").exists()
+    with open(final_rundir / "stdout.txt", "r") as file:
         stdout = file.read()
     assert "GVEC SUCESSFULLY FINISHED" in stdout
 
