+++ python/gvec/core/run.py
@@ -173,13 +173,14 @@ class Run:
             self.parameters["maxIter"] = self.totaliter
 
         picard_current = self.parameters.get("picard_current", "off")
+
         if "I_tor" in self.parameters and picard_current != "off":
             self.curr_constraint = True
-        elif picard_current == "off":
+        elif "I_tor" in self.parameters and picard_current == "off":
             raise KeyError(
                 "'I_tor' is provided but 'picard_current' is set to 'off' or not provided. Please provide a valid 'picard_current', e.g. 'auto'."
             )
-        elif "I_tor" not in self.parameters:
+        elif "I_tor" not in self.parameters and picard_current != "off":
             raise KeyError(
                 "Expected 'I_tor' in the parameters since 'picard_current' is not 'off'."
                 + " Please set 'picard_current' to 'off' if you want to use a fixed 'iota' profile or provide 'I_tor'."
@@ -208,6 +209,7 @@ class Run:
                 self.stages[0]["whichInitEquilibrium"] = 1
                 self.parameters["whichInitEquilibrium"] = 0
 
+            # add autogenerated stages
             if picard_current == "auto":
                 self.logger.info("Using `picard_current` automatic mode. Generating stages ...")
 
@@ -215,6 +217,8 @@ class Run:
                 self.stages += auto_generate_stages(minimize_tol, 1e-10)
                 self.parameters["picard_current"] = cidict()
                 self.stages[0]["picard_current"] = "off"
+            else:  # ensure at least one stage beyond initial stage
+                self.stages += [cidict()]
 
             parameters_stages = self.parameters.copy()
             parameters_stages["stages"] = self.stages
@@ -612,7 +616,13 @@ class Run:
 
         final_iota_str = ""
         if self.curr_constraint:
-            final_iota_str += f"\n and rms Δiota = {self.rms_iota.item():.2e} (iota_tol={self._state_parameters['picard_current']['iota_tol']:.2e})"
+            final_iota_str += f"\n and rms Δiota = {self.rms_iota.item():.2e}"
+            # in case current constrained was turned off during the final stage
+            if self._state_parameters["picard_current"] != "off":
+                final_iota_str += (
+                    f"(iota_tol={self._state_parameters['picard_current']['iota_tol']:.2e})"
+                )
+
         final_message = (
             f"GVEC finished after {end_time - start_time:5.1f} seconds",
             f" using {self.GVEC_iter_used} iterations (totalIter = {self.totaliter})",
