### 2025-09-12 f2c655c
update DOIs
