+++ JOSS_paper/paper.bib
@@ -111,7 +111,8 @@
   institution  = {Princeton Univ., N.J. Project Matterhorn},
   place        = {United States},
   year         = {1953},
-  month        = {05}
+  month        = {05},
+  doi          = {10.2172/4285864}
 }
 
 
@@ -213,7 +214,8 @@
   number={6},
   pages={905870604},
   year={2021},
-  publisher={Cambridge University Press}
+  publisher={Cambridge University Press},
+  doi={10.1017/S0022377821001082}
 }
 
 %BOUT++
@@ -456,7 +458,8 @@ keywords = {Plasma simulation, Finite element exterior calculus, Particle-in-cel
   date = {2015-01-01},
   booktitle = {IDIHOM: Industrialization of High-Order Methods-A Top-Down Approach},
   pages = {133--152},
-  publisher = {Springer}
+  publisher = {Springer},
+  doi = {10.1007/978-3-319-12886-3_8}
 }
 
 @article{holeSteppedPressureProfile2006,
