### 2025-09-15 60bf5de
add union operator to CaseInsensitiveDict
