### 2025-09-15 c65623c
fix float parameter conversion
Floats like '1500.' were not parsed as floats and therefore written as strings to yaml/toml. Did not affect any output
