### 2025-09-15 3dc7522
use warnings module in core.run
