### 2025-09-15 c8b3235
automatic boundary perturbation for restarts
* The gvec.Run object, as created by gvec.run now automatically sets the boundary perturbation and iota profile if appropriate and restarting from another state.
* As this requires the parameters for the previous state a new positional argument for it was added to the command line 'pygvec run'.
