### 2025-09-15 4032f44
fix perturbed restart
