### 2025-09-22 25fad99
fix derivatives of mod_B
