+++ test-CI/unit-pygvec/test_compute.py
@@ -27,9 +27,18 @@ except ImportError:
 # === FIXTURES === #
 
 
-@pytest.fixture()
+@pytest.fixture(scope="session")
 def teststate(testfiles):
     return State(*testfiles)
+
+
+@pytest.fixture(
+    params=[False, True],
+    ids=["surf", "drho"],
+    scope="session",
+)
+def radial_derivative(request):
+    return request.param
 
 
 @pytest.fixture()
@@ -175,6 +184,32 @@ def ev_profile(request, ev_r_only, ev_rtz, ev_rtz_mixed, ev_rtz_1d):
         return ev_rtz_1d
 
 
+@pytest.fixture(scope="session")
+def ev_boozer(teststate, radial_derivative):
+    ds = EvaluationsBoozer(
+        rho=[0.5, 1.0],
+        theta_B=20,
+        zeta_B=18,
+        state=teststate,
+        radial_derivative=radial_derivative,
+        MNfactor=5,
+    )
+    return ds
+
+
+@pytest.fixture(scope="session")
+def ev_boozer_drho(teststate):
+    ds = EvaluationsBoozer(
+        rho=[0.5, 1.0],
+        theta_B=20,
+        zeta_B=18,
+        state=teststate,
+        radial_derivative=True,
+        MNfactor=5,
+    )
+    return ds
+
+
 # === TESTS === #
 
 
@@ -225,7 +260,6 @@ def test_evaluations_init(teststate):
     [2, 1.0, xr.DataArray([0.5, 0.6], dims="tor"), np.array([0.5, 0.6]), [0.5, 0.6]],
     ids=["int", "float", "xr", "np", "list"],
 )
-@pytest.mark.parametrize("radial_derivative", [True, False])
 def test_boozer_init(teststate, rho, theta_B, zeta_B, radial_derivative):
     ds = EvaluationsBoozer(
         rho, theta_B, zeta_B, teststate, radial_derivative=radial_derivative, MNfactor=1
@@ -233,11 +267,8 @@ def test_boozer_init(teststate, rho, theta_B, zeta_B, radial_derivative):
     teststate.compute(ds, "X1")
 
 
-@pytest.mark.parametrize("radial_derivative", [True, False])
-def test_boozer(teststate, radial_derivative):
-    ds = EvaluationsBoozer(
-        [0.5, 1.0], 20, 18, teststate, radial_derivative=radial_derivative, MNfactor=5
-    )
+def test_boozer(teststate, ev_boozer, radial_derivative):
+    ds = ev_boozer.copy()
     assert np.allclose(ds.rho, [0.5, 1.0])
     assert {"rho", "theta_B", "zeta_B"} == set(ds.coords)
     assert {"rad", "pol", "tor"} == set(ds.dims)
@@ -304,7 +335,6 @@ def test_EvaluationsBoozer_2D(teststate):
         assert "long_name" in ds[var].attrs, f"no long_name defined in {var} attributes"
 
 
-@pytest.mark.parametrize("radial_derivative", [True, False])
 def test_EvaluationsBoozer_fieldlines(teststate, radial_derivative):
     """Test EvaluationsBoozer with a 3D array for theta_B, as required for fieldline coordinates."""
[…]
