### 2025-09-29 6903632
fix documentation, magnetic field contravariant components now correct.
