### 2025-09-29 c9f9e62
fix documentation, in equation for lambda.
