### 2025-09-29 #8909 @danielskatz
Just looking at this GitLab instance, where I see https://gitlab.mpcdf.mpg.de/users/sign_in?redirect_to_referer=yes which indicates that "anyone can interact with the software, submit issues, etc." is false here...
