### 2025-09-30 0bcb9ed
improve vmec parameter conversion
additionally add utils for evaluating the boundary and axis as specified in a parameter file
