+++ python/gvec/util.py
@@ -432,6 +432,73 @@ def effective_minor_radius(
     return np.sqrt(np.mean(areas) / np.pi)
 
 
+def evaluate_boundary(
+    theta: np.ndarray, zeta: np.ndarray, parameters: Mapping
+) -> tuple[np.ndarray, np.ndarray]:
+    """Evaluate the boundary at the given (theta, zeta) points.
+
+    Args:
+        theta (1D np.ndarray): The poloidal angles at which to evaluate the boundary.
+        zeta (1D np.ndarray): The toroidal angles at which to evaluate the boundary.
+        parameters (Mapping): The parameters defining the boundary.
+
+    Returns:
+        tuple[2D np.ndarray, 2D np.ndarray]: The (X^1, X^2) coordinates of the boundary at the given (theta, zeta) points.
+    """
+    theta = np.asarray(theta)
+    zeta = np.asarray(zeta)
+    if theta.ndim != 1 or zeta.ndim != 1:
+        raise ValueError("theta and zeta must be 1D arrays")
+    nfp = parameters.get("nfp", 1)
+    theta, zeta = np.meshgrid(theta, zeta, indexing="ij")
+    x1 = np.zeros_like(theta)
+    x2 = np.zeros_like(theta)
+    for (m, n), value in parameters.get("X1_b_cos", {}).items():
+        x1 += value * np.cos(m * theta - n * nfp * zeta)
+    for (m, n), value in parameters.get("X1_b_sin", {}).items():
+        x1 += value * np.sin(m * theta - n * nfp * zeta)
+    for (m, n), value in parameters.get("X2_b_cos", {}).items():
+        x2 += value * np.cos(m * theta - n * nfp * zeta)
+    for (m, n), value in parameters.get("X2_b_sin", {}).items():
+        x2 += value * np.sin(m * theta - n * nfp * zeta)
+    return x1, x2
+
+
+def evaluate_axis(zeta: np.ndarray, parameters: Mapping) -> tuple[np.ndarray, np.ndarray]:
+    """Evaluate the magnetic axis at the given zeta points.
+
+    Args:
+        zeta (1D np.ndarray): The toroidal angles at which to evaluate the axis.
+        parameters (Mapping): The parameters defining the axis.
+
+    Returns:
+        tuple[1D np.ndarray, 1D np.ndarray]: The (X^1, X^2) coordinates of the axis at the given zeta points.
+    """
+    zeta = np.asarray(zeta)
+    if zeta.ndim != 1:
+        raise ValueError("zeta must be a 1D array")
+    nfp = parameters.get("nfp", 1)
+    x1 = np.zeros_like(zeta)
+    x2 = np.zeros_like(zeta)
+    for (m, n), value in parameters.get("X1_a_cos", {}).items():
+        if m != 0:
+            raise ValueError("Axis X1_a_cos should only have m=0 modes")
+        x1 += value * np.cos(-n * nfp * zeta)
+    for (m, n), value in parameters.get("X1_a_sin", {}).items():
+        if m != 0:
+            raise ValueError("Axis X1_a_sin should only have m=0 modes")
+        x1 += value * np.sin(-n * nfp * zeta)
+    for (m, n), value in parameters.get("X2_a_cos", {}).items():
+        if m != 0:
+            raise ValueError("Axis X2_a_cos should only have m=0 modes")
+        x2 += value * np.cos(-n * nfp * zeta)
+    for (m, n), value in parameters.get("X2_a_sin", {}).items():
+        if m != 0:
+            raise ValueError("Axis X2_a_sin should only have m=0 modes")
+        x2 += value * np.sin(-n * nfp * zeta)
+    return x1, x2
+
+
[…]
