+++ docs/dev/derivations.md
@@ -32,3 +32,129 @@ $$
 for $\alpha\in\left\{\rho,\thet,\zeta\right\}$.
 
 $\Rightarrow$ `dmod_B_dr`, `dmod_B_dt`, `dmod_B_dz`, `grad_mod_B`
+
+## Derivatives of $\vec{B}$
+
+$$
+\begin{align}
+\nabla \vec{B} &:= \frac{\partial \vec{B}}{\partial \vec{x}}
+= \sum_\alpha \frac{\partial \vec{B}}{\partial \alpha} \nabla\alpha
+= \frac{\partial \vec{B}}{\partial \rho} \nabla\rho + \frac{\partial \vec{B}}{\partial \thet} \nabla\thet + \frac{\partial \vec{B}}{\partial \zeta} \nabla\zeta \\
+\frac{\partial \vec{B}}{\partial \alpha} &:= \sum_\beta \frac{\partial B^\beta}{\partial \alpha} \vec{e}_\beta + B^\beta \vec{k}_{\beta\alpha} \\
+&= \cancel{\frac{\partial B^\rho}{\partial \alpha}} \erho + \cancel{B^\rho} \vec{k}_{\rho\alpha}
++ \frac{\partial B^\thet}{\partial \alpha} \ethet + B^\thet \vec{k}_{\thet\alpha}
++ \frac{\partial B^\zeta}{\partial \alpha} \ezeta + B^\zeta \vec{k}_{\zeta\alpha} \\
+\nabla \vec{B} &= \sum_{\alpha\beta} \left(\frac{\partial B^\beta}{\partial \alpha} \vec{e}_\beta + B^\beta \vec{k}_{\beta\alpha} \right) \nabla\alpha \\
+\end{align}
+$$
+
+for $\alpha,\beta\in\left\{\rho,\thet,\zeta\right\}$.
+
+Note that $\vec{e}_\beta \nabla\alpha$ and $\vec{k}_{\beta\alpha} \nabla\alpha$ are *outer products* (?) and $\nabla\vec{B} \in  \R^3\times\R^3$ (?).
+
+Due to `xarray`'s limitations with multiple dimensions of the same name, $\nabla\vec{B}$ is not directly available as a computable quantitiy, and its components should be used instead.
+
+$\Rightarrow$ `dB_dr`, `dB_dt`, `dB_dz`
+
+---
+
+$$
+\begin{align}
+\lVert\nabla\vec{B}\rVert^2 &:= \nabla\vec{B} : \nabla\vec{B} \\
+&= \sum_{ij} \sum_{\alpha\beta} \left(\frac{\partial \vec{B}}{\partial \alpha}\right)_i \left(\nabla\alpha\right)_j \left(\frac{\partial \vec{B}}{\partial \beta}\right)_i \left(\nabla\beta\right)_j
+\end{align}
+$$
+???
+
+for $\alpha,\beta\in\left\{\rho,\thet,\zeta\right\}$.
+
+$\Rightarrow$ `grad_B_norm2`
+
+---
+
+From
+
+$$
+\begin{align}
+B^\thet &= \frac{1}{\Jac} \dPhidr \left(\iota-\frac{\partial\lambda}{\partial\zeta} \right) \\
+B^\zeta &= \frac{1}{\Jac} \dPhidr \left(1+\frac{\partial\lambda}{\partial\thet} \right) \\
+\end{align}
+$$
+
+it follows
+
+$$
+\begin{align}
+\frac{\partial B^\thet}{\partial \rho} &=
+    \frac{1}{\Jac} \frac{d^2\Phi}{d\rho^2} \left(\iota-\frac{\partial\lambda}{\partial\zeta} \right)
[…]
