### 2025-09-30 2ea9b97
fix dmod_B_di
