### 2025-09-30 #8909 @fhindenlang
@kyleniemeyer thanks for looking into the specific items, and I am happy to get started!

- Regarding the 3 missing DOI's, they seem to belong to publications that were not cited in the manuscript.
  Should we remove the publications that do not have a citation from the `paper.bib`?
- The citation of `SeLaLib` does not have a doi, since its a reference to the code webpage.
- the editorial bot counts 1222 words in `paper.md` ...
  Starting from the summary, its 1006, and without the acknowledgement section its 823. Hope this is ok.
- Above, I posted a list of potential reviewers
