### 2025-10-01 #8909 @kyleniemeyer
@fhindenlang it isn't necessary to remove the unused citation from the `.bib` file but it will prevent the extraneous warnings.
