### 2025-10-02 c815c91
fix run stages refactoring & boozer tests
