### 2025-10-06 300fb96
Added routines for plotting from GVEC state files. These include: scalar quantities on the axis (plot_on_axis), radial profile of scalar quantity (in rho) (plot_radial_profile), scalar value on a flux surface (given a rho label) (plot_flux_surface), scalar quantity on a poloidal cross-section (plot_poloidal_slice). Two 3D plots using plotly are also available, a 3d surface of a given rho value (plot_3d_surface) and a 3d plot of the GVEC boundary (plot_boundary), which is a wrapper of the first function.
