+++ python/gvec/plotting/plots.py
@@ -1,0 +1,321 @@
+from numpy import array, max, min, ndarray
+
+import matplotlib.pyplot as pyplot
+
+from gvec.core.state import State
+
+from gvec_plotting.utils import _get_coord_range, _get_scalars_for_plotting
+
+
+pyplot.rcParams.update({"text.usetex": True})
+
+
+def _plot_line_quantities_from_dict(
+    plotting_quantities, x_axis_value, subplot_grid, share_axis, xlabel
+):
+    """
+    Plot the quantities from `plotting_quantities`.
+
+    Return both the `matplotlib.pyplot.figure` object and the `array([axes])`.
+    """
+
+    f, axs = pyplot.subplots(*subplot_grid, sharex=share_axis)
+
+    # Ensure this is an array so we can iterate over it
+    if not isinstance(axs, ndarray):
+        axs = array([axs])
+
+    if axs.size == 1:
+        # If there is only one axis we plot all quantities on the single axis
+        ax_lines = [
+            axs[0].plot(x_axis_value, values, label=quantity)
+            for (quantity, values) in plotting_quantities.items()
+        ]
+        axs[0].legend(plotting_quantities.keys())
+    else:
+        # Plot one value per axis
+        for i, quantity in enumerate(plotting_quantities):
+            ax_lines = axs[i].plot(x_axis_value, plotting_quantities[quantity])
+            # Top right of each subplot will have a text box with the quantity being plotted
+            axs[i].annotate(
+                quantity,
+                xy=(1, 1),
+                xycoords="axes fraction",
+                xytext=(-0.6, -0.6),
+                textcoords="offset fontsize",
+                verticalalignment="top",
+                horizontalalignment="right",
+                bbox=dict(facecolor="white", edgecolor="black"),
+            )
+            axs[i].set_xlabel(xlabel)
+
+    return f, axs
+
+
+def plot_poloidal_slice(
+    state: State,
+    nrho: int = 51,
+    ntheta: int = 51,
+    nzeta: int | float | ndarray = 1,
+    quantities: str = "mod_B",
+    subplot_grid: list = [1, 1],
+    share_axis: bool = False,  # TODO: Fix this so that only the minor radius is shared along the x-axis
+    # show_origin = False #TODO
+):
+    """
+    plot_poloidal_slice
+
+    Plot a poloidal slice with some equilibrium quantity on it. Defaults to plotting $\|\mathbf{B}\|$
+
+    Parameters
+    ----------
+    state: GVEC state file
+    nrho: int, optional
+        The radial resolution of the slices. Default is 51
+    ntheta: int, optional
+        The poloidal resolution of the slices. Default is 51
+    nzeta: int, float, ndarray, optional
+        Poloidal slices to plot. Default is 1.
+        - int: Number of equally spaced slices.
+        - float: The specific slice.
+        - ndarray: The specific slices.
+    quantities: str
+        The quantity to plot. Default is "mod_B"
+    subplot_grid: list, conditionally optional
+        The grid shape for the subplots. Default is `[1,1]`. Required if `len(nzeta)>1`.
+    share_axis: bool
+        If true, all subplots will share their `X1` and `X2` axis positions.
+        TODO: Fix this so only the minor radius is shared.
+
+    Returns
+    -------
+    `matplotlib.pyplot.figure` object
+    """
+
+    rho = _get_coord_range("rho", state.nfp, nrho)
+    theta = _get_coord_range("theta", state.nfp, ntheta)
+    zeta = _get_coord_range("zeta", state.nfp, nzeta)
+
[…]
