+++ python/gvec/plotting/__init__.py
@@ -1,0 +1,7 @@
+from gvec_plotting.plots3d import plot_3d_surface, plot_boundary
+from gvec_plotting.plots import (
+    plot_poloidal_slice,
+    plot_flux_surface,
+    plot_radial_profile,
+    plot_on_axis,
+)
