### 2025-10-07 38075b4
Rewrote some things so the plotting can actually be run.
