+++ python/gvec/plotting/__init__.py
@@ -1,5 +1,5 @@
-from gvec_plotting.plots3d import plot_3d_surface, plot_boundary
-from gvec_plotting.plots import (
+from .plots3d import plot_3d_surface, plot_boundary
+from .plots import (
     plot_poloidal_slice,
     plot_flux_surface,
     plot_radial_profile,
