### 2025-10-09 8dc4cb2
remove uncited references
