+++ JOSS_paper/paper.bib
@@ -105,15 +105,6 @@
     doi = {10.1063/5.0020743},
 }
 
-@techreport{osti_4285864,
-  author       = {Goldman, L M and Spitzer, L Jr},
-  title        = {PRELIMINARY EXPERIMENTAL RESULTS WITH THE MODEL A STELLARATOR},
-  institution  = {Princeton Univ., N.J. Project Matterhorn},
-  place        = {United States},
-  year         = {1953},
-  month        = {05},
-  doi          = {10.2172/4285864}
-}
 
 
 @article{vmec_83,
@@ -204,18 +195,6 @@
   year={2020},
   publisher={IOP Publishing},
   doi = {10.1088/1361-6587/aba858}
-}
-
-@article{wilms2021global,
-  title={Global electromagnetic turbulence simulations of {W7-X}-like plasmas with {GENE-3D}},
-  author={Wilms, Felix and Navarro, Alejandro Ba{\~n}{\'o}n and Merlo, Gabriele and Leppin, Leonhard and G{\"o}rler, Tobias and Dannert, Tilman and Hindenlang, Florian and Jenko, Frank},
-  journal={Journal of Plasma Physics},
-  volume={87},
-  number={6},
-  pages={905870604},
-  year={2021},
-  publisher={Cambridge University Press},
-  doi={10.1017/S0022377821001082}
 }
 
 %BOUT++
@@ -449,17 +428,6 @@ keywords = {Plasma simulation, Finite element exterior calculus, Particle-in-cel
   journal={Journal of Plasma Physics},
   year={2024},
   pages={905900303}
-}
-
-@incollection{HOPR,
-  title = {Mesh Curving Techniques for High Order Discontinuous Galerkin Simulations},
-  author = {F. Hindenlang and T. Bolemann and C-D. Munz},
-  year = {2015},
-  date = {2015-01-01},
-  booktitle = {IDIHOM: Industrialization of High-Order Methods-A Top-Down Approach},
-  pages = {133--152},
-  publisher = {Springer},
-  doi = {10.1007/978-3-319-12886-3_8}
 }
 
 @article{holeSteppedPressureProfile2006,
