### 2025-10-10 f42b200
checking chain-rule derivatives against high-order Finite differences.
- add a `compute_FD` function to the gvec utils
- in the `test_compute.py`, replaced the 2nd order FD by an 8th order FD
- absolute tolerance for comparison now 1e-8, scaled with max(|quantity|)
- locally tested on different examples (w7x is chosen for CI)
