+++ test-CI/unit-pygvec/test_compute.py
@@ -18,6 +18,7 @@ try:
         compute,
         volume_integral,
     )
+    from gvec.util import compute_FD
 except ImportError:
     xr = type(pytest)("xarray")
     xr.DataArray = lambda *args, **kwargs: NotImplemented
@@ -466,20 +467,27 @@ def test_compute_Jac_B_consistency(teststate, ev_rtz):
 def test_derivatives(teststate, q):
     ds = gvec.Evaluations(
         rho=np.linspace(0.95, 1, 51),
-        theta=np.linspace(0.0, np.pi / 10, 51),
-        zeta=np.linspace(0.0, np.pi / 10 / teststate.nfp, 51),
+        theta=np.linspace(np.pi / 10, np.pi / 5, 31),
+        zeta=np.linspace(0.0, np.pi / 10 / teststate.nfp, 41),
     )
     compute(ds, q, f"d{q}_dr", f"d{q}_dt", f"d{q}_dz", state=teststate)
 
     assert ds[q].dims == ("rad", "pol", "tor")
-    dQdr = np.gradient(ds[q], ds.rho, axis=0)[1:-1, 1:-1, 1:-1]
-    dQdt = np.gradient(ds[q], ds.theta, axis=1)[1:-1, 1:-1, 1:-1]
-    dQdz = np.gradient(ds[q], ds.zeta, axis=2)[1:-1, 1:-1, 1:-1]
-    ds = ds.isel(rad=slice(1, -1), pol=slice(1, -1), tor=slice(1, -1))
 
-    np.testing.assert_allclose(ds[f"d{q}_dr"], dQdr, atol=1e-3)
-    np.testing.assert_allclose(ds[f"d{q}_dt"], dQdt, atol=1e-3)
-    np.testing.assert_allclose(ds[f"d{q}_dz"], dQdz, atol=1e-3)
+    # using 8th order central FD on the equispaced grid (careful: 4 boundary points cannot be correctly computed)
+    FDpos = [-4, -3, -2, -1, 1, 2, 3, 4]
+    FDcoefs = np.array([1 / 280, -4 / 105, 1 / 5, -4 / 5, 4 / 5, -1 / 5, 4 / 105, -1 / 280])
+    dr = ds.rho.values[1] - ds.rho.values[0]
+    dt = ds.theta.values[1] - ds.theta.values[0]
+    dz = ds.zeta.values[1] - ds.zeta.values[0]
+    dQdr = compute_FD(np.array(ds[q]), FDpos, FDcoefs / dr, axis=0)
+    dQdt = compute_FD(ds[q].values, FDpos, FDcoefs / dt, axis=1)
+    dQdz = compute_FD(ds[q].values, FDpos, FDcoefs / dz, axis=2)
+    ds = ds.isel(rad=slice(4, -4), pol=slice(4, -4), tor=slice(4, -4))
+    Q_scale = np.amax(np.abs(ds[q]))
+    np.testing.assert_allclose(ds[f"d{q}_dr"], dQdr[4:-4, 4:-4, 4:-4], atol=1e-8 * Q_scale)
+    np.testing.assert_allclose(ds[f"d{q}_dt"], dQdt[4:-4, 4:-4, 4:-4], atol=1e-8 * Q_scale)
+    np.testing.assert_allclose(ds[f"d{q}_dz"], dQdz[4:-4, 4:-4, 4:-4], atol=1e-8 * Q_scale)
 
 
 def test_volume_integral(teststate, ev_rtz_int, ev_rtz):
