### 2025-10-13 7be674b
radial derivative added to boozer transform
* we carefully added a Finite Difference radial derivative of lambda and nu and its theta/zeta derivatives
* `LA` and `NU_B` are computed in the boozer transform, however only on a set of specified flux surfaces `rho_i`.
* we recompute lambda and nu at all rho_eps=rho_i-epsilon (epsilon=1e-8), and then evalute `LA` and `NU_B` at the same (theta,zeta) positions,
  which are found for each flux surface by a newton search to match corresponding theta_B,zeta_B coordinates.
* then we use a finite Difference to compute the radial derivatives `dLA_dr,dLA_dr
[…]
