### 2025-10-14 ff238d6
add derivations in the documentation
