### 2025-10-15 12fce73
improve vmec parameter conversion
* proper error if f90nml is not installed
* correctly read start indices for boundary coefficients
