### 2025-10-15 28a3ebe
enable vmec parameter conversion for QUASR
