### 2025-10-15 95c1e6f
fix vmec parameter conversion for curtor=0
