### 2025-10-16 287199c
new interface to create GENE-GIST files
