+++ python/gvec/scripts/gene_gist.py
@@ -1,0 +1,378 @@
+# Copyright (c) 2025 GVEC Contributors, Max Planck Institute for Plasma Physics
+# License: MIT
+"""gene_gist.py - generate GIST files to be used with GENE."""
+
+from typing import Literal
+from collections.abc import Sequence, Mapping
+from pathlib import Path
+import logging
+import argparse
+
+import numpy as np
+import xarray as xr
+import f90nml
+
+import gvec
+
+parser = argparse.ArgumentParser(
+    prog="pygvec-to-gist",
+    description="Produce a GENE-GIST input file from a GVEC state.",
+)
+parser.add_argument(
+    "--rundir",
+    type=Path,
+    help="GVEC run directory",
+    default=Path("."),
+)
+parser.add_argument(
+    "-o",
+    "--outputfile",
+    type=Path,
+    help="output file name (default: '{projectname}_s{s}.gist.txt')",
+    default=None,
+)
+srho = parser.add_mutually_exclusive_group(required=True)
+srho.add_argument(
+    "-s",
+    type=float,
+    help="position of the target flux surface (in normalized toroidal flux, 0 < s <= 1)",
+)
+srho.add_argument(
+    "-r",
+    "--rho",
+    type=float,
+    help="position of the target flux surface (in square root of the normalized toroidal flux, 0 < rho <= 1)",
+)
+parser.add_argument(
+    "--npol",
+    type=float,
+    default=1,
+    help="number of poloidal turns (default 1)",
+)
+parser.add_argument(
+    "--gridpoints",
+    type=int,
+    default=128,
+    help="number of grid points along the fieldline (default 128)",
+)
+parser.add_argument(
+    "--MNfactor",
+    type=int,
+    default=3,
+    help="multiplication factor for the maximum fourier modes for the boozer transform (default 3)",
+)
+parser.add_argument(
+    "-x",
+    "--flip",
+    choices=[None, "pol", "tor", "both"],
+    default=None,
+    help="flip the poloidal or toroidal direction with respect (left-handed Boozer coordinates).",
+)
+parser.add_argument(
+    "-p",
+    "--plot",
+    action="store_true",
+    help="plot the output quantities ('{projectname}_s{s}.gist.png')",
+)
+verbosity = parser.add_mutually_exclusive_group()
+verbosity.add_argument(
+    "-v",
+    "--verbose",
+    action="count",
+    default=0,
+    help="verbosity level: -v for info, -vv for debug",
+)
+verbosity.add_argument("-q", "--quiet", action="store_true", help="suppress output")
+
+# === Main function === #
+
+
+def warning_assert_allclose(logger, name, a, b):
+    try:
+        np.testing.assert_allclose(a, b)
+    except AssertionError as e:
+        logger.warning(f"Assertion for {name} failed{e}")
+
+
+def generate_fieldline_coordinates(state, s0, gridpoints, n_pol, flip, boozer_kwargs={}):
+    rho = np.array([np.sqrt(s0)])  # radial position
+    alpha = np.array([0.0])  # fieldline label
+    # (poloidal) angle along the fieldline
+    theta_B = np.linspace(-np.pi * n_pol, np.pi * n_pol, gridpoints, endpoint=False)
+
+    # flip theta_B: [-pi, pi) -> [pi, -pi)
[…]
