### 2025-10-24 70ef90d
fix FinalizeMHD3D for errors within InitMHD3D
