### 2025-10-24 ce4e3ef
add test for non-existing netCDF file
