### 2025-10-24 1cf4fe8
clean abort messages.
* instead of writing things to screen before an abort, include that information in the abort message, so that its seen when running from python.
* this was especially bad for errors regarding parameter readin, which should be now much better.
* tried to make some abort messages more informative.
