+++ src/globals/readintools.f90
@@ -169,9 +169,8 @@ IF (CntStr.EQ.0) THEN
   IF (PRESENT(Proposal)) THEN
     CntStr=Proposal
   ELSE
-    SWRITE(UNIT_StdOut,*) 'Inifile missing necessary keyword item : ',TRIM(TmpKey)
     CALL abort(__STAMP__, &
-         'Code stopped during inifile parsing!')
+         "missing necessary parameter '"//TRIM(TmpKey)//"'")
   END IF
 END IF
 END FUNCTION CNTSTR
@@ -209,9 +208,8 @@ ELSE
 END IF
 READ(HelpStr,*,IOSTAT=ioerr)GetInt
 IF(ioerr.NE.0)THEN
-  WRITE(UNIT_stdout,*)'PROBLEM IN READIN OF LINE (integer):'
-  WRITE(UNIT_stdout,*) TRIM(key),' = ',TRIM(helpStr)
-  CALL abort(__STAMP__,"")
+  CALL abort(__STAMP__, &
+       "Problem reading parameter '"//TRIM(key)//"', expected integer, got '= "//TRIM(helpStr)//"'")
 END IF
 quiet_def=.FALSE.
 IF(PRESENT(quiet_def_in))THEN
@@ -259,9 +257,8 @@ END IF
 ! Find values of pi in the string
 READ(HelpStr,*,IOSTAT=ioerr)GetReal
 IF(ioerr.NE.0)THEN
-  WRITE(UNIT_stdout,*)'PROBLEM IN READIN OF LINE (real):'
-  WRITE(UNIT_stdout,*) TRIM(key),' = ',TRIM(helpStr)
-  CALL abort(__STAMP__,"")
+  CALL abort(__STAMP__,&
+       "Problem reading parameter '"//TRIM(key)//"', expected real, got '= "//TRIM(helpStr)//"'")
 END IF
 quiet_def=.FALSE.
 IF(PRESENT(quiet_def_in))THEN
@@ -310,7 +307,8 @@ READ(HelpStr,*,IOSTAT=ioerr)GetLogical
 IF(ioerr.NE.0)THEN
   WRITE(UNIT_stdout,*)'PROBLEM IN READIN OF LINE (logical):'
   WRITE(UNIT_stdout,*) TRIM(key),' = ',TRIM(helpStr)
-  CALL abort(__STAMP__,"")
+  CALL abort(__STAMP__, &
+       "Problem reading parameter '"//TRIM(key)//"', expected logical, got '= "//TRIM(helpStr)//"'")
 END IF
 quiet_def=.FALSE.
 IF(PRESENT(quiet_def_in))THEN
@@ -345,7 +343,7 @@ INTEGER                   :: GetIntArray(nIntegers)      !! Integer array read f
 !-----------------------------------------------------------------------------------------------------------------------------------
 ! LOCAL VARIABLES
 CHARACTER(LEN=MAXLEN)           :: HelpStr,ProposalStr
-CHARACTER(LEN=8)                :: DefMsg
+CHARACTER(LEN=8)                :: DefMsg,tmpstr
 INTEGER                         :: iInteger
 INTEGER                         :: ioerr
 LOGICAL                         :: quiet_def
@@ -358,18 +356,17 @@ ELSE
   CALL FindStr(Key,HelpStr,DefMsg)
 END IF
 !count number of components
-iInteger=1+count_sep(helpstr,",")
+iInteger=1+count_sep(Key,helpstr,",")
 
 IF(iInteger.NE.nIntegers)THEN
-  WRITE(UNIT_stdout,'(A,I4,A,I4)')'PROBLEM IN READIN OF LINE (integer Array), number of elements : ', iInteger, ' .NE. ',nIntegers
[…]
