### 2025-10-27 596079a
- Renamed input `equilibrium_quantities` to `quantities` - Changed how scalars are detected by `_get_scalars_for_plotting` -- Now   checks if the maximum of the `shape` is the same as the product of   `shape`, if the object is a scalar then this should be the same (I   think) - Updated `plotting.py` in examples
