+++ python/examples/plotting/plotting.py
@@ -1,0 +1,183 @@
+# %% [markdown]
+#
+# We present some examples for how to use the basic plotting functions in the python interface for GVEC. Note that
+# these are not intended for advanced plotting, but for quickly visualising some basic properties.
+
+# %%
+
+import gvec
+import gvec.plotting as gp
+import numpy as np
+
+# %%
+
+
+major_radius = 5.0
+minor_radius = 0.5
+radius_scale = 1.0
+
+params = {
+    "ProjectName": "test_plotting",
+    "which_hmap": 1,
+    "PhiEdge": 1.0,
+    "iota": {"type": "polynomial", "coefs": [0.625, 0.35]},
+    "pres": {"type": "polynomial", "coefs": [0.0]},
+    "which_hmap": 1,
+    "nfp": 1,
+    "X1_b_cos": {(0, 0): major_radius, (1, 0): minor_radius * radius_scale},
+    "X2_b_sin": {(1, 0): minor_radius * radius_scale},
+    "X1_a_cos": {(0, 0): major_radius},
+    "sgrid_nElems": 2,
+    "X1_mn_max": [3, 1],
+    "X2_mn_max": [3, 1],
+    "LA_mn_max": [3, 1],
+    "X1X2_deg": 3,
+    "LA_deg": 2,
+    "totalIter": 1000,
+    "minimize_tol": 1.0e-6,
+}
+
+
+try:
+    state = gvec.find_state()
+except:
+    run = gvec.run(params)
+    state = run.state
+
+
+# %% [markdown]
+#
+# Note that all 1D and 2D plotting functions return the `matplotlib.figure.Figure` object, so you may edit the output if you wish.
+#
+# # 1D plots
+# Two are two convience functions for plotting 1D properties.
+# The first is to plot radial profiles, i.e. scalar functions of $\rho$. Note that only scalar quantities can be evaluated directly.
+
+# %%
+
+p_plot_radial_profile = gp.plot_radial_profile(state, 11)
+p_plot_radial_profile.show()
+
+# %% [markdown]
+# Properties along the magnetic axis can be plotted with `plot_on_axis`.
+# Note that some derived quantities are not well resolved at $\rho=0$.
+
+# %%
+
+p_on_axis = gp.plot_on_axis(state, quantities=["mod_B"])
+p_on_axis.show()
+
+
+# %% [markdown]
+# We can also set the layout of the subplots by providing the `subplot_grid` input. By default, these plots will not share their axes (so the scales on the
+#   y axis are the same), but we can force it with the `share_axis` argument.
+
+# %%
+
+p_on_axis_grid = gp.plot_on_axis(
+    state, quantities=["mod_B", "X1"], subplot_grid=[1, 2], share_axis=True
+)
+p_on_axis_grid.show()
+
+
+# %% [markdown]
+# # 2D plotting
+#
+# ## Poloidal slice plots
+#
+# If mutiple slices are requested by the `nzeta` input, you _must_ also include the `subplot_grid` input.
+# This will tell the plotting routine how to structure the subfigures. The second and third inputs is the number of discrete
+#   $\theta$ and $\zeta$ points.
+#
+# Note that the inputs for the 1D and 2D plots are very similar.
+
+# %%
+
+p_poloidal_slice_mod_B = gp.plot_poloidal_slice(
+    state, 11, 11, subplot_grid=[2, 2], share_axis=True, nzeta=4
+)
+p_poloidal_slice_mod_B.show()
+
+
+# %% [markdown]
+# ## Flux surface plots
+# As in the case with the 1D plots, the poloidal slice and flux surface plots share the same inputs.
[…]
