### 2025-10-27 4555457
fix scikit-build-core build-type
also test Debug mode for unit tests in the CI
