### 2025-10-27 6ef7887
make cmake options available in pygvec
* generate `gvec._compile_options` with the CMake script
* show some compile options in the output of `pygvec --version`
* test the `build-type` in the CI
