### 2025-10-27 e53f6a5
fixes for netcdf handling, and selalib abort is redirected to GVEC abort.
* the selalib errors only appear when Compiling in Debug mode and running the unit tests. Can now be catched by pygvec, as we use the GVEC abort.
