+++ test-CI/unit-pygvec/test_scripts.py
@@ -104,10 +104,17 @@ def test_run_recover_from_error():
     run = gvec.run(parameters)
 
 
-def test_run_netcdf_error():
+@pytest.mark.parametrize("which_read", ["hmap", "boundaryFromFile"])
+def test_run_netcdf_error(which_read):
     parameters = gvec.util.read_parameters("parameter.ini")
-    parameters["which_hmap"] = 21
-    parameters["hmap_ncfile"] = "non_existing_file.nc"
+    run = gvec.run(parameters)
+    if which_read=="hmap":
+        parameters["which_hmap"] = 21
+        parameters["hmap_ncfile"] = "non_existing_file.nc"
+    if which_read=="boundaryFromFile":
+        parameters["getBoundaryFromFile"] = 1
+        parameters["boundaryFromFile"] = "non_existing_file.nc"
+        
     with pytest.raises((RuntimeError, UnicodeDecodeError)):
         gvec.run(parameters)
 
