### 2025-10-28 535f8bb
fix issue #67, fix cleanup, checks.
* iota profile is now always defined at rho=0, using a quadratic extrapolation in radial coordinate s.
* check if knots / interpolation points for profiles start at 0 and end at 1.
* check if profiles are allocated before evaluation from pygvec.
