+++ src/pygvec/state.f90
@@ -870,12 +870,20 @@ SUBROUTINE evaluate_profile(n_s, s, deriv, var, result)
   ! CODE ------------------------------------------------------------------------------------------------------------------------!
   SELECT CASE(TRIM(var))
     CASE("iota")
+      IF(.NOT.ALLOCATED(iota_profile)) CALL abort(__STAMP__, &
+          'ERROR in profile evaluation: iota profile not initialized')
       input_profile = iota_profile
     CASE("p")
+      IF(.NOT.ALLOCATED(pres_profile)) CALL abort(__STAMP__, &
+          'ERROR in profile evaluation: pressure profile not initialized')
       input_profile = pres_profile
     CASE("chi")
+      IF(.NOT.ALLOCATED(chi_profile)) CALL abort(__STAMP__, &
+          'ERROR in profile evaluation: chi profile not initialized')
       input_profile = chi_profile
     CASE("Phi")
+      IF(.NOT.ALLOCATED(Phi_profile)) CALL abort(__STAMP__, &
+          'ERROR in profile evaluation: Phi profile not initialized')
       input_profile = Phi_profile
     CASE DEFAULT
       CALL abort(__STAMP__, &
