+++ src/functionals/mhd3d/mhd3d_evalfunc.f90
@@ -1525,26 +1525,31 @@ SUBROUTINE FinalizeMHD3D_EvalFunc()
     NULLIFY(DLA_tt); NULLIFY(DLA_tz); NULLIFY(DLA_zz)
     SDEALLOCATE(D_buf)
     IF(MPIroot)THEN
-      SELECT TYPE(precond_X1); TYPE IS(sll_t_spline_matrix_banded)
-        DO iMode=1,X1_Base%f%modes
-          CALL precond_X1(iMode)%free()
-        END DO !iMode
-      END SELECT !TYPE
-      SDEALLOCATE(precond_X1)
+      IF(ALLOCATED(precond_X1))THEN
+        SELECT TYPE(precond_X1); TYPE IS(sll_t_spline_matrix_banded)
+          DO iMode=1,SIZE(precond_X1)
+            CALL precond_X1(iMode)%free()
+          END DO !iMode
+        END SELECT !TYPE
+        DEALLOCATE(precond_X1)
+      END IF
+      IF(ALLOCATED(precond_X2))THEN
+        SELECT TYPE(precond_X2); TYPE IS(sll_t_spline_matrix_banded)
+          DO iMode=1,SIZE(precond_X2)
+            CALL precond_X2(iMode)%free()
+          END DO !iMode
+        END SELECT !TYPE
+        DEALLOCATE(precond_X2)
+      END IF
 
-      SELECT TYPE(precond_X2); TYPE IS(sll_t_spline_matrix_banded)
-        DO iMode=1,X2_base%f%modes
-          CALL precond_X2(iMode)%free()
-        END DO !iMode
-      END SELECT !TYPE
-      SDEALLOCATE(precond_X2)
-
-      SELECT TYPE(precond_LA); TYPE IS(sll_t_spline_matrix_banded)
-        DO iMode=1,LA_base%f%modes
-          CALL precond_LA(iMode)%free()
-        END DO !iMode
-      END SELECT !TYPE
-      SDEALLOCATE(precond_LA)
+      IF(ALLOCATED(precond_LA))THEN
+        SELECT TYPE(precond_LA); TYPE IS(sll_t_spline_matrix_banded)
+          DO iMode=1,SIZE(precond_LA)
+            CALL precond_LA(iMode)%free()
+          END DO !iMode
+        END SELECT !TYPE
+        DEALLOCATE(precond_LA)
+      END IF
     END IF !MPIroot
   END IF !PrecondType>0
 
