### 2025-10-28 d17c10d
fixed reason for segfault in state.90, when compiled with debug.
* par_Init and par_finalize were missing, such that variables were left unallocated/uninitialized.
* was not really a problem, since running without mpi, these variables are not used!
* debug logging extended in the binding/unbinding
