### 2025-10-29 beeba52
update gene-gist interface
