+++ python/gvec/scripts/gene_gist.py
@@ -24,13 +24,6 @@ parser.add_argument(
     help="GVEC run directory",
     default=Path("."),
 )
-parser.add_argument(
-    "-o",
-    "--outputfile",
-    type=Path,
-    help="output file name (default: '{projectname}_s{s}.gist.txt')",
-    default=None,
-)
 srho = parser.add_mutually_exclusive_group(required=True)
 srho.add_argument(
     "-s",
@@ -69,10 +62,23 @@ parser.add_argument(
     help="flip the poloidal or toroidal direction with respect (left-handed Boozer coordinates).",
 )
 parser.add_argument(
+    "-o",
+    "--outputfile",
+    type=Path,
+    help="output file name (default: '{projectname}_s{s}.gist.txt')",
+    default=None,
+)
+parser.add_argument(
     "-p",
     "--plot",
     action="store_true",
     help="plot the output quantities ('{projectname}_s{s}.gist.png')",
+)
+parser.add_argument(
+    "--projectname",
+    type=str,
+    help="override the project name for the output files (default: use GVEC state name)",
+    default=None,
 )
 verbosity = parser.add_mutually_exclusive_group()
 verbosity.add_argument(
@@ -141,121 +147,109 @@ def compute_gist_quantities(ev, state, flip):
         "iota",
         "diota_dr",
         "p",
+        "dp_dr",
         "minor_radius",
         "Jac_B",
         "mod_B",
         "dmod_B_dr_B",
         "dmod_B_dt_B",
         "dmod_B_dz_B",
-        "dp_dr",
         "mu0",
     )
 
     ev = ev.squeeze()
 
-    rho = ev.rho
+    # === GVEC coordinates === #
+
+    rho = ev.rho.item()
     theta = ev.theta_B
     a = ev.minor_radius.item()
-    s0 = ev.rho**2
-    sqrt_s0 = ev.rho
-    iota = ev.iota
-    diota_dr = ev.diota_dr
-    diota_ds = diota_dr / (2 * ev.rho)
-    p0 = ev.p.item()
+    iota = ev.iota.item()
+    diota_dr = ev.diota_dr.item()
+    p = ev.p.item()
+    dp_dr = ev.dp_dr.item()
     grad_rho = ev.grad_rho
-    grad_vartheta = ev.grad_theta_B
+    grad_theta = ev.grad_theta_B
     grad_zeta = ev.grad_zeta_B
-    dB_drho = ev.dmod_B_dr_B
-    dB_dvartheta = ev.dmod_B_dt_B
-    dB_dzeta = ev.dmod_B_dz_B
-    Phi_lcfs = state.evaluate("Phi", rho=1.0, theta=None, zeta=None).Phi.item()
+    dmodB_drho = ev.dmod_B_dr_B
+    dmodB_dtheta = ev.dmod_B_dt_B
+    dmodB_dzeta = ev.dmod_B_dz_B
+    Phi_edge = state.evaluate("Phi", rho=1.0, theta=None, zeta=None).Phi.item()
     mu0 = ev.mu0.item()
-    B = ev.mod_B
+    modB = ev.mod_B
     Jac_B = ev.Jac_B  # only used for assertion
+
+    # === flipping coordinates === #
 
     if flip in ("tor", "both"):
         iota = -iota
         diota_dr = -diota_dr
-        diota_ds = -diota_ds
         grad_zeta = -grad_zeta
-        dB_dzeta = -dB_dzeta
-        Phi_lcfs = -Phi_lcfs
[…]
