+++ docs/dev/derivations.md
@@ -31,7 +31,7 @@ $$
 
 for $\alpha\in\left\{\rho,\thet,\zeta\right\}$.
 
-$\Rightarrow$ `dmod_B_dr`, `dmod_B_dt`, `dmod_B_dz`, `grad_mod_B`
+$\Rightarrow$ `dmodB_dr`, `dmodB_dt`, `dmodB_dz`, `grad_modB`
 
 ## Derivatives of $\vec{B}$
 
@@ -113,48 +113,390 @@ $\Rightarrow$ `dB_contra_t_dr`, `dB_contra_t_dt`, `dB_contra_t_dz`, `dB_contra_z
 
 ## export GIST files for GENE
 
-Left-handed coordinate systems $\left(s,\alpha,\theta\right)$ and $\left(x^1, x^2, x^3\right) = \left(\sqrt{s}, \frac{\sqrt{s_0}}{a}\alpha,\theta\right)$.
+?-handed coordinate systems $\left(s,\alpha,\theta\right)$ and $\left(x^1, x^2, x^3\right) = \left(a \sqrt{s}, a\frac{\sqrt{s_0}}{q_0}\alpha, a\theta\right)$.
 
-With respect to a left-handed $\left(\rho, \vartheta, \zeta\right)$, the fieldline coordinates are defined as
+With respect to a ?-handed $\left(\rho, \vartheta, \zeta\right)$, the fieldline coordinates are defined as
+
 $$
 \begin{align}
 s &= \rho^2 &
 \rho &= \sqrt{s} \\
-\alpha &= \vartheta/\iota - \zeta &
-\zeta &= \theta/\iota - \alpha \\
-\theta &= \vartheta &
-\vartheta &= \theta \\
+\alpha &= \sigma\vartheta/\iota - \sigma\zeta &
+\zeta &= \theta/\iota - \sigma\alpha \\
+\theta &= \sigma\vartheta &
+\vartheta &= \sigma\theta \\
 \end{align}
 $$
-which can be obtained from the right-handed GVEC angles by flipping the toroidal or poloidal direction.
+
+<!-- which can be obtained from the right-handed GVEC angles by flipping the toroidal or poloidal direction. -->
 
 $$
 \begin{align}
 \frac{\partial s}{\partial \rho} &= 2\rho &
 \frac{\partial \rho}{\partial s} &= \frac{1}{2\sqrt{s}} \\
-\frac{\partial \alpha}{\partial \rho} &= -\frac{d\iota}{d\rho} \frac{\vartheta}{\iota^2} &
+\frac{\partial \alpha}{\partial \rho} &= -\sigma \frac{d\iota}{d\rho} \frac{\vartheta}{\iota^2} &
 \frac{\partial \zeta}{\partial s} &= -\frac{d\iota}{ds} \frac{\theta}{\iota^2} \\
-\frac{\partial \alpha}{\partial \vartheta} &= \frac{1}{\iota} &
+\frac{\partial \alpha}{\partial \vartheta} &= \sigma\frac{1}{\iota} &
 \frac{\partial \zeta}{\partial \theta} &= \frac{1}{\iota} \\
-\frac{\partial \alpha}{\partial \zeta} &= -1 &
-\frac{\partial \zeta}{\partial \alpha} &= -1 \\
-\frac{\partial \theta}{\partial \vartheta} &= 1 &
[…]
