### 2025-10-29 da1ecf3
- Changed function name of `plot_poloidal_slice` to   `plot_poloidal_plane` - Added warnings if dataset contains NaNs, which can happen if certain   quantities are evaluated on the axis sometimes - Updated how axes are linked, now if there's a grid it will try and   link the x-axis always for the 1D plots - Plotting poloidal planes now takes an optional input to set the   minimium rho value in case NaNs appear in dataset - rho and theta^star contours can be plotted on poloidal planes. TODO:   let the user input the number of contours - Added an option to the 3D plotting to plot either the full
[…]
