+++ python/gvec/plotting/plots.py
@@ -1,4 +1,6 @@
-from numpy import array, max, min, ndarray
+from numpy import array, max, min, ndarray, linspace, pi, any, isnan
+
+from warnings import warn
 
 import matplotlib.pyplot as pyplot
 
@@ -10,20 +12,30 @@ from .utils import _get_coord_range, _get_scalars_for_plotting
 pyplot.rcParams.update({"text.usetex": True})
 
 
-def _plot_line_quantities_from_dict(
-    plotting_quantities, x_axis_value, subplot_grid, share_axis, xlabel
-):
+def _plot_line_quantities_from_dict(plotting_quantities, x_axis_value, subplot_grid, xlabel):
     """
     Plot the quantities from `plotting_quantities`.
 
-    Return both the `matplotlib.pyplot.figure` object and the `array([axes])`.
+    Return both the `matplotlib.pyplot.figure`
     """
 
-    f, axs = pyplot.subplots(*subplot_grid, sharex=share_axis)
+    link_xaxis = False
+    hide_inner_axis = False
+
+    if (subplot_grid[0] != 1) & (subplot_grid[1] == 1):
+        link_xaxis = True
+    elif (subplot_grid[0] != 1) & (subplot_grid[1] != 1):
+        link_xaxis = True
+        hide_inner_axis = True
+
+    f, axs = pyplot.subplots(*subplot_grid, sharex=link_xaxis)
 
     # Ensure this is an array so we can iterate over it
     if not isinstance(axs, ndarray):
         axs = array([axs])
+    if len(axs) != 1:
+        # Otherwise we cannot iterate over this easily
+        axs = axs.flatten()
 
     if axs.size == 1:
         # If there is only one axis we plot all quantities on the single axis
@@ -38,7 +50,7 @@ def _plot_line_quantities_from_dict(
             ax_lines = axs[i].plot(x_axis_value, plotting_quantities[quantity])
             # Top right of each subplot will have a text box with the quantity being plotted
             axs[i].annotate(
-                quantity,
+                quantity,  # TODO: convert this to latex
                 xy=(1, 1),
                 xycoords="axes fraction",
                 xytext=(-0.6, -0.6),
@@ -47,19 +59,27 @@ def _plot_line_quantities_from_dict(
                 horizontalalignment="right",
                 bbox=dict(facecolor="white", edgecolor="black"),
             )
-            axs[i].set_xlabel(xlabel)
+
+            if (not link_xaxis) | (i == (len(axs) - 1)):
+                # If the axis are linked we only need the label on the last plot
+                axs[i].set_xlabel(xlabel)
+            if (hide_inner_axis) & (i - len(axs) + subplot_grid[1] >= 0):
+                # If there are multiple plots we need to set the x-axis labels only on the bottom row
+                axs[i].set_xlabel(xlabel)
 
     return f, axs
 
 
-def plot_poloidal_slice(
+def plot_poloidal_plane(
     state: State,
     nrho: int = 51,
     ntheta: int = 51,
     nzeta: int | float | ndarray = 1,
     quantities: str = "mod_B",
     subplot_grid: list = [1, 1],
-    share_axis: bool = False,  # TODO: Fix this so that only the minor radius is shared along the x-axis
+    share_axis: bool = False,
+    show_contours=False,  # TODO: update so a limited number of contours can be specified
+    min_rho=0.0,
     # show_origin = False #TODO
 ):
     """
@@ -92,15 +112,25 @@ def plot_poloidal_slice(
     `matplotlib.pyplot.figure` object
     """
 
-    rho = _get_coord_range("rho", state.nfp, nrho)
+    rho = _get_coord_range("rho", state.nfp, nrho, min_val=min_rho)
[…]
