+++ python/gvec/plotting/__init__.py
@@ -1,6 +1,6 @@
 from .plots3d import plot_3d_surface, plot_boundary
 from .plots import (
-    plot_poloidal_slice,
+    plot_poloidal_plane,
     plot_flux_surface,
     plot_radial_profile,
     plot_on_axis,
