### 2025-10-29 3ff401d
subregion logging in fortran code, for abort message.
* in the globals module, a global variable is used for tracking region.subregion of the code, for the abort error message.
* subroutines `reset_subregion`, `enter_subregion`, `exit_subregion` control resetting / adding / removing strings for each subregion in the fortran code.
