+++ src/globals/globals.f90
@@ -41,11 +41,15 @@ INTEGER                     :: testlevel =-1             !! flag for testing rou
 INTEGER                     :: ntestCalled=0             !! counter for called tests
 INTEGER                     :: nfailedMsg=0              !! counter for messages on failed tests
 INTEGER                     :: testUnit                  !! unit for out.test file
-INTEGER                     :: ProgressBar_oldpercent    !! for progressBar
-REAL(wp)                    :: ProgressBar_starttime     !! for progressBar
+!MPI--------------------------------------------------------------------------------------------------------------------------------
 LOGICAL                     :: MPIRoot=.TRUE.            !! flag whether process is MPI root process
 INTEGER                     :: myRank=0                  !! rank of the MPI task
 INTEGER                     :: nRanks=1                  !! total number of MPI tasks
+!-----------------------------------------------------------------------------------------------------------------------------------
+CHARACTER(LEN=20)           :: active_region(5)=(/"", "", "", "", ""/) !! for abort messages, to identify which (sub-)region was currently
+INTEGER                     :: iregion=0                 !! which active_region to fill
+INTEGER                     :: ProgressBar_oldpercent    !! for progressBar
+REAL(wp)                    :: ProgressBar_starttime     !! for progressBar
 !-----------------------------------------------------------------------------------------------------------------------------------
 #ifndef NOISOENV
 INTEGER, PARAMETER          :: UNIT_stdIn  = input_unit  !! Terminal input
@@ -57,6 +61,19 @@ INTEGER, PARAMETER          :: UNIT_stdOut = 6           !! Terminal output
 INTEGER, PARAMETER          :: UNIT_errOut = 0           !! For error output
 #endif
 INTEGER, PARAMETER          :: MAXLEN  = 4096       !! max length of strings, needed for string handling when compiled with NVHPC
+
+INTERFACE reset_subregion
+  MODULE PROCEDURE reset_subregion
+END INTERFACE
+
+INTERFACE enter_subregion
+  MODULE PROCEDURE enter_subregion
+END INTERFACE
+
+INTERFACE exit_subregion
+  MODULE PROCEDURE exit_subregion
+END INTERFACE
+
 INTERFACE Abort
    MODULE PROCEDURE Abort
 END INTERFACE
@@ -107,12 +124,91 @@ END INTERFACE
 
 CONTAINS
 
+
+!==================================================================================================================================
+!> reset global variables for the subregion output to default
+!==================================================================================================================================
+SUBROUTINE reset_subregion
+! MODULES
+  IMPLICIT NONE
+!==================================================================================================================================
+  iregion=0
+  active_region=""
+END SUBROUTINE reset_subregion
+
+!==================================================================================================================================
+!> add the current subregion to the active_regions (maximum depth is 5)
+!! This information is collected uniquely for the abort error message
+!!
+!==================================================================================================================================
+SUBROUTINE enter_subregion(subregion_name)
+! MODULES
+  IMPLICIT NONE
+!----------------------------------------------------------------------------------------------------------------------------------
+! INPUT/OUTPUT VARIABLES
+  CHARACTER(LEN=*), INTENT(IN) :: subregion_name
+!----------------------------------------------------------------------------------------------------------------------------------
[…]

[…]
