### 2025-10-29 4f1e470
add Error Type to relevant abort calls.
* The abort function takes an optional input called TypeInfo. If not specified, the error type is "RuntimeError".
* the following error types are used:
  "MissingParameterError","InvalidParameterError","FileNotFoundError","InitializationError"
* only abort calls where these error types apply, now pass the corresponding TypeInfo
* for errors in the netcdf, we use as default a "netCDF_error"
* The abort message now has three parts, separated by " | ":
  code region | errortype | errormessage
