+++ src/readstate/boundaryFromFile.f90
@@ -207,7 +207,8 @@ SUBROUTINE bff_convert_to_modes(sf,x1_fbase_in,x2_fbase_in,X1_b,X2_b,scale_minor
   WRITE(UNIT_stdOut,'(A)')'  CONVERT BOUNDARY FROM POINTS TO MODES:'
 
   IF(sf%nfp.NE.x1_fbase_in%nfp)CALL abort(__STAMP__,&
-                 " error in convert boundary to modes, nfp from GVEC parameterfile does not match to nfp of boundary file")
+                 " error in convert boundary to modes, nfp from GVEC parameterfile does not match to nfp of boundary file",&
+                 TypeInfo="InvalidParameterError")
   mn_max_pts(1:2)=(/(sf%ntheta-1)/2,(sf%nzeta-1)/2/)
   WRITE(UNIT_stdOut,'(6X,2(A,I4),2A)')' from X,Y(ntheta= ',sf%ntheta,', nzeta= ', &
                                                sf%nzeta, '), lasym=',MERGE("symmetric","asymetric",(sf%lasym.EQ.0))
