+++ src/hmap/hmap_knot.f90
@@ -120,7 +120,8 @@ FUNCTION hmap_knot_init_params(knot_kl,knot_R0,knot_delta) RESULT(sf)
 
   IF (.NOT.((sf%R0 - ABS(sf%delta)) > 0.0_wp)) THEN
      CALL abort(__STAMP__, &
-          "hmap_knot init: condition R0 - |delta| > 0 not fulfilled!")
+          "hmap_knot init: condition R0 - |delta| > 0 not fulfilled!", &
+          TypeInfo="InvalidParameterError")
   END IF
 
   sf%initialized=.TRUE.
