### 2025-10-30 4dc793e
add shift boundary util & adapt convert-params script
