+++ python/gvec/util.py
@@ -564,6 +564,24 @@ def flip_boundary_zeta(parameters: MutableMapping) -> MutableMapping:
     return output_params
 
 
+def shift_boundary_theta_pi(parameters: MutableMapping) -> MutableMapping:
+    """
+    Shift the theta origin of the boundary by pi.
+
+    cos(m(θ+π)-nζ) = (-1)^m cos(mθ-nζ)
+    sin(m(θ+π)-nζ) = (-1)^m sin(mθ-nζ)
+    """
+    parameters = copy.deepcopy(parameters)
+    for var in ["X1", "X2"]:
+        for sc in ["cos", "sin"]:
+            key = f"{var}_b_{sc}"
+            if key in parameters:
+                for (m, n), value in list(parameters[key].items()):
+                    if m % 2 == 1:
+                        parameters[key][m, n] = -value
+    return parameters
+
+
 def flip_parameters_theta(parameters: MutableMapping) -> MutableMapping:
     parameters = flip_boundary_theta(parameters)
 
