+++ python/gvec/scripts/main.py
@@ -11,7 +11,7 @@ import argparse
 from collections.abc import Sequence
 
 import gvec
-from gvec.scripts import cas3d, gist as gist, run, quasr
+from gvec.scripts import cas3d, convert, gist, quasr, run
 
 # === Arguments === #
 
@@ -23,6 +23,7 @@ subparsers = parser.add_subparsers(
     title="mode",
     description="which mode/subcommand to run",
     dest="mode",
+    metavar="MODE",
 )
 parser.add_argument(
     "-V",
@@ -30,63 +31,6 @@ parser.add_argument(
     action="version",
     version=f"pyGVEC v{gvec.__version__} from {Path(gvec.__file__).parent} (python {platform.python_version()})",
 )
-
-# --- convert parameterfile --- #
-
-convert_parser = subparsers.add_parser(
-    "convert-params",
-    help="convert the GVEC parameterfile between different formats",
-    formatter_class=argparse.RawDescriptionHelpFormatter,
-    description="Convert GVEC parameterfiles between different formats.\n"
-    "The INI (classical) parameter files do not support stages or the current optimization!\nAlso the formatting is lost upon conversion.",
-)
-convert_parser.add_argument(
-    "input",
-    type=Path,
-    help="input GVEC or VMEC parameterfile",
-)
-convert_parser.add_argument(
-    "output",
-    type=Path,
-    nargs="?",
-    help="output GVEC parameterfile",
-    default="parameter.yaml",
-)
-convert_parser.add_argument(
-    "--vmec",
-    action="store_true",
-    help="input parameterfile is a VMEC namelist",
-)
-convert_parser.add_argument(
-    "-x",
-    "--flip",
-    choices=["auto", "none", "tor", "pol", "both"],
-    default="auto",
-    help="flip the coordinates in the specified direction(s), possible values are: 'auto' (default), 'none', 'tor', 'pol', 'both'.",
-    metavar="FLIP",
-)
-convert_parser.add_argument(
-    "-t",
-    "--shift-theta",
-    choices=["auto", "pi", "0"],
-    default="auto",
-    help="shift the theta coordinate origin, possible values are 'auto' (default), 'pi' or '0'.",
-    metavar="SHIFT",
-)
-convert_parser.add_argument(
-    "--stellsym",
-    action="store_true",
-    help="enforce stellarator symmetry in the output parameterfile",
-)
-verbosity = convert_parser.add_mutually_exclusive_group()
-verbosity.add_argument(
-    "-v",
-    "--verbose",
-    action="count",
-    default=0,
-    help="verbosity level: -v for info, -vv for debug",
-)
-verbosity.add_argument("-q", "--quiet", action="store_true", help="suppress output")
 
 # --- scripts --- #
 
@@ -97,6 +41,24 @@ run_parser = subparsers.add_parser(
     description=run.parser.description,
     parents=[run.parser],
     add_help=False,
+)
+
+convert_parser = subparsers.add_parser(
+    "convert-params",
+    help="convert the GVEC parameterfile between different formats",
+    formatter_class=convert.parser.formatter_class,
+    description=convert.parser.description,
+    parents=[convert.parser],
+    add_help=False,
+)
+
+quasr_parser = subparsers.add_parser(
+    "load-quasr",
+    help=quasr.parser.description,
+    description=quasr.parser.description,
+    parents=[quasr.parser],
+    add_help=False,
+    usage=quasr.parser.usage,
 )
 
 cas3d_parser = subparsers.add_parser(
@@ -115,111 +77,25 @@ gist_parser = subparsers.add_parser(
     add_help=False,
 )
 
-quasr_parser = subparsers.add_parser(
-    "load-quasr",
[…]
