### 2025-10-30 99debd4
add documentation for gene-gist
