+++ docs/dev/derivations.md
@@ -186,271 +186,3 @@ $$
 \nabla \zeta_B &= \pdv{\zeta_B}{\rho}\nabla \rho + \pdv{\zeta_B}{\thet}\nabla \thet + \pdv{\zeta_B}{\zeta}\nabla \zeta \\
 \end{align}
 $$
-
-## export GIST files for GENE
-
-### Boozer coordinates
-
-Boozer coordinates $\left(\rho,\theta_B,\zeta_B\right)$ are a particular kind of straight-fieldline coordinates, where the following relations hold for the magnetic field $\vec{B}$:
-
-$$
-\begin{align}
-\vec{B}\cdot\vec{e}^{\rho_B} = B^{\rho_B} &= 0 \\
-\vec{B}\cdot\vec{e}^{\theta_B} = B^{\theta_B} &= \frac{1}{\Jac_{B}}\frac{d\Phi}{d\rho}\iota \\
-\vec{B}\cdot\vec{e}^{\zeta_B} = B^{\zeta_B} &= \frac{1}{\Jac_{B}}\frac{d\Phi}{d\rho} \\
-\vec{B}\cdot\nabla\theta_B = B_{\theta_B}(\rho_B) \\
-\vec{B}\cdot\nabla\zeta_B = B_{\zeta_B}(\rho_B) \\
-\end{align}
-$$
-
-Here $\rho_B\in[0,1]$ is a radial, radius-like coordinate, typically chosen such that $\rho^2$ is the normalized toroidal flux.
-$\theta_B, \zeta_B \in [0,2\pi)$ are the poloidal and toroidal periodic, angle-like coordinates.
-
-### changing coordinate directions
-
-In GVEC Boozer coordinates are right-handed and follow the directions of the logical (flux aligned) coordinates $\left(\rho,\theta,\zeta\right)$.
-The directions of the two angular coordinates are arbitrary and need to be flipped in certain cases.
-
-Flipping toroidally $\zeta \to -\zeta$ changes $\Phi \to -\Phi, \iota \to -\iota, \Jac \to -\Jac$.
-
-Flipping poloidally $\theta \to -\theta$ changes $\chi \to -\chi, \iota \to -\iota, \Jac \to -\Jac$.
-
-The geometry also needs to be flipped accordingly and all derived quantities change, but the formulas stay the same.
-
-### Fieldline coordinates
-
-The transformation to the fieldline coordinates $\left(\rho_\alpha,\alpha,\phi_\alpha\right)$ can be defined as:
-
-$$
-\begin{align}
-\rho_\alpha &= \rho_B \\
-\alpha &= \theta_B / \iota - \zeta_B \\
-\phi_\alpha &= \theta_B
-\end{align}
-$$
-
-$$
-\begin{align}
-\pdv{\rho_\alpha}{\rho_B} &= 1 &
-\pdv{\rho_\alpha}{\theta_B} &= 0 &
-\pdv{\rho_\alpha}{\zeta_B} &= 0 \\
-\pdv{\alpha}{\rho_B} &= -\frac{\theta_B}{\iota^2} \frac{d\iota}{d\rho} &
-\pdv{\alpha}{\theta_B} &= \frac{1}{\iota} &
-\pdv{\alpha}{\zeta_B} &= -1 \\
-\pdv{\phi_\alpha}{\rho_B} &= 0 &
[…]
