### 2025-10-30 1101f2b
add f90nml to dependencies
