+++ pyproject.toml
@@ -23,6 +23,7 @@ dependencies = [
     "tomlkit",
     "tqdm",
     "xarray",
+    "f90nml",
 ]
 authors = [
     {name = "Florian Hindenlang"},
@@ -59,7 +60,7 @@ Issues = "https://gitlab.mpcdf.mpg.de/gvec-group/gvec/-/issues"
 
 [project.optional-dependencies]
 dev = ["pytest"]
-examples = ["f90nml", "matplotlib", "jupyter", "ipywidgets", "ipympl", "simsopt"]
+examples = ["matplotlib", "jupyter", "ipywidgets", "ipympl", "simsopt"]
 
 [project.scripts]
 pygvec = "gvec.scripts.main:main"
