### 2025-10-31 ad1bace
visualization from G-Frame files. check field periodicity in quasr.
* moved real_dft from quasr to fourier, also the test.
* in quasr script, check the input surface, especially its field periodicity (by rotation of 2pi/nfp).
* add visualization from a G-Frame file, gvec.util.gframe_to_vtk()... tested on existing Gframe files in `test-CI/data`, and also added it to one quasr test.
