+++ python/gvec/scripts/quasr.py
@@ -150,62 +150,74 @@ def check_args(parser, args):
 # === Functions === #
 
 
-def real_dft_mat(x_in, x_out, nfp=1, modes=None, deriv=0):
-    """
-    Flexible Direct Fourier Transform for real data
-    takes an input array of equidistant points in [0,2pi/nfp[ (exclude endpoint!),
-    evaluate the discrete fourier transform with the given 1d mode vector (all >=0) using the input points x_in, then evaluate the inverse transform (or its derivative deriv>0) on the output points x_out anywhere...
-    len(x_in) must be > 2*max(modes)
-    output is the matrix that transforms real function to real function [derivative]:
-     f^deriv(x_out) = Mat f(x_in) (can then be used to do 2d transforms with matmul!)
-
-    nfp is the number of field periods, default 1 (int), all modes are multiples of nfp
-
-    """
-    if modes is None:
-        modes = np.arange((len(x_in) - 1) // 2 + 1)  # all modes up to Nyquist
-    assert np.allclose(x_in[-1] + (x_in[1] - x_in[0]) - x_in[0], 2 * np.pi / nfp)
-    assert np.all(modes >= 0), "modes must be positive"
-    zeromode = np.where(modes == 0)
-    assert len(zeromode) <= 1, "only one zero mode allowed"
-    maxmode = np.amax(modes)
-    assert len(x_in) > 2 * maxmode, (
-        f"number of sampling points ({len(x_in)}) > 2*maxmodenumber ({maxmode})"
+def check_field_periodicity(xyz: np.ndarray, nfp: int, atol=1e-12):
+    assert xyz.shape[-1] == 3, (
+        "last dimension must be the cartesian components surface positions!"
     )
-    # matrix for forward transform
-    Fmat = np.exp(1j * nfp * (modes[:, None] * x_in[None, :]))
-    mass_re = Fmat.real @ Fmat.real.T
-    mass_im = Fmat.imag @ Fmat.imag.T
-    diag_re = np.copy(np.diag(mass_re))
-    diag_im = np.copy(np.diag(mass_im))
+    if nfp == 1:
+        return  # nothing to check
+    assert np.mod(xyz.shape[0], nfp) == 0, (
+        "number of points in zeta direction  must be divisible by nfp!"
+    )
+    nzeta_fp = xyz.shape[0] // nfp
 
-    assert np.all(np.abs(mass_re - np.diag(diag_re)) < 1.0e-8), "massre must be diagonal"
-    assert np.all(np.abs(mass_im - np.diag(diag_im)) < 1.0e-8), "massim must be diagonal"
-    diag_im[zeromode] = 1  # imag (=sin) is zero at zero mode
-    assert np.all(diag_re > 0.0)
-    assert np.all(diag_im > 0.0)
+    # find rotation direction:
+    xyz_rot_pos = rodrigues(xyz[0, 0, :], 2 * np.pi / nfp)
+    dist_pos = np.sqrt(np.sum((xyz_rot_pos - xyz[nzeta_fp, 0, :]) ** 2))
+    xyz_rot_neg = rodrigues(xyz[0, 0, :], -2 * np.pi / nfp)
+    dist_neg = np.sqrt(np.sum((xyz_rot_neg - xyz[nzeta_fp, 0, :]) ** 2))
+    if dist_pos < atol:
+        angle_sign = 1
+    elif dist_neg < atol:
+        angle_sign = -1
+    else:
+        raise ValueError(
+            f"the first point of the surface [0,0] is not rotationally symmetric with the next field period. nfp={nfp}, absolute distance={np.amax([dist_pos, dist_neg])} not within tolerance {atol}!"
+        )
 
[…]
