### 2025-10-31 16324e3
moved the gframe_to_vtk function from util to vtk.
* seems that util should not have any dependency on other gvec "inner" files.
