+++ python/gvec/vtk.py
@@ -6,6 +6,7 @@ import logging
 from pyevtk.hl import gridToVTK
 import xarray as xr
 import numpy as np
+from gvec import fourier
 
 
 def ev2vtk(
@@ -147,3 +148,186 @@ def ev2vtk(
 
     if not quiet:
         return Path(fn)
+
+
+def gframe_to_vtk(
+    file: str | Path,
+    prefix="visu",
+    zeta_visu: np.ndarray = None,
+    theta_visu: np.ndarray = None,
+    phi_visu: np.ndarray = None,
+    box_axis=None,
+    filetype="vts",
+):
+    """
+    Reads a netcdf file that defines the G-Frame and possibly the boundary X1,X2 in that frame.
+    The file is for example produced by the GVEC quasr script, and then used in GVEC for initialization.
+    Writes vtk-visualization files from the data.
+    Input:
+        * file : netcdf file that contains axis and boundary data. Format as produced by the GVEC quasr script.
+        * prefix : prefix of the output files. Default is `visu_`.
+        * zeta_visu : 1d zeta positions of the axis and boundary surface output. If not specified, the ones from the file are used.
+        * theta_visu : 1d theta positions of the boundary surface output. If not specified, the ones from the file are used.
+        * box_axis=[a,b] : visualize G-Frame with a box of with distances +a -a in N direction and +b -b in B direction. If None, axis curve is visualized as a line.
+        * filetype : can be "vts"  (VTK) or "nc" (netcdf)
+    Output:
+        * writes `prefix_axis.filetype` : if 'axis' group exists in `file`, provides the origin curve position in 3D and N,B vectors on that curve. On full torus or on given `zeta_visu` positions
+        * writes `prefix_boundary.filetype` : if 'boundary' group exists `file`, provides the boundary surface position in 3D. On one field period, or on given `zeta_visu` positions
+    """
+
+    ds_main = xr.open_dataset(file, engine="netcdf4")
+    nfp = ds_main.NFP.data
+    try:
+        ds_axis = xr.open_dataset(file, engine="netcdf4", group="axis")
+    except Exception as e:
+        raise RuntimeError(f"Could not open axis group in {file}") from e
+
+    zeta_fp = ds_axis["zeta(:)"].data
+    pos_axis = ds_axis["xyz(::)"].data
+    N_axis = ds_axis["Nxyz(::)"].data
+    B_axis = ds_axis["Bxyz(::)"].data
+    ds_axis.close()
+
+    nzeta_fp_axis = len(zeta_fp)
+    nzeta_full = N_axis.shape[1]
+    assert nzeta_full == nzeta_fp_axis * nfp, (
+        f"axis data must be given on a full turn, with nfp being a factor in the number of points! nfp={nfp}, nzeta_full={nzeta_full}, nzeta of one fp={nzeta_fp_axis}"
+    )
+    zeta_axis = zeta_fp[0] + np.linspace(0, 2 * np.pi, nzeta_full, endpoint=False)
+
+    assert np.allclose(zeta_fp, zeta_axis[0:nzeta_fp_axis]), "zeta on axis must be equidistant"
+    if zeta_visu is not None:
+        zdft = fourier.real_dft_mat(zeta_axis, zeta_visu, nfp=1)
+        zeta_out = zeta_visu
+        pos = pos_axis @ zdft["BF"].T
+        N = N_axis @ zdft["BF"].T
+        B = B_axis @ zdft["BF"].T
+    else:
+        zeta_out = zeta_axis
+        pos = pos_axis
+        N = N_axis
[…]
