### 2025-10-31 74040e6
update documentation and in code documentation.
* describing construction of G-Frame configuration from other surface data.
* how to visualize
