### 2025-10-31 7caa422
fix calling scipy.optimize.root_scalar, from deprecation warning.
