### 2025-10-31 6f1011a
extend ev2vtk capabilities to 1d,2d datasets.
* input dimension names can now also be 'rho' ->'rad', 'theta'->'pol' , 'zeta'->'tor' (renamed on a copy of the dataset, internally)
* the input dataset needs if any of the 'rad','theta','zeta' dimensions are missing, the pos dataarray is simply extended by a dimension of size 1.
