+++ python/gvec/vtk.py
@@ -33,16 +33,17 @@ def ev2vtk(
 
     Notes
     -----
-    The following dimension are expected to be in the dataset:
-    - 'pos' : the cartesian components of grid points
+    The following data / dimensions are expected to be in the dataset:
     - 'xyz' : the dimension name for the cartesian components of grid points
-    - 'rad' : the radial dimension name
-    - 'pol' : the poloidal dimension name
-    - 'tor' : the toroidal dimension name
+    - 'rad' /'rho'   : the radial dimension name. If not present in xrds, it is added as {"rad":[0]}
+    - 'pol' /'theta' : the poloidal dimension name. If not present in xrds, it is added as {"pol":[0]}
+    - 'tor' /'zeta'  : the toroidal dimension name. If not present in xrds, it is added as {"tor":[0]}
+    - 'pos' : datarray with the cartesian components of grid points, with dimension 'xyz' and at least one of ['rad', 'pol', 'tor']
 
-    Scalar variables without the 'xyz' dimension are broadcasted to the 'rad', 'pol', 'tor' dimensions.
+    Scalar variables are dataarrays without the 'xyz' dimension, and  are broadcasted to the 'rad', 'pol', 'tor' dimensions.
+    Vector variables are dataarrays with the 'xyz' dimension, and are broadcasted to the 'xyz', 'rad', 'pol', 'tor' dimensions.
 
-    If a variable does not have the expected dimensions, it is ignored.
+    If a dataarray (except 'pos') does not have the expected dimensions, it is ignored.
 
     Examples
     --------
@@ -62,50 +63,71 @@ def ev2vtk(
 
     # make sure dimensions are in the expected order
     dimension_order = ["xyz", "rad", "pol", "tor"]
-
+    ds_out = xrds.copy()
     assert (
-        "pos" in xrds
+        "pos" in ds_out
     ), """Expected 'pos' in 'xrds', please make sure you are working with a pygvec evaluation dataset
     or rename your variable for the  cartesian components of grid points to 'pos'."""
 
+    # rename dimensions
+    for dim, dimnew in (("rho", "rad"), ("theta", "pol"), ("zeta", "tor")):
+        if dim in ds_out.pos.dims:
+            ds_out = ds_out.rename_dims({dim: dimnew})
+
+    assert np.any([dim in ds_out.pos.dims for dim in ("rad", "pol", "tor")]), (
+        'pos data array must contain at least one of "rad", "pol", "tor" dimensions.'
+    )
+    # add missing dimensions to pos
+    expanded_dim = []
+    for dim in ("rad", "pol", "tor"):
+        if dim not in ds_out.pos.dims:
+            expanded_dim.append(dim)
+            ds_out["pos"] = ds_out.pos.expand_dims({dim: [0.0]})
+
     expected_dimension = {"rad": "radial", "pol": "poloidal", "tor": "toroidal"}
+
     for dim in expected_dimension:
         assert (
-            dim in xrds.dims
+            dim in ds_out.dims
         ), f"""Expected '{dim}' in 'xrds' dimensions, please make sure you are working with a pygvec evaluation dataset
         or rename your {expected_dimension[dim]} dimension to '{dim}'."""
 
     outvars = []
     ignored_variables = []
-    for var in xrds.data_vars:
-        if set(xrds[var].dims).issubset(dimension_order) and len(xrds[var].dims) >= 1:
+    for var in ds_out.data_vars:
+        if set(ds_out[var].dims).issubset(dimension_order) and len(ds_out[var].dims) >= 1:
[…]
