### 2025-10-31 3aac3a8
catch & log gvec errors
