+++ src/pygvec/binding.f90
@@ -20,7 +20,7 @@ CONTAINS
 !================================================================================================================================!
 SUBROUTINE redirect_stdout(filename)
   ! MODULES
-  USE MODgvec_Globals, ONLY: Unit_stdOut, UNIT_errOut,abort
+  USE MODgvec_Globals, ONLY: Unit_stdOut, abort
   ! INPUT/OUTPUT VARIABLES ------------------------------------------------------------------------------------------------------!
   CHARACTER(LEN=*), INTENT(IN) :: filename
   ! LOCAL VARIABLES -------------------------------------------------------------------------------------------------------------!
@@ -37,7 +37,7 @@ END SUBROUTINE redirect_stdout
 !================================================================================================================================!
 SUBROUTINE flush_stdout()
   ! MODULES
-  USE MODgvec_Globals, ONLY: Unit_stdOut, UNIT_errOut,abort
+  USE MODgvec_Globals, ONLY: Unit_stdOut, abort
   ! LOCAL VARIABLES -------------------------------------------------------------------------------------------------------------!
   INTEGER :: ios
   ! CODE ------------------------------------------------------------------------------------------------------------------------!
@@ -51,9 +51,10 @@ END SUBROUTINE flush_stdout
 !================================================================================================================================!
 SUBROUTINE redirect_abort()
   ! MODULES
-  USE MODgvec_Globals, ONLY: RaiseExceptionPtr
+  USE MODgvec_Globals, ONLY: RaiseExceptionPtr, print_backtrace
   ! CODE ------------------------------------------------------------------------------------------------------------------------!
   RaiseExceptionPtr => F90WRAP_ABORT
+  print_backtrace = .FALSE.
 END SUBROUTINE redirect_abort
 
 END MODULE MODgvec_py_binding
