+++ python/gvec/core/run.py
@@ -18,6 +18,7 @@ from pandas import read_csv
 
 import gvec
 from gvec.core.state import State
+from gvec.errors import catch_gvec_errors
 from gvec.util import CaseInsensitiveDict as cidict
 from gvec.lib import modgvec_py_run as _run
 from gvec.lib import modgvec_py_binding as _binding
@@ -178,11 +179,11 @@ class Run:
         if "I_tor" in self.parameters and picard_current != "off":
             self.curr_constraint = True
         elif "I_tor" in self.parameters and picard_current == "off":
-            raise KeyError(
+            raise gvec.errors.MissingParameterError(
                 "'I_tor' is provided but 'picard_current' is set to 'off' or not provided. Please provide a valid 'picard_current', e.g. 'auto'."
             )
         elif "I_tor" not in self.parameters and picard_current != "off":
-            raise KeyError(
+            raise gvec.errors.MissingParameterError(
                 "Expected 'I_tor' in the parameters since 'picard_current' is not 'off'."
                 + " Please set 'picard_current' to 'off' if you want to use a fixed 'iota' profile or provide 'I_tor'."
             )
@@ -191,7 +192,9 @@ class Run:
 
         # Automatically generate the stages for the current optimization
         if "stages" in self.parameters and picard_current == "auto":
-            raise ValueError("Picard current is set to 'auto' but 'stages' is specified!")
+            raise gvec.errors.InvalidParameterError(
+                "Picard current is set to 'auto' but 'stages' is specified!"
+            )
         if "stages" in self.parameters:
             self.stages = self.parameters["stages"]
             del self.parameters["stages"]
@@ -305,7 +308,7 @@ class Run:
 
         Raises
         ------
-        ValueError
+        gvec.errors.InvalidParameterError
             If an unknown profile type is provided.
         """
         if (
@@ -326,7 +329,7 @@ class Run:
                 if (
                     abs(coefs[-1]) > 1e-8
                 ):  # poly1d is reverse to GVEC, e.g. coefs is ordered x²+x+1
-                    raise ValueError(
+                    raise gvec.errors.InvalidParameterError(
                         f"Toroidal current profile not zero at magnetic axis!  I_tor(rho=0): {coefs[-1]}"
                     )
 
@@ -340,7 +343,7 @@ class Run:
                 I_tor_bspl = BSpline(knots, coefs, deg)
                 self.I_tor_target = I_tor_bspl(self.rho**2)
                 if abs(I_tor_bspl(0.0)) > 1e-8:
-                    raise ValueError(
+                    raise gvec.errors.InvalidParameterError(
                         f"Toroidal current profile not zero at magnetic axis! I_tor(rho=0): {I_tor_bspl(0.0)}"
                     )
 
@@ -355,12 +358,14 @@ class Run:
                 I_tor_bspl = make_splrep(rho2_vals, y_vals)
                 self.I_tor_target = I_tor_bspl(self.rho**2)
                 if abs(I_tor_bspl(0.0)) > 1e-8:
-                    raise ValueError(
+                    raise gvec.errors.InvalidParameterError(
                         f"Toroidal current profile not zero at magnetic axis! I_tor(rho=0): {I_tor_bspl(0.0)}"
                     )
 
             case _:
-                raise ValueError(f"Unknown Itor type: {params['I_tor']['type']}")
+                raise gvec.errors.InvalidParameterError(
+                    f"Unknown Itor type: {params['I_tor']['type']}"
+                )
 
     def run_single_minimization(self):
         """Run a single GVEC energy minimization using the current parameters. The run-state is updated after the run."""
@@ -559,7 +564,7 @@ class Run:
                     and "I_tor" not in self._state_parameters
[…]
