### 2025-11-03 3ed7cc1
fix CAS3D test & add GIST test
