### 2025-11-03 6e3ae15
fixed parameter readin of integer.
* ifx compiler truncates a real to an integer with `READ(parameterfile,*)`, and does not throw an error.
* now, for integers, we use a formatted read instead: `READ(parameterfile,'(I8)')
* for arrays, one must first read to an array of strings, then parse for integer.
