+++ src/globals/readintools.f90
@@ -207,7 +207,7 @@ IF (PRESENT(Proposal)) THEN
 ELSE
   CALL FindStr(Key,HelpStr,DefMsg)
 END IF
-READ(HelpStr,*,IOSTAT=ioerr)GetInt
+READ(HelpStr,'(I8)',IOSTAT=ioerr)GetInt
 IF(ioerr.NE.0)THEN
   CALL abort(__STAMP__, &
        "Problem reading parameter '"//TRIM(key)//"', expected integer, got '= "//TRIM(helpStr)//"'", &
@@ -226,8 +226,8 @@ END FUNCTION GETINT
 
 
 !===================================================================================================================================
-!> Read real named "key" from setup file and store in "GETINT". If keyword "Key" is not found in ini file,
-!! the default value "Proposal" is used for "GETINT" (error if "Proposal" not given).
+!> Read real named "key" from setup file and store in "GETREAL". If keyword "Key" is not found in ini file,
+!! the default value "Proposal" is used for "GETREAL" (error if "Proposal" not given).
 !! Ini file was read in before and is stored as list of character strings starting with "FirstString".
 !!
 !===================================================================================================================================
@@ -276,8 +276,8 @@ END FUNCTION GETREAL
 
 
 !===================================================================================================================================
-!> Read logical named "key" from setup file and store in "GETINT". If keyword "Key" is not found in ini file,
-!! the default value "Proposal" is used for "GETINT" (error if "Proposal" not given).
+!> Read logical named "key" from setup file and store in "GETLOGICAL". If keyword "Key" is not found in ini file,
+!! the default value "Proposal" is used for "GETLOGICAL" (error if "Proposal" not given).
 !! Ini file was read in before and is stored as list of character strings starting with "FirstString".
 !!
 !===================================================================================================================================
@@ -347,6 +347,7 @@ INTEGER                   :: GetIntArray(nIntegers)      !! Integer array read f
 !-----------------------------------------------------------------------------------------------------------------------------------
 ! LOCAL VARIABLES
 CHARACTER(LEN=MAXLEN)           :: HelpStr,ProposalStr
+CHARACTER(LEN=8)                :: tmpstrarray(nIntegers)
 CHARACTER(LEN=8)                :: DefMsg,tmpstr
 INTEGER                         :: iInteger
 INTEGER                         :: ioerr
@@ -368,12 +369,21 @@ IF(iInteger.NE.nIntegers)THEN
        "Problem reading parameter '"//TRIM(key)//"', expected integer array of size "//TRIM(tmpstr)//", got '= "//TRIM(helpStr)//"'", &
        TypeInfo="InvalidParameterError")
 END IF
-READ(HelpStr,*,IOSTAT=ioerr)GetIntArray
+
+READ(HelpStr,*,IOSTAT=ioerr)tmpstrarray
 IF(ioerr.NE.0)THEN
   CALL abort(__STAMP__,&
        "Problem reading parameter '"//TRIM(key)//"', expected integer array, got '= "//TRIM(helpStr)//"'", &
        TypeInfo="InvalidParameterError")
 END IF
+DO iInteger=1,nIntegers
[…]
