### 2025-11-03 86eb43a
fix testGIST
remove `--plot` as this requires matplotlib
