### 2025-11-03 ac2b76b
add L_gradB, well & mirror_ratio
