### 2025-11-03 7a98463
add a reader for the G-frame netcdf file, with test.
