+++ python/gvec/scripts/quasr.py
@@ -54,6 +54,7 @@ from collections.abc import Sequence
 import logging
 
 import numpy as np
+import xarray as xr
 from scipy.optimize import root_scalar
 
 from gvec.util import write_parameters, logging_setup
@@ -500,38 +501,16 @@ def write_Gframe_ncfile(filename: str | Path, dict_in):
 
     for vecvar, vecval in zip(["axis/xyz", "axis/Nxyz", "axis/Bxyz"], ["xyz", "Nxyz", "Bxyz"]):
         assert np.all(dict_in["axis"][vecval].shape == (3, dict_in["axis"]["nzetaFull"])), (
-            f"shape of axis/{vecval} must be (3,nzetaFull_axis), but it is {dict_in['axis'][vecval].shape}"
+            f"shape of axis/{vecval} must be (3,{dict_in['axis']['nzetaFull']}), but it is {dict_in['axis'][vecval].shape}"
         )
         ncvars[vecvar + "_var"] = ncfile.createVariable(
             vecvar + "(::)", "f8", ("vec", "nzetaFull_axis")
         )
         ncvars[vecvar + "_var"][:, :] = dict_in["axis"][vecval]
 
-    boundary_ntheta = dict_in["boundary"]["ntheta"]
-    boundary_m_max = (boundary_ntheta - 1) // 2
-    boundary_nzeta = dict_in["boundary"]["nzeta"]  # same for now
-    boundary_n_max = (boundary_nzeta - 1) // 2
-
-    boundary_lasym = 1 * dict_in["boundary"]["lasym"]
-
-    for ivar, ival in zip(
-        [
-            "boundary/ntheta",
-            "boundary/nzeta",
-            "boundary/m_max",
-            "boundary/n_max",
-            "boundary/lasym",
-        ],
-        [
-            boundary_ntheta,
-            boundary_nzeta,
-            boundary_m_max,
-            boundary_n_max,
-            boundary_lasym,
-        ],
-    ):
-        ncvars[ivar + "_var"] = ncfile.createVariable(ivar, "i8")
-        ncvars[ivar + "_var"].assignValue(ival)
+    for ivar in ["ntheta", "nzeta", "m_max", "n_max", "lasym"]:
+        ncvars["boundary/" + ivar + "_var"] = ncfile.createVariable("boundary/" + ivar, "i8")
+        ncvars["boundary/" + ivar + "_var"].assignValue(1 * dict_in["boundary"][ivar])
 
     ncfile.createDimension("ntheta_boundary", dict_in["boundary"]["ntheta"])
 
@@ -548,9 +527,10 @@ def write_Gframe_ncfile(filename: str | Path, dict_in):
     ncvars["zeta_var"][:] = dict_in["boundary"]["zeta"]
 
     for vecvar, vecval in zip(["boundary/X", "boundary/Y"], ["X1", "X2"]):
-        assert np.all(dict_in["boundary"][vecval].shape == (boundary_ntheta, boundary_nzeta)), (
-            f"shape of boundary/{vecval} must be (ntheta_boundary,nzeta_boundary)"
-        )
+        assert np.all(
+            dict_in["boundary"][vecval].shape
+            == (dict_in["boundary"]["ntheta"], dict_in["boundary"]["nzeta"])
+        ), f"shape of boundary/{vecval} must be (ntheta_boundary,nzeta_boundary)"
         ncvars[vecvar + "_var"] = ncfile.createVariable(
             vecvar + "(::)", "f8", ("ntheta_boundary", "nzeta_boundary")
         )
@@ -605,6 +585,55 @@ def write_Gframe_ncfile(filename: str | Path, dict_in):
[…]
