### 2025-11-03 baf4f3c
add some boundary surface parameters as test cases
