+++ python/gvec/util.py
@@ -962,3 +962,173 @@ def compute_FD(f: np.ndarray, pos, coefs, axis=0):
     for roll, c in zip(pos[1:], coefs[1:]):
         df += np.roll(f, -roll, axis=axis) * c
     return df
+
+
+def boundary_generator(case: str, nfp, X1_00=1.0, a0=0.5, ellipticity=0.4, helix_r=0.5):
+    """
+    Define parameters for some simple boundaries for testing.
+    To evaluate the boundary at the given (theta, zeta) points, use
+    `evaluate_boundary(theta,zeta,boundary_generator(name,nfp))`
+    Args:
+        case (str): The name of the boundary:
+                    - "ellip_cyl" : elliptic cross-section with no change in zeta
+                    - "ellip_cyl_breathe" : elliptic cross-section with cross-section area changing with zeta
+                    - "ellip_cyl_rot" : elliptic cross section of constant ellipticity that only rotates with zeta
+                    - "ellip_cyl_rot2" : cross section of constant ellipticity that only rotates with zeta, but theta=0 origin remains on positive X1 direction
+                    - "ellip_cyl_stretch" : elliptic cross section of  changing ellipticity with zeta, not orientation of theta=0
+                    # "ellip_cyl_helix":  cross section of constant ellipticity, where the axis/boundary moves like a helix over zeta
+                    # "ellip_cyl_helix_rot": cross section of constant ellipticity, where the axis/boundary moves like a helix, and rotates along it, over zeta.
+                    # "ellip_cyl_helix_rot2": cross section of constant ellipticity, where the axis/boundary moves like a helix, and rotates along it, over zeta, but theta=0 origin remains on positive X1 direction
+
+        X1_00 : =major radius if $X^1=R$
+        a0    : =minor radius scale
+        ellipticity : =ellipticity of the cross section
+    Returns:
+        parameter dictionary describing $X^1$ and $X^2$
+    """
+    assert case in ["ellip_cyl", "ellip_cyl_breathe"]
+    params = {"nfp": nfp}
+    if case == "ellip_cyl":
+        params["X1_mn_max"] = [1, 0]
+        params["X2_mn_max"] = [1, 0]
+        params["X1_a_cos"] = {(0, 0): X1_00}
+        params["X1_b_cos"] = {
+            (0, 0): X1_00,
+            (1, 0): a0 * (1.0 - ellipticity),
+        }
+        params["X2_b_sin"] = {
+            (1, 0): a0 * (1.0 + ellipticity),
+        }
+    if case == "ellip_cyl_breathe":
+        breathe = 0.1
+        params["X1_mn_max"] = [1, 1]
+        params["X2_mn_max"] = [1, 1]
+        params["X1_a_cos"] = {(0, 0): X1_00}
+        params["X1_b_cos"] = {
+            (0, 0): X1_00,
+            (1, 0): a0 * (1.0 - ellipticity) * (1 + breathe),
+            (1, -1): -a0 * (1.0 - ellipticity) * 0.5 * breathe,
+            (1, 1): -a0 * (1.0 - ellipticity) * 0.5 * breathe,
+        }
+        params["X2_b_sin"] = {
+            (1, 0): a0 * (1.0 + ellipticity) * (1 + breathe),
+            (1, -1): -a0 * (1.0 + ellipticity) * 0.5 * breathe,
+            (1, 1): -a0 * (1.0 + ellipticity) * 0.5 * breathe,
+        }
+    if case == "ellip_cyl_rot":
[…]
