### 2025-11-03 9613683
Minimal corrections
