### 2025-11-04 ee2916b
adjust post-tests / fix testGIST
