### 2025-11-04 406b884
use aux dataset for vacuum well
this ensures that the correct points are used for the edge and (almost) axis position
