### 2025-11-06 af553ff
new aspect_ratio & effective elongation
* compute V, A_surface, L_axis
  * independent of coordinate frame
  * L_axis does depend on the equilibrium solution (axis position)
* r_maj, r_min, elongation, aspect_ratio are determined by an equivalent torus with elliptical cross section and circular axis
