### 2025-11-06 9f08efe
add iota_avg2, shear_avg & shear_avg2
