+++ test-CI/unit-pygvec/test_compute.py
@@ -631,8 +631,8 @@ def test_volume_integral(teststate, ev_rtz_int, ev_rtz):
         "V",
         "dV_dPhi_n",
         "dV_dPhi_n2",
-        "minor_radius",
-        "major_radius",
+        "aspect_ratio",
+        "elongation",
         "iota_avg",
         "iota_curr",
         "iota_0",
@@ -662,8 +662,8 @@ def test_iota_curr(teststate, ev_rtz_int):
         "V",
         "dV_dPhi_n",
         "dV_dPhi_n2",
-        "minor_radius",
-        "major_radius",
+        "aspect_ratio",
+        "elongation",
         "iota_avg",
         "iota_curr",
         "iota_0",
