### 2025-11-07 3065b6c
add shift to fourier module, tests with fixes.
* fourier fft1d and fft2d can get a shift parameter, such that periodic data that was sampled at angles which do not start at 0 can be still tranformed to the modes on [0,2pi]
* add shift_1d function to shift the data, via using numpy rfft1d
* add tests for these
* fix xarray call when loading data from a file. instead of `xr.dataset_open`, one should use `xr.dataset_load`, such that the file is not left open.
