### 2025-11-07 c86960e
re-allow fac_nyq=-1
