### 2025-11-07 3a0f635
overhauled quasr script. new options, run in tests
* BUGFIX: Mmax,Nmax in output and parameterfile were interchanged!
* more functionality of the convert_quasr script:
  * input surface can have even number of points, and points sampled on halfgrid (changed to odd and grid at [0,2pi) )
  * input surface can be filtered with an extra tolerance, called "clean". Will change the surface, and a maximum distance to the original boundary can be given
  * input surface can be forced to be stellarator-symmetric
  * cutoff number of modes in G-frame. before construction of the G-Frame, toroidal surface 
[…]
