+++ python/gvec/scripts/quasr.py
@@ -85,13 +85,34 @@ verbosity.add_argument(
     help="verbosity level: -v for info, -vv for debug",
 )
 verbosity.add_argument("-q", "--quiet", action="store_true", help="suppress output")
-parser.add_argument("--nt", type=int, help="number of theta points (only for ID or -s)")
-parser.add_argument("--nz", type=int, help="number of zeta points (only for ID or -s)")
+parser.add_argument(
+    "--nt", type=int, help="number of theta points, must be odd (only for ID or -s)"
+)
+parser.add_argument(
+    "--nz", type=int, help="number of zeta points, must be odd (only for ID or -s)"
+)
 parser.add_argument(
     "--tol",
     type=float,
     default=1e-8,
-    help="tolerance for determining minimal necessary (M, N)",
+    help="tolerance for determining minimal necessary (M, N) for the output Fourier modes of X1,X2. default is 1e-8",
+)
+parser.add_argument(
+    "--clean",
+    type=float,
+    default=0.0,
+    help="tolerance for determining minimal necessary (M, N) for the input surface. Default is 0., which means no cleaning.",
+)
+parser.add_argument(
+    "--symm",
+    action="store_true",
+    help="if set, imposes stellarator symmetry for the input surface. Use this with great care!",
+)
+parser.add_argument(
+    "--cutoff",
+    type=int,
+    default=-1,
+    help="cutoff toroidal mode number only for G-Frame construction, reduces the number of Fourier modes in the G-Frame. Default is -1, which means no cutoff.",
 )
 param_type = parser.add_mutually_exclusive_group()
 param_type.add_argument(
@@ -127,10 +148,14 @@ def check_args(parser, args):
             args.nt = 81
         elif args.nt < 1:
             raise parser.error("Number of theta points must be greater than 0.")
+        elif args.nt % 2 == 0:
+            raise parser.error("Number of theta points (nt) must be odd.")
         if args.nz is None:
             args.nz = 81
         elif args.nz < 1:
             raise parser.error("Number of zeta points must be greater than 0.")
+        elif args.nz % 2 == 0:
+            raise parser.error("Number of zeta points (nz) must be odd.")
     else:
         if not args.file.exists():
             raise parser.error(f"File {args.file} does not exist.")
@@ -142,8 +167,10 @@ def check_args(parser, args):
             raise parser.error(
                 "Number of zeta points cannot manually be set with a boundary file."
             )
-    if args.tol is not None and args.tol <= 0:
+    if args.tol is not None and args.tol <= 0.0:
         raise parser.error("Tolerance must be greater than 0.")
+    if args.clean is not None and args.clean < 0.0:
+        raise parser.error("Cleaning tolerance must be greater than 0.")
     if args.param_type is None:
         args.param_type = "toml"
 
@@ -152,11 +179,16 @@ def check_args(parser, args):
 
 
 def check_field_periodicity(xyz: np.ndarray, nfp: int, atol=1e-12):
+    """
+    checks if all  xyz positions of the surface on a full turn, xyz[0:nz*nfp,0:nt,0:2], have the field periodicity with nfp.
[…]
