### 2025-11-08 4720688
Finalizing G-frame interface. Cutoff works!
* added the cutoff of toroidal modes, but only for the construction of the G-Frame. Seems to work very well!
* for the input surface in `gframe.construct_gframe_from_surface`, allow for even number of points and also shifted origin... origin is then changed, together with upsampling to an odd number.
* functions from quasr.py moved to new `gvec/gframe.py` file: all scripts related to construction of the G-Frame and read and write of the netcdf file.
* updated the documentation with the new parameters!
* some fixes in the visualization
