+++ python/gvec/scripts/quasr.py
@@ -55,17 +55,16 @@ import logging
 
 import numpy as np
 import xarray as xr
-from scipy.optimize import root_scalar
 
-from gvec.util import write_parameters, logging_setup
-from gvec import fourier
+from gvec.util import logging_setup
+from gvec import gframe
 
 # === Argument Parser === #
 
 parser = argparse.ArgumentParser(
     prog="pygvec-load-quasr",
     description="Load a QUASR configuration and convert it to a G-Frame and boundary for use with GVEC.",
-    usage="%(prog)s [-h] (ID | -s FILE | -f FILE) [-v | -q] [--nt NT] [--nz NZ] [--tol TOL] [--yaml | --toml] [--save-xyz]",
+    usage="%(prog)s [-h] (ID | -s FILE | -f FILE) [-v | -q] [--nt NT] [--nz NZ] [--clean=CLEANTOL] [--symm] [--cutoff=CUTOFF] [--tol=TOL] [--yaml | --toml] [--save-xyz]",
 )
 parser.add_argument("ID", type=int, nargs="?", help="ID of the QUASR configuration")
 parser.add_argument(
@@ -86,22 +85,20 @@ verbosity.add_argument(
 )
 verbosity.add_argument("-q", "--quiet", action="store_true", help="suppress output")
 parser.add_argument(
-    "--nt", type=int, help="number of theta points, must be odd (only for ID or -s)"
+    "--nt",
+    type=int,
+    help="number of theta points, must be odd (only for ID or -s). Default is 81.",
 )
 parser.add_argument(
-    "--nz", type=int, help="number of zeta points, must be odd (only for ID or -s)"
-)
-parser.add_argument(
-    "--tol",
-    type=float,
-    default=1e-8,
-    help="tolerance for determining minimal necessary (M, N) for the output Fourier modes of X1,X2. default is 1e-8",
+    "--nz",
+    type=int,
+    help="number of zeta points, must be odd (only for ID or -s). Default is 81.",
 )
 parser.add_argument(
     "--clean",
     type=float,
     default=0.0,
-    help="tolerance for determining minimal necessary (M, N) for the input surface. Default is 0., which means no cleaning.",
+    help="tolerance to reduce necessary Fourier modes (M, N) for the input surface. Default is 0., which means no cleaning.",
 )
 parser.add_argument(
     "--symm",
@@ -113,6 +110,12 @@ parser.add_argument(
     type=int,
     default=-1,
     help="cutoff toroidal mode number only for G-Frame construction, reduces the number of Fourier modes in the G-Frame. Default is -1, which means no cutoff.",
+)
+parser.add_argument(
+    "--tol",
+    type=float,
+    default=1e-8,
+    help="tolerance for determining minimal necessary (M, N) for the output Fourier modes of X1,X2. default is 1e-8",
 )
 param_type = parser.add_mutually_exclusive_group()
 param_type.add_argument(
@@ -176,145 +179,6 @@ def check_args(parser, args):
 
 
 # === Functions === #
-
-
-def check_field_periodicity(xyz: np.ndarray, nfp: int, atol=1e-12):
-    """
-    checks if all  xyz positions of the surface on a full turn, xyz[0:nz*nfp,0:nt,0:2], have the field periodicity with nfp.
[…]
