### 2025-11-10 5e1fc70
add generation of boundary xyz data from G-Frame file
