### 2025-11-11 4c43b70
further extensions of the gframe module.
* add a `to_RZ` function, which computes the RZ cross-sections of input xyz surface data. returns a dictionary, with truncated fourier modes given by a tolerance and stellarator-symmetry check.
* extended the `to_surface` function. it computes the xyz surface object from a G-frame file, but now returns a dictionary, with truncated fourier modes given by a tolerance and stellarator-symmetry check.
* visualization with a cross-section comparison of RZ and X1,X2 planes, visualized from the data of the dictionaries returned by `to_surface` and `to_RZ`
* fix
[…]
