+++ python/gvec/gframe.py
@@ -233,7 +233,7 @@ def find_zeta_cuts(zeta_in, origin, normal, xyz, dft_dict, zeta_bracket: float):
     def eval_dist(zeta_in):
         return eval_distance_to_curve(zeta_in, origin, normal, xyz, dft_dict).item()
 
-    for factor in [0.01, 0.1, 1]:
+    for factor in [0.1, 0.5, 0.99]:
         try:
             return root_scalar(
                 eval_dist,
@@ -243,10 +243,20 @@ def find_zeta_cuts(zeta_in, origin, normal, xyz, dft_dict, zeta_bracket: float):
             ).root
         except ValueError:
             pass
-    raise RuntimeError("Could not find zeta cuts with the given bracket (or 1/10, 1/100)")
+    raise RuntimeError(
+        f"Could not find zeta cuts with the given bracket (or 1/2,1/10), for initial guess zeta_in={zeta_in}"
+    )
 
 
 def get_xyz_cut(zeta_start, origins, normals, xyz_in, dft_dict, nfp):
+    """
+
+    Inputs:
+        zeta_start: initial guess for zeta position for each cut shape [nz_out]
+        origins: origin of the cutting plane, shape [nz_out, 3]
+        normals: normal of the cutting plane  [nz_out,3]
+        dft_dict: dictionary containing the dft from the zeta points of xyz_in, to be used for evaluation of theta=const curves at arbitrary zeta.
+    """
     nz_out = origins.shape[0]
     nt = xyz_in.shape[1]
     zeta_out = np.zeros(nz_out)
@@ -806,15 +816,23 @@ def minimal_modes(X, Y, Z=None, tolerance=1e-8):
     return mcan[mask].item(), ncan[mask].item()
 
 
-def to_surface(dict_in: dict, nzeta: int = 81, ntheta: int = 81):
+def to_surface(dict_in: dict, nzeta: int = 81, ntheta: int = 81, tolerance: float = 1e-8):
     """
     Convert a gframe file with axis+boundary to boundary surface in cartesian coordinates.
     Input:
         dict_in: dictionary of the Gframe netcdf file, from `gvec.gframe.read_Gframe_ncfile(filename)`
-        ntheta: number of theta points
-        nzeta: number of zeta points in one field period
+        nzeta: number of zeta positions for the output, to sample on one field period
+        ntheta: number of theta positions for the output
     Output:
+    dictionary with:
         xyz: boundary surface in cartesian coordinates, with shape [0:nzeta*nfp,0:ntheta,0:2]
+        X1, X2: boundary in G-Frame, shape [0:ntheta,0:nzeta]
+        zetafull: zeta values of the boundary surface
+        theta: theta values of the boundary surface
+        lasym : logical for asymmetry, =false if stellarator symmetry is found
+        nfp : number of field periods
+        Mmax, Nmax: maximum mode numbers needed for the given tolerance
+        X1c, X1s, X2c, X2s: boundary  modes in G-Frame, up to Mmax, Nmax
     """
     nfp = dict_in["nfp"]
     theta_out = np.linspace(0, 2 * np.pi, ntheta, endpoint=False)
[…]
