### 2025-11-13 15bb7b3
add derivations for well, aspect ratio & elongation
