### 2025-11-13 33115c4
add compiler to version info
also make version info available from within python
