### 2025-11-14 c2ea630
changes requested in the MR.
* renamed parameter `symm` to `stellsym`
* docstring and match case for `boundary_generator`
* add gframe.py to __init__.py , made version number 301 consistently.
* add writefile flag for `construct_gframe_from_surface`, to allow to switch it off.
* FIX: in `construct_gframe_from_surface`:
  cleanup of surface modes will now NOT reduce number of points, making the cuts more accurate, and thus allowing more modes in X1,X2 than in xyz.
  This can happen when the cutoff for the Gframe modes is set too small, and becomes therefore a useful information.
