+++ python/gvec/util.py
@@ -1001,168 +1001,176 @@ def ellipse_circumference_factor(epsilon: float) -> float:
 def boundary_generator(case: str, nfp, X1_00=1.0, a0=0.5, ellipticity=0.4, helix_r=0.5):
     """
     Define parameters for some simple boundaries for testing.
-    To evaluate the boundary at the given (theta, zeta) points, use
-    `evaluate_boundary(theta,zeta,boundary_generator(name,nfp))`
-    Args:
-        case (str): The name of the boundary:
-                    - "ellip_cyl" : elliptic cross-section with no change in zeta
-                    - "ellip_cyl_breathe" : elliptic cross-section with cross-section area changing with zeta
-                    - "ellip_cyl_rot" : elliptic cross section of constant ellipticity that only rotates with zeta
-                    - "ellip_cyl_rot2" : cross section of constant ellipticity that only rotates with zeta, but theta=0 origin remains on positive X1 direction
-                    - "ellip_cyl_stretch" : elliptic cross section of  changing ellipticity with zeta, not orientation of theta=0
-                    # "ellip_cyl_helix":  cross section of constant ellipticity, where the axis/boundary moves like a helix over zeta
-                    # "ellip_cyl_helix_rot": cross section of constant ellipticity, where the axis/boundary moves like a helix, and rotates along it, over zeta.
-                    # "ellip_cyl_helix_rot2": cross section of constant ellipticity, where the axis/boundary moves like a helix, and rotates along it, over zeta, but theta=0 origin remains on positive X1 direction
 
-        X1_00 : =major radius if $X^1=R$
-        a0    : =minor radius scale
-        ellipticity : =ellipticity of the cross section
-    Returns:
-        parameter dictionary describing $X^1$ and $X^2$
+    Parameters
+    ----------
+    case : str
+        The name of the boundary:
+            - "ellip_cyl" : elliptic cross-section with no change in zeta
+            - "ellip_cyl_breathe" : elliptic cross-section with cross-section area changing with zeta
+            - "ellip_cyl_rot" : elliptic cross section of constant ellipticity that only rotates with zeta
+            - "ellip_cyl_rot2" : cross section of constant ellipticity that only rotates with zeta, but theta=0 origin remains on positive X1 direction
+            - "ellip_cyl_stretch" : elliptic cross section of  changing ellipticity with zeta, not orientation of theta=0
+            - "ellip_cyl_helix":  cross section of constant ellipticity, where the axis/boundary moves like a helix over zeta
+            - "ellip_cyl_helix_rot": cross section of constant ellipticity, where the axis/boundary moves like a helix, and rotates along it, over zeta
+            - "ellip_cyl_helix_rot2": cross section of constant ellipticity, where the axis/boundary moves like a helix, and rotates along it, over zeta, but theta=0 origin remains on positive X1 direction
+
+    X1_00 : float, optional
+        =major radius if $X^1=R$
+    a0 : float, optional
+        =minor radius scale
+    ellipticity : float, optional
+        =ellipticity of the cross section
+
+    Returns
+    -------
+    parameter dictionary describing $X^1$ and $X^2$
     """
[…]
