### 2025-11-14 34a2982
add mercier criterion
