+++ docs/dev/derivations.md
@@ -170,6 +170,56 @@ positive values of $d_{well}$ indicate $\frac{d^2V}{d\Phi_n^2} < 0$ which is fav
 
 $\Rightarrow$ `vacuum_magnetic_well_depth`
 
+## Mercier criterion
+
+We follow the formulas reported in *Landreman & Jorge* (2020), given in *Bauer et al.* (1984); *Ichiguchi et al.* (1993).
+
+$$
+D_\text{Merc} = D_\text{M,Shear} + D_\text{M,Curr} + D_\text{M,Well} + D_\text{M,Geod},
+$$
+
+where a positive value of $D_\text{Merc}$ indicates stability and
+
+$$
+\begin{align}
+D_\text{M,Shear} &= \frac{1}{16\pi^2} \pqty{\dv{\iota}{\Phi}}^2 \\
+% D_\text{M,Shear} &= \frac{1}{16\pi^2} \left( \frac{d\iota}{d\rho} \right)^2 \left( \frac{d\Phi}{d\rho} \right)^{-2} \\
+D_\text{M,Curr} &= -\frac{s_\chi}{\pqty{2\pi}^4} \dv{\iota}{\Phi} \int\dd S \frac{\vec{\Xi} \cdot \vec{B}}{\abs{\grad\Phi}^3} \\
+% D_\text{M,Curr} &= -\frac{s_\chi}{\left(2\pi\right)^4} \frac{d\iota}{d\rho} \left(\frac{d\Phi}{d\rho}\right)^{-1} \int dS \frac{\vec{\Xi} \cdot \vec{B}}{\left|\frac{d\Phi}{d\rho} \nabla\rho \right|^3} \\
+% D_\text{M,Curr} &= -\frac{s_\chi}{\left(2\pi\right)^4} \frac{d\iota}{d\rho} \left(\frac{d\Phi}{d\rho}\right)^{-1}\left|\frac{d\Phi}{d\rho}\right|^{-3} \int dS \frac{\vec{\Xi} \cdot \vec{B}}{\left|\nabla\rho\right|^3} \\
+% D_\text{M,Curr} &= -\frac{s_\chi s_\Phi}{\left(2\pi\right)^4} \frac{d\iota}{d\rho} \left(\frac{d\Phi}{d\rho}\right)^{-4} \int dS \frac{\vec{\Xi} \cdot \vec{B}}{\left|\nabla\rho\right|^3} \\
+D_\text{M,Well} &= \frac{\mu_0}{\pqty{2\pi}^6} \dv{p}{\Phi} \pqty{s_\Phi \dv[2]{V}{\Phi} - \mu_0 \dv{p}{\Phi} \int dS \frac{1}{\modB^2 \abs{\grad\Phi}}}\int\dd S \frac{\modB^2}{\abs{\grad\Phi}^3} \\
+D_\text{M,Geod} &= \frac{1}{\pqty{2\pi}^6} \bqty{
+    \pqty{\int\dd S \frac{\mu_0 \vec{J}\cdot\vec{B}}{\abs{\grad\Phi}^3}}^2
+   -\pqty{\int\dd S \frac{\modB^2}{\abs{\grad\Phi}^3}}
+    \pqty{\int\dd S \frac{\pqty{\mu_0 \vec{J}\cdot\vec{B}}^2}{\modB^2 \abs{\grad\Phi}^3}}
+} \\
+\end{align}
+$$
+
+with
+
+$$
+\begin{align}
+s_\chi &= \text{sgn}{\chi} \\
[…]
