### 2025-11-14 9427c22
use text in quantity symbols
