+++ python/gvec/quantities.py
@@ -203,7 +203,7 @@ for var, long_name, symbol in [
 
 
 @register(
-    attrs=dict(long_name="number of field periods", symbol=r"N_{FP}"),
+    attrs=dict(long_name="number of field periods", symbol=r"N_\text{FP}"),
 )
 def N_FP(ds: xr.Dataset, state: State):
     ds["N_FP"] = state.nfp
@@ -912,7 +912,7 @@ for v in [
     requirements=("B", "e_theta", "dLA_dt", "iota", "B_theta_avg", "B_zeta_avg"),
     attrs=dict(
         long_name="poloidal derivative of the Boozer potential computed from the magnetic field",
-        symbol=r"\left." + latex_partial(r"\nu_B", "t") + r"\right|_{def.}",
+        symbol=r"\left." + latex_partial(r"\nu_B", "t") + r"\right|_\text{def.}",
     ),
 )
 def dNU_B_dt(ds: xr.Dataset):
@@ -926,7 +926,7 @@ def dNU_B_dt(ds: xr.Dataset):
     requirements=("B", "e_zeta", "dLA_dz", "iota", "B_theta_avg", "B_zeta_avg"),
     attrs=dict(
         long_name="toroidal derivative of the Boozer potential computed from the magnetic field",
-        symbol=r"\left." + latex_partial(r"\nu_B", "z") + r"\right|_{def.}",
+        symbol=r"\left." + latex_partial(r"\nu_B", "z") + r"\right|_\text{def.}",
     ),
 )
 def dNU_B_dz(ds: xr.Dataset):
@@ -1407,7 +1407,7 @@ def dA(ds: xr.Dataset):
 
 
 @register(
-    attrs=dict(long_name="surface area", symbol=r"A_{surface}"),
+    attrs=dict(long_name="surface area", symbol=r"A_\text{surface}"),
 )
 def A_surface(ds: xr.Dataset, state: State):
     aux = state.evaluate("dA", rho=1.0, theta="int", zeta="int")
@@ -1415,7 +1415,7 @@ def A_surface(ds: xr.Dataset, state: State):
 
 
 @register(
-    attrs=dict(long_name="length of the magnetic axis", symbol=r"L_{axis}"),
+    attrs=dict(long_name="length of the magnetic axis", symbol=r"L_\text{axis}"),
 )
 def L_axis(ds: xr.Dataset, state: State):
     aux = state.evaluate("e_zeta", rho=0.0, theta=0.0, zeta="int")
@@ -1425,7 +1425,7 @@ def L_axis(ds: xr.Dataset, state: State):
 
 @register(
     requirements=("L_axis",),
-    attrs=dict(long_name="effective major radius", symbol=r"r_{maj,eff}"),
+    attrs=dict(long_name="effective major radius", symbol=r"r_\text{maj,eff}"),
 )
 def r_maj(ds: xr.Dataset):
     ds["r_maj"] = ds.L_axis / (2 * np.pi)
@@ -1433,7 +1433,7 @@ def r_maj(ds: xr.Dataset):
 
 @register(
     requirements=("V", "L_axis"),
-    attrs=dict(long_name="effective minor radius", symbol=r"r_{min,eff}"),
+    attrs=dict(long_name="effective minor radius", symbol=r"r_\text{min,eff}"),
 )
 def r_min(ds: xr.Dataset):
     ds["r_min"] = np.sqrt(ds.V / (np.pi * ds.L_axis))
@@ -1441,7 +1441,7 @@ def r_min(ds: xr.Dataset):
 
 @register(
     requirements=("r_maj", "r_min"),
-    attrs=dict(long_name="effective aspect ratio", symbol=r"a_{eff}"),
[…]
