### 2025-11-14 9452369
add gframe tests. some set to xfail.
* add test of the `construct_gframe_from_surface` and  `to_surface` functions.
  we can produce a RZphi boundary from known modes (from `util.generate_boundary`), then put this into the G-Frame algorithm, and should then recover the same boundary at the end, up to a tolerance.
* `_helix` cases do not yet work, so they are still run but marked as xfail, until we know what the problem is.
