+++ python/gvec/gframe.py
@@ -70,12 +70,17 @@ def rodrigues(
     """
     Rodrigues rotation function.
 
-    Parameters:
-    - xyz: cartesian point positions. if multidimensional, the LAST dimension must contain the cartesian components.
-    - angle: The rotation angle.
+    Parameters
+    ----------
+    xyz : ndarray
+        cartesian point positions. if multidimensional, the LAST dimension must contain the cartesian components.
+    angle : float
+        The rotation angle.
 
-    Returns:
-    - pos_rot: The rotated position vector (cartesian components).
+    Returns
+    -------
+    pos_rot : ndarray
+        The rotated position vector (cartesian components).
     """
     assert np.array(xyz.shape[-1]) == 3, (
         "last dimension must be the cartesian components surface positions!"
@@ -95,20 +100,24 @@ def rodrigues(
 
 def xyz_to_xyz_hat(xyz_in: np.ndarray, zeta: np.ndarray, sign_rot: float):
     """
-    change from cartesian xyz positions of the surface (can be a full torus or a single field period) on to 'hat' coordinates, which are periodic on field period.
-    One needs the corresponding 1d zeta positions, and sign of the angle, to compute the transform:
-    ```
-    xhat = x*cos(zeta)+y*sign*sin(zeta)
-    yhat = y*cos(zeta)-x*sign*sin(zeta)
-    zhat = z
-    ```
-    Input:
-    xyz_in: xyz positions of the surface, xyz[0:nz,0:nt,0:2], sampled at zeta positions, must exclude the endpoint
-    zeta: 1d array of zeta positions belonging to the surface (without endpoint), size [0:nz]
-    sign_rot: direction of zeta for the rotation into hat coordinates, +1 or -1
-    Returns:
-    xhat,yhat,zhat : hat coordinates, periodic on one field period, size [0:nz,0:nt,0:2].
+    Change from cartesian xyz positions of the surface (can be a full torus or a single field period) on to 'hat' coordinates,
+    which are periodic on field period.
+
+    Parameters
+    ----------
+    xyz_in : ndarray
+        xyz positions of the surface, xyz[0:nz,0:nt,0:2], sampled at zeta positions, must exclude the endpoint
+    zeta : ndarray
+        1d array of zeta positions belonging to the surface (without endpoint), size [0:nz]
+    sign_rot : float
+        direction of zeta for the rotation into hat coordinates, +1 or -1
+
+    Returns
+    -------
+    xhat, yhat, zhat : ndarray
+        hat coordinates, periodic on one field period, size [0:nz,0:nt,0:2]
     """
+
     assert sign_rot == -1 or sign_rot == 1, f"sign_rot must be -1 or 1, but is {sign_rot}"
     assert xyz_in.shape[0] == zeta.shape[0], (
         f"xyz and zeta must have the same length, but are {xyz_in.shape[0]} and {zeta.shape[0]} respectively"
@@ -125,20 +134,25 @@ def xyz_hat_to_xyz(
     xhat: np.ndarray, yhat: np.ndarray, zhat: np.ndarray, zeta: np.ndarray, sign_rot: float
 ):
     """
-    change from xyz 'hat' coordinates  cartesian xyz positions, with their associated zeta positions.
-    Inverse function of xyz_to_xyz_hat
-        ```
-    x = xhat*cos(zeta)-yhat*sign*sin(zeta)
-    y = yhat*cos(zeta)+xhat*sign*sin(zeta)
-    z = zhat
-    ```
+    Change from xyz 'hat' coordinates to cartesian xyz positions.
 
-    Inputs:
-    xhat,yhat,zhat : hat coordinates , size [0:nz,0:nt,0:2].
-    zeta: 1d array of zeta positions corresponding to xyz positions (without endpoint), size [0:nz]
[…]
