### 2025-11-14 26bfb58
small fixes for odd/even point numbers.
* added also test for odd/even number of nt and nz, and also shifted theta zeta starting points.
