+++ test-CI/unit-pygvec/test_gframe.py
@@ -2,7 +2,8 @@ import pytest
 
 try:
     import numpy as np
-    from gvec import util, gframe, fourier
+    import gvec
+    from gvec import gframe
 except ImportError:
     pass  # tests will be skipped via the `check_import` fixture
 
@@ -23,37 +24,59 @@ def mark_xfail(request):
 
 @pytest.mark.parametrize(
     "case",
-    util.boundary_generator_cases().keys(),
+    gvec.util.boundary_generator_cases().keys(),
 )
 @pytest.mark.parametrize("nfp", [1, 2, 3])
-def test_gframe_boundary(case, nfp):
+@pytest.mark.parametrize("nz_nt", [[20, 10], [21, 11]], ids=["even", "odd"])
+@pytest.mark.parametrize("shft", [0.0, 0.5], ids=["noshift", "halfshift"])
+def test_gframe_boundary(case, nz_nt, shft, nfp):
+    def eval_bnd_xyz(theta_in, zeta_in, bnd_params):
+        X1, X2 = gvec.util.evaluate_boundary(theta_in, zeta_in, bnd_params)
+        # simply use RZ hmap here:
+        xyz = np.zeros((len(zeta_in), len(theta_in), 3))
+        xyz[:, :, 0] = X1.T * np.cos(zeta_in[:, None])
+        xyz[:, :, 1] = -X1.T * np.sin(zeta_in[:, None])
+        xyz[:, :, 2] = X2.T
+        return xyz
+
     tol = 1e-8
-    bnd_params = util.boundary_generator(case)
+    bnd_params = gvec.util.boundary_generator(case)
+    M_in, N_in = bnd_params["X1_mn_max"]
     bnd_params["nfp"] = nfp
     ntheta = 11
     nzeta = 21
     theta = np.linspace(0, 2 * np.pi, ntheta, endpoint=False)
     zeta = np.linspace(0, 2 * np.pi / nfp, nzeta, endpoint=False)
-    M_in, N_in = bnd_params["X1_mn_max"]
-    X1_fp, X2_fp = util.evaluate_boundary(theta, zeta, bnd_params)
+
+    X1_fp, X2_fp = gvec.util.evaluate_boundary(theta, zeta, bnd_params)
     M, N = gframe.minimal_modes(X1_fp, X2_fp, tolerance=tol)
     assert M <= M_in and N <= N_in, (
         f"boundary surface modes (M={M},N={N}) are larger than they should be (M={M_in},N={N_in})"
     )
     zetafull = np.linspace(0, 2 * np.pi, nzeta * nfp, endpoint=False)
-    X1, X2 = util.evaluate_boundary(theta, zetafull, bnd_params)
-    # simply use RZ hmap here:
-    xyz = np.zeros((nzeta * nfp, ntheta, 3))
-    xyz[:, :, 0] = X1.T * np.cos(zetafull[:, None])
-    xyz[:, :, 1] = -X1.T * np.sin(zetafull[:, None])
-    xyz[:, :, 2] = X2.T
+    # generate data to compare with
+    xyz_target = eval_bnd_xyz(theta, zetafull, bnd_params)
+    # test odd/even input points and shifts
+    t = np.linspace(0, 2 * np.pi, nz_nt[1], endpoint=False)
+    z = np.linspace(0, 2 * np.pi, nz_nt[0] * nfp, endpoint=False)
+    t = t + shft * t[1]
+    z = z + shft * z[1]
+
+    xyz_in = eval_bnd_xyz(t, z, bnd_params)
 
     params, dict_out = gframe.construct_gframe_from_surface(
-        xyz, nfp, name=f"test_{case}", writeFiles=False, tolerance_clean_surface=tol
[…]
