### 2025-11-14 cbab212
skip test_gframe completely for CI tests
* we typically skip all unit tests if the gvec package is not available, as is the case when testing only the fortran part in the CI
* in test_gframe we need util functions already at test-collection
* util is provided as a fixture which modifies the PYTHONPATH to import the util module directly without gvec being installed
* the easiest workaround is to skip the whole test_gframe module
