### 2025-11-17 d5cd418
replace r_min, r_maj -> r_minor, r_major
