### 2025-11-17 63021ff
fix docs: geometric quantities
