+++ docs/dev/derivations.md
@@ -106,38 +106,38 @@ The gradient $\nabla\vec{B}$ can be used to compute the *magnetic gradient lengt
 
 ## effective geometric quantities
 
-The plasma volume $V$, surface are $A_{surface}$ and length of the magnetic axis $L_{axis}$ are defined as
+The plasma volume $V$, surface are $A_\text{surface}$ and length of the magnetic axis $L_\text{axis}$ are defined as
 
 $$
 \begin{align}
 V &= \int_0^1 d\rho \int_0^{2\pi} d\theta \int_0^{2\pi} d\zeta \Jac \\
-A_{surface} &= \left. \int_0^{2\pi} d\theta \int_0^{2\pi} d\zeta \left|\vec{e}_\theta \times \vec{e}_\zeta\right| \right|_{\rho=1}\\
-L_{axis} &= \left. \int_0^{2\pi} d\zeta \left|\vec{e}_\zeta\right| \right|_{\rho=0,\theta=0}\\
+A_\text{surface} &= \left. \int_0^{2\pi} d\theta \int_0^{2\pi} d\zeta \left|\vec{e}_\theta \times \vec{e}_\zeta\right| \right|_{\rho=1}\\
+L_\text{axis} &= \left. \int_0^{2\pi} d\zeta \left|\vec{e}_\zeta\right| \right|_{\rho=0,\theta=0}\\
 \end{align}
 $$
 
-All three are independent of the coordinate frame and parametrization, and $V,A_{surface}$ only depend on the boundary shape and not the full equilibrium.
+All three are independent of the coordinate frame and parametrization, and $V,A_\text{surface}$ only depend on the boundary shape and not the full equilibrium.
 
 From these we can compute the *effective minor radius*, *effective major radius*, *effective aspect ratio* and *effective elongation*, by relating them to an equivalent torus with elliptical cross section.
 In particular we define the equivalent torus as a torus with circular axis and constant elliptical cross-section, which has the same volume, surface area and axis length as the configuration of interest.
 
 $$
 \begin{align}
-V &= 2\pi^2 r_\text{minor}^2 r_\text{major} \\
-A_{surface} &= 4\pi^2 r_\text{minor} r_\text{major} \tilde{C}(E_{eff}) \\
-L_{axis} &= 2\pi r_\text{major}
+V &= 2\pi^2 r_\text{minor,eff}^2 r_\text{major,eff} \\
+A_\text{surface} &= 4\pi^2 r_\text{minor,eff} r_\text{major,eff} \tilde{C}(E_\text{eff}) \\
+L_\text{axis} &= 2\pi r_\text{major,eff}
 \end{align}
 $$
 
-with the effective elongation $E_{eff} := \frac{a}{b}$ defined as the ratio of the cross-sections semi-major axis and semi-minor axis ($a \geq b$).
+with the effective elongation $E_\text{eff} := \frac{a}{b}$ defined as the ratio of the cross-sections semi-major axis and semi-minor axis ($a \geq b$).
 The circumference of the ellipse does not have a closed form.
 We use Ramanujan's approximation
 
 $$
 \begin{align}
[…]
