### 2025-11-17 f6b7bc2
add s_Phi back into D_Merc_Well
