+++ python/gvec/quantities.py
@@ -1716,6 +1716,7 @@ def D_Merc(ds: xr.Dataset):
     diota_dPhi = ds.diota_dr / ds.dPhi_dr
     dp_dPhi = ds.dp_dr / ds.dPhi_dr
     s_chi = np.sign(ds.chi)
+    s_Phi = np.sign(ds.Phi)
     dV_dr = fluxsurface_integral(ds.Jac)
     dV_drr = fluxsurface_integral(ds.dJac_dr)
     d2V_dPhi2 = dV_drr / ds.dPhi_dr**2 - dV_dr * ds.dPhi_drr / ds.dPhi_dr**3
@@ -1732,7 +1733,7 @@ def D_Merc(ds: xr.Dataset):
     ds["D_Merc_Shear"] = 1 / (16 * np.pi**2) * diota_dPhi**2
     ds["D_Merc_Curr"] = -s_chi / twopi**4 * diota_dPhi * (JBint - dB_theta_avg_dPhi * B2int)
     ds["D_Merc_Well"] = (
-        ds.mu0 / twopi**6 * dp_dPhi * (d2V_dPhi2 - ds.mu0 * dp_dPhi * Bi2int) * B2int
+        ds.mu0 / twopi**6 * dp_dPhi * (s_Phi * d2V_dPhi2 - ds.mu0 * dp_dPhi * Bi2int) * B2int
     )
     ds["D_Merc_Geod"] = 1 / twopi**6 * (JBint**2 - B2int * JB2int)
     ds["D_Merc"] = (
