### 2025-11-17 61f16a4
automatically set `mn_nyq` depending on hmap.
* introduce a `hmap%n_max`, which is zero for `hmap_RZ` and `hmap_cyl`, and depends on the input for `hmap_knot`, `hmap_frenet` and `hmap_axisNB`.
* This now allows to automatically set the default mn_nyq in the startup of GVEC, to include `n_max` from the hmap
* additional modification in the gframe script:
  - axis n_max now correctly computed in `write_Gframe_ncfile`
  - do not set `mn_nyq` in the parameterfile, since we have now an automatic one.
  - if boundary is not stellarator-symmetric, also switch on non-stellarator-symmetric modes in GVE
[…]
