+++ src/hmap/hmap_knot.f90
@@ -117,6 +117,7 @@ FUNCTION hmap_knot_init_params(knot_kl,knot_R0,knot_delta) RESULT(sf)
   sf%l=REAL(knot_kl(2), wp)
   sf%R0=knot_R0
   sf%delta=knot_delta
+  sf%n_max=MAX(sf%k,sf%l)
 
   IF (.NOT.((sf%R0 - ABS(sf%delta)) > 0.0_wp)) THEN
      CALL abort(__STAMP__, &
