### 2025-11-17 5b1af07
Update 7 files
- /test-CI/examples/axisNB_N2-12/parameter.ini
- /test-CI/examples/axisNB_N2-63/parameter.ini
- /test-CI/examples/axisNB_N3_knot/parameter.ini
- /test-CI/examples/axisNB_N3-test/parameter.ini
- /test-CI/examples/axisNB_QH_nfp7/parameter.ini
- /test-CI/examples/hmap_knot/parameter.ini
- /test-CI/examples/frenet_QH_nfp7_untwisted/parameter.ini
