+++ test-CI/examples/frenet_QH_nfp7_untwisted/parameter.ini
@@ -31,6 +31,9 @@ LA_deg    = 5
 LA_mn_max = (/ 4, 15/)
 LA_sin_cos = _sin_
 
+fac_nyq=-1
+mn_nyq=(/17,137/)
+
 nfp = 7 !still needed in older gvec versions
 
 which_hmap = 20
