+++ test-CI/examples/axisNB_N2-63/parameter.ini
@@ -32,6 +32,9 @@ X2_mn_max = (/ 1, 4/)
 LA_mn_max = (/ 1, 4/)
 !LA_sin_cos = _sin_
 
+fac_nyq=-1
+mn_nyq=(/5,31/)
+
 nfp = 2 !still needed in older gvec versions
 
 which_hmap = 21
