+++ test-CI/examples/hmap_knot/parameter.ini
@@ -31,6 +31,9 @@ LA_deg    = 5
 LA_mn_max = (/ 2, 0/)
 LA_sin_cos = _sin_
 
+fac_nyq=-1
+mn_nyq=(/9,13/)
+
 nfp = 3  ! = l of knot
 
 which_hmap = 10
