### 2025-11-19 #8909 @fhindenlang
> [@fhindenlang](https://github.com/fhindenlang) Thanks for your patience. I will be editing this submission. I would like to confirm that you still want reviewers to review version tag v1.2.0. The submission indicates v1.2.0, but you have released v1.3.0 recently. Let me know if you want to switch to v1.3.0 or any specific branch. (Otherwise we will ask reviewers to check out and review only v1.2.0 snapshot.)

Hi @piyueh,
We would prefer it to have v1.3.0 reviewed, which we just released. `v1.3` mostly contained improvements to stability, postprocessing and interfaces to other codes, but did not require any modifictions to the submitted manuscript.
We also rebased `joss_paper_2025` on top of this.

> [@fhindenlang](https://github.com/fhindenlang) Apart from the [question](https://github.com/openjournals/joss-reviews/issues/8909#issuecomment-3549149234) about the version to be reviewed, another point that needs clarification from you is the authorship. There are two non-coder authors, Dean Muir and Eric Sonnendrücker (at least they do not appear in the commit history). Could you please describe their contributions to the code development? [This link](https://joss.readthedocs.io/en/latest/submitting.html#authorship) is the JOSS authorship policy. Thanks!

Dean Muir contributed to testing, analysis and writing the manuscript (and is also a current active developer) and Eric Sonnendrücker contributed to the initial software design and requirements, to the theoretical derivations as well as guidance for the project direction.
Best,
Florian Hindenlang in the name of all co-authors
