### 2025-11-21 0aea10c
harden State.bind
