### 2025-11-21 065c515
restructure with_binding decorator
* cleaner structure
* State.bind() now accepts a 'force' argument
