### 2025-11-21 2ec2806
- Updated plotting routines - Moved 1d, 2d and 3d plots into seperate files
