+++ python/gvec/plotting/plots2d.py
@@ -6,68 +6,10 @@ import matplotlib.pyplot as pyplot
 
 from gvec.core.state import State
 
-from .utils import _get_coord_range, _get_scalars_for_plotting
+from .utils import _get_coord_range
 
 
 pyplot.rcParams.update({"text.usetex": True})
-
-
-def _plot_line_quantities_from_dict(plotting_quantities, x_axis_value, subplot_grid, xlabel):
-    """
-    Plot the quantities from `plotting_quantities`.
-
-    Return both the `matplotlib.pyplot.figure`
-    """
-
-    link_xaxis = False
-    hide_inner_axis = False
-
-    if (subplot_grid[0] != 1) & (subplot_grid[1] == 1):
-        link_xaxis = True
-    elif (subplot_grid[0] != 1) & (subplot_grid[1] != 1):
-        link_xaxis = True
-        hide_inner_axis = True
-
-    f, axs = pyplot.subplots(*subplot_grid, sharex=link_xaxis)
-
-    # Ensure this is an array so we can iterate over it
-    if not isinstance(axs, ndarray):
-        axs = array([axs])
-    if len(axs) != 1:
-        # Otherwise we cannot iterate over this easily
-        axs = axs.flatten()
-
-    if axs.size == 1:
-        # If there is only one axis we plot all quantities on the single axis
-        ax_lines = [
-            axs[0].plot(x_axis_value, values, label=quantity)
-            for (quantity, values) in plotting_quantities.items()
-        ]
-        axs[0].legend(plotting_quantities.keys())
-    else:
-        # Plot one value per axis
-        for i, quantity in enumerate(plotting_quantities):
-            ax_lines = axs[i].plot(x_axis_value, plotting_quantities[quantity])
-            # Top right of each subplot will have a text box with the quantity being plotted
-            axs[i].annotate(
-                quantity,  # TODO: convert this to latex
-                xy=(1, 1),
-                xycoords="axes fraction",
-                xytext=(-0.6, -0.6),
-                textcoords="offset fontsize",
-                verticalalignment="top",
-                horizontalalignment="right",
-                bbox=dict(facecolor="white", edgecolor="black"),
-            )
-
-            if (not link_xaxis) | (i == (len(axs) - 1)):
-                # If the axis are linked we only need the label on the last plot
-                axs[i].set_xlabel(xlabel)
-            if (hide_inner_axis) & (i - len(axs) + subplot_grid[1] >= 0):
-                # If there are multiple plots we need to set the x-axis labels only on the bottom row
-                axs[i].set_xlabel(xlabel)
-
-    return f, axs
 
 
 def plot_poloidal_plane(
@@ -284,133 +226,5 @@ def plot_flux_surface(
         #   vmin=colour_scale[0], vmax=colour_scale[1])
         if share_axis:
             ax.label_outer()  # Removes any axis labels on subplots on the interior of the grid
-
-    return f
-
-
-def plot_radial_profile(
-    state: State,
-    nrho: int | ndarray,
-    quantities: str | list = "mod_B",
-    subplot_grid: list = [1, 1],
-    post_process: dict | dict = None,
-    xaxis="rho",
-):
-    """
-    Plot the radial profile of given equilibrium quantities.
-
-    Parameters
-    ----------
-    state: GVEC state file
-    nrho: int, numpy.ndarray
-        The number of or specific list of radial points to plot at.
-    quantities: str, list, optional
-        Default is "mod_B".
-    subplot_grid: list, conditionally optional
-        The grid shape for the subplots. Default is `[1,1]`. Required if `len(rho)>1`.
-    share_axis: bool
-        Default is True.
-    post_process: dict, optional
[…]
