+++ python/gvec/plotting/__init__.py
@@ -1,7 +1,3 @@
+from .plots1d import plot_radial_profile, plot_on_axis
+from .plots2d import plot_poloidal_plane, plot_flux_surface
 from .plots3d import plot_3d_surface, plot_boundary
-from .plots import (
-    plot_poloidal_plane,
-    plot_flux_surface,
-    plot_radial_profile,
-    plot_on_axis,
-)
