### 2025-11-21 002c4de
- Fixed bug in `_subplots` - Removed ability to give axis object to plots. This may be readded once   I work out how to update the legend - Converted example to jupyter notebook with jupytext.
