### 2025-11-28 ce8b0e8
Updated plots: - Switched from plotting with dicts to using the xarray (see   `_plot_line_quantities_from_xarray` in `plots1d.py`) -- TODO: Need to   reimplement the post processing methods for this implementation - Renamed `plot_flux_surface` to `plot_on_flux_surface` - Removed unnecessary numpy imports - Added function to automatically determine subplot grid size (see   `_design_subgrid` in `utils.py`) - Switched to using quadratic extrapolation for `plot_on_axis` (see   `_extrapolate_axis` in `utils.py`
