+++ python/gvec/plotting/utils.py
@@ -1,24 +1,51 @@
 import matplotlib.pyplot as plt
-from numpy import array, asarray, linspace, max, ndarray, pi, prod
+from numpy import array, asarray, floor, linspace, max, ndarray, pi, prod, sqrt
 
 
-def _get_coord_range(coordinate, nfp, points, min_val=0.0):
+def _extrapolate_axis(state, quantities, zeta):
+    """
+    Generate an evaluations object and replace the axis values with those recovered from quadratic extrapolation
+    """
+    ev_axis = state.evaluate(
+        *quantities,
+        rho=[1e-4, 1.1e-4, 2.2e-4, 3.3e-4, 4.4e-4],  # must be >=1e-4
+        theta=[0.0],
+        zeta=zeta,
+    )
+    ev_axis = ev_axis.sel(theta=ev_axis.theta[0])
+    # quadratic extrapolation to rho=0: y(0)= 3*(y(1) - y(2)) + y(3)
+    ev_axis_1d_extrapol = ev_axis.sel(rho=ev_axis.rho[0])
+    for var in ev_axis.data_vars:
+        r1 = ev_axis.sel(rho=ev_axis.rho[1])[var].data
+        r2 = ev_axis.sel(rho=ev_axis.rho[2])[var].data
+        r3 = ev_axis.sel(rho=ev_axis.rho[3])[var].data
+        # r4=ev_axis.sel(rho=ev_axis.rho[4])[var].data
+        # print(f" check extrapolation, variable:{var} max diff:{np.amax(np.abs(r1 -  (3*(r2 - r3) + r4)))}")
+        # OVERWRITE DATA WITH EXTRAPOLATION works with a 1d array...
+        ev_axis_1d_extrapol[var].data = 3 * (r1 - r2) + r3
+
+    return ev_axis_1d_extrapol
+
+
+def _get_coord_range(coordinate, nfp, points, bounds=None):
     """
     Boiler plate code for building the arrays required for plotting
 
     Return is always of type `numpy.ndarray`.
     """
 
-    # Limits chosen based on the coordinate
-    if coordinate == "rho":
-        auto_limit = [min_val, 1]
-    elif coordinate == "theta":
-        auto_limit = [min_val, 2 * pi]
-    elif coordinate == "zeta":
-        auto_limit = [min_val, 2 * pi / nfp]
-
     # The output will be adjusted depending on the input type
     if isinstance(points, int):
+        # Limits chosen based on the coordinate
+        if bounds is None:
+            if coordinate == "rho":
+                auto_limit = [1e-4, 1.0]
+            elif coordinate == "theta":
+                auto_limit = [0.0, 2 * pi]
+            elif coordinate == "zeta":
+                auto_limit = [0.0, 2 * pi / nfp]
+        else:
+            auto_limit = bounds
         nodes = linspace(*auto_limit, points)
     elif isinstance(points, ndarray):
         nodes = points
@@ -35,6 +62,38 @@ def _get_scalars_for_plotting(evaluations, equilibrium_quantities, post_process,
     Return is a `dict{"quantity": array([values])}
 
     TODO: Allow post processing even if the quantity is a scalar.
+    TODO: Switch from using dicts to xarray object
+    """
+
+    plotting_quantities = {}
+    for quantity in equilibrium_quantities:
+        # Loop over the quantities and store what we're plotting in the dict
+        if max(evaluations[quantity].shape) != prod(evaluations[quantity].shape):
+            # If this quantity is not a scalar
+            #   we check to see if it is going to be remapped, or we error out
+            if post_process is None:
+                raise TypeError("The plotted quantities must be scalars.")
+            else:
+                # Post process the quantity and give the new quantity the given name in plotting
[…]

[…]
