### 2025-11-28 279e668
draft: find_pest_angles
