### 2025-12-01 f803111
finish & test: find_pest_angles
