+++ test-CI/unit-pygvec/test_state.py
@@ -5,6 +5,7 @@ import pytest
 
 try:
     import numpy as np
+    import xarray as xr
 
     import gvec
     from gvec.core.state import State, load_state, find_state, find_states
@@ -309,10 +310,33 @@ def test_get_boozer_angles(teststate):
     assert tz.shape == (2, 3, 4)
 
 
-def test_get_pest_angles(teststate):
+def test_get_pest_angles_1D(teststate):
     rho = np.asarray([0.1, 0.4, 0.6, 1.0])
     theta_P = np.asarray([0.1, 0.5, 0.9])
     zeta = np.asarray([0.1, 0.5, 0.9])
-    tz = teststate.get_pest_angles(rho, np.stack([theta_P, zeta]))
-    assert tz.shape == (2, rho.size, theta_P.size)
-    np.testing.assert_allclose(tz[0, :, :], zeta)
+    theta = teststate.get_pest_angles(rho, np.stack([theta_P, zeta]))
+    assert theta.shape == (theta_P.size, rho.size)
+
+    Z, R = np.meshgrid(zeta, rho, indexing="ij")
+    rtz = np.stack([R.flatten(), theta.flatten(), Z.flatten()])
+    LA = teststate.evaluate_base_list_rtz_all("LA", rtz)[0].reshape(theta_P.size, rho.size)
+    np.testing.assert_allclose(
+        theta + LA, np.tile(theta_P.reshape(-1, 1), (1, rho.size)), atol=1e-12
+    )
+
+
+def test_get_pest_angles_2D_rt(teststate):
+    rho = np.asarray([0.1, 0.4, 0.6, 1.0])
+    theta_P = np.asarray([0.1, 0.5, 0.9])
+    T, R = np.meshgrid(theta_P, rho, indexing="ij")
+    zeta = np.asarray([0.1, 0.5, 0.9])
+    Z, _ = np.meshgrid(zeta, rho, indexing="ij")
+    tz = np.stack([T, Z])
+    theta = teststate.get_pest_angles(rho, tz)
+    assert theta.shape == (theta_P.shape[0], rho.size)
+
+    rtz = np.stack([R.flatten(), theta.flatten(), Z.flatten()])
+    LA = teststate.evaluate_base_list_rtz_all("LA", rtz)[0].reshape(theta_P.size, rho.size)
+    np.testing.assert_allclose(
+        theta + LA, np.tile(theta_P.reshape(-1, 1), (1, rho.size)), atol=1e-12
+    )
