+++ python/gvec/core/compute.py
@@ -633,6 +633,145 @@ def EvaluationsBoozer(
     return ds
 
 
+def EvaluationsPEST(
+    rho: Literal["int"] | CoordinateSpec,
+    theta_P: CoordinateSpec,
+    zeta: Literal["int"] | CoordinateSpec,
+    state: State,
+):
+    """Create an Evaluations dataset with a grid in PEST coordinates.
+
+    PEST coordinates are straight-fieldline coordinates with both the radial and toroidal coordinate being identical to their logical coordinates,
+    i.e. rho_P = rho and zeta_P = zeta. Note that for GVEC the toroidal coordinate is not necessarily the cylindrical angle.
+
+    This factory function generates a mesh in logical coordinates (rho, theta, zeta) based on a grid in PEST coordinates.
+    The grid has dimensions ("rad", "pol", "tor"), corresponding to the radial, poloidal, and toroidal directions.
+
+    If a 2D or 3D array for theta_P or zeta is passed, the corresponding coordinate for the poloidal/toroidal dimension
+    needs to be set manually afterwards (e.g. `ev["alpha"] = ("pol", values)` and `ev = ev.set_coords("alpha").set_xindex("alpha")`).
+
+    Parameters
+    ----------
+    rho : "int" | int | float | 1D array (DataArray, ndarray, list)
+        The specification of the radial, radius-like coordinate. "int" will use the integration points from the state object.
+    theta_P : int | float | 1D, 2D or 3D array (DataArray, ndarray, list)
+        The specification of the poloidal, angle-like PEST coordinate.
+        1D assumes dimension "pol", 2D assumes ("pol", "tor"), 3D assumes ("rad", "pol", "tor").
+    zeta : int | float | 1D, 2D or 3D array (DataArray, ndarray, list)
+        The specification of the toroidal, angle-like logical coordinate.
+        1D assumes dimension "tor", 2D assumes ("pol", "tor"), 3D assumes ("rad", "pol", "tor").
+    state : State
+        The gvec.State object to create the grid for. Used to perform the PEST transform.
+    """
+    match rho:
+        case str() if rho == "int":
+            intp = [state.get_integration_points(q) for q in ["X1", "X2", "LA"]]
+            if any([not np.allclose(intp[0][j], intp[i][j]) for i in (1, 2) for j in (0, 1)]):
+                raise ValueError("Integration points for rho do not align for X1, X2 and LA.")
+            rho = ("rad", intp[0][0])
+        case xr.DataArray():
+            rho = rho
+        case np.ndarray() | Sequence():
+            rho = np.asarray(rho)
+            if rho.ndim != 1:
+                raise ValueError(f"rho can only be 1D, but is {rho.ndim}D.")
+            rho = ("rad", rho)
+        case int():
+            rho = ("rad", np.linspace(0, 1, rho + 1)[1:])
+        case float():
+            rho = ("rad", np.array([rho]))
+        case _:
+            raise ValueError(f"Could not parse rho, got {rho}.")
+    match theta_P:
+        case xr.DataArray():
+            theta_P = theta_P
+        case np.ndarray() | Sequence():
+            theta_P = np.asarray(theta_P)
+            if theta_P.ndim == 1:
+                theta_P = ("pol", theta_P)
+            elif theta_P.ndim == 2:
+                theta_P = (("pol", "tor"), theta_P)
+            elif theta_P.ndim == 3:
+                theta_P = (("rad", "pol", "tor"), theta_P)
+            else:
[…]
