### 2025-12-02 1c77998
add evaluate_sfl(..., sfl="pest")
