### 2025-12-02 1177ef7
extend quantities for  PEST coordinates
* tangent vectors in PEST coordinates added
* factory functions for B and J extended to take any vector quantity and add an argument which sfl coordinate is used (Boozer or PEST)
* add quantities to unit test. Check average of fields and B_contra_t_P with equivalent quantity in GVEC coordinates.
* removed requirement "xyz" where it is not necessary
