+++ python/gvec/quantities.py
@@ -529,7 +529,14 @@ def theta_P(ds: xr.Dataset):
 
 
 @register(
-    requirements=("xyz", "theta_sfl", "dLA_dr", "dLA_dt", "dLA_dz"),
+    requirements=(
+        "dLA_dr",
+        "dLA_dt",
+        "dLA_dz",
+        "grad_theta",
+        "grad_rho",
+        "grad_zeta",
+    ),
     attrs=dict(
         long_name="poloidal reciprocal basis vector in PEST coordinates",
         symbol=r"\nabla \theta_P",
@@ -549,11 +556,57 @@ def Jac_P(ds: xr.Dataset):
     ds["Jac_P"] = ds.Jac / (1 + ds.dLA_dt)
 
 
+@register(
+    requirements=(
+        "dLA_dr",
+        "dLA_dt",
+        "e_rho",
+        "e_theta",
+    ),
+    attrs=dict(
+        long_name="poloidal tangent basis vector in PEST coordinates",
+        symbol=r"\mathbf{e}_{\theta_P}",
+    ),
+)
+def e_rho_P(ds: xr.Dataset):
+    ds["e_rho_P"] = ds.e_rho - ds.e_theta * (ds.dLA_dr / (1 + ds.dLA_dt))
+
+
+@register(
+    requirements=(
+        "dLA_dt",
+        "e_theta",
+    ),
+    attrs=dict(
+        long_name="poloidal tangent basis vector in PEST coordinates",
+        symbol=r"\mathbf{e}_{\theta_P}",
+    ),
+)
+def e_theta_P(ds: xr.Dataset):
+    ds["e_theta_P"] = ds.e_theta / (1 + ds.dLA_dt)
+
+
+@register(
+    requirements=(
+        "dLA_dt",
+        "dLA_dz",
+        "e_theta",
+        "e_zeta",
+    ),
+    attrs=dict(
+        long_name="toroidal tangent basis vector in PEST coordinates",
+        symbol=r"\mathbf{e}_{\zeta_P}",
+    ),
+)
+def e_zeta_P(ds: xr.Dataset):
+    ds["e_zeta_P"] = ds.e_zeta - ds.e_theta * (ds.dLA_dz / (1 + ds.dLA_dt))
+
+
 # === derived ========================================================================== #
 
 
 @register(
-    requirements=("xyz", "e_q1", "e_q2", "dX1_dr", "dX2_dr"),
+    requirements=("e_q1", "e_q2", "dX1_dr", "dX2_dr"),
     attrs=dict(long_name="radial tangent basis vector", symbol=r"\mathbf{e}_\rho"),
 )
 def e_rho(ds: xr.Dataset):
@@ -561,7 +614,7 @@ def e_rho(ds: xr.Dataset):
 
 
 @register(
-    requirements=("xyz", "e_q1", "e_q2", "dX1_dt", "dX2_dt"),
+    requirements=("e_q1", "e_q2", "dX1_dt", "dX2_dt"),
     attrs=dict(long_name="poloidal tangent basis vector", symbol=r"\mathbf{e}_\theta"),
 )
 def e_theta(ds: xr.Dataset):
@@ -569,7 +622,7 @@ def e_theta(ds: xr.Dataset):
 
 
 @register(
-    requirements=("xyz", "e_q1", "e_q2", "e_q3", "dX1_dz", "dX2_dz"),
+    requirements=("e_q1", "e_q2", "e_q3", "dX1_dz", "dX2_dz"),
     attrs=dict(long_name="toroidal tangent basis vector", symbol=r"\mathbf{e}_\zeta"),
 )
 def e_zeta(ds: xr.Dataset):
[…]
