+++ python/gvec/quantities.py
@@ -1219,36 +1219,97 @@ def _k_ij_B(ds: xr.Dataset):
 
 
 @register(
-    requirements=("normal", "k_tt_B"),
-    attrs=dict(
-        long_name="poloidal Boozer component of the second fundamental form",
-        symbol=r"\mathrm{II}_{\theta_B\theta_B}",
-    ),
+    quantities=("k_tt_P", "k_tz_P", "k_zz_P"),
+    requirements=["e_theta", "e_zeta", "k_tt", "k_tz", "k_zz"]
+    + sum([[f"dLA_d{ij}"] for ij in ("t", "z", "tt", "tz", "zz")], start=[]),
+    attrs={
+        f"k_{a}{b}_P": dict(
+            long_name=f"{a}{b} PEST curvature vector",
+            symbol=rf"\mathbf{{k}}_{{{rtz_symbols[a]}_P {rtz_symbols[b]}_P}}",
+        )
+        for a, b in ["tt", "tz", "zz"]
+    },
 )
-def II_tt_B(ds: xr.Dataset):
-    ds["II_tt_B"] = xr.dot(ds.normal, ds.k_tt_B, dim="xyz")
+def _k_ij_P(ds: xr.Dataset):
+    r"""Factory function for curvature vectors in PEST coordinates.
+
+    The curvature vector is computed in cartesian space, with
+
+    k_{\alpha\beta} = \frac{\partial}{\partial \alpha} \left(\frac{\partial x}{\partial \beta}\right)
+
+    for the choices of $\alpha,\beta$  being $\vartheta_P,\vartheta_P$, $\vartheta_P,\zeta_P$ and $\zeta_P,\zeta_P$.
+    The chain rule is applied to express the quantities in terms of the logical coordinate derivatives,
+    together with the derivatives of the PEST coordinate transform.
+    """
+    dtP_dt = 1 + ds.dLA_dt
+    dtP_dz = ds.dLA_dz
+    # dzP_dt = 0
+    # dzP_dz = 1
+
+    dtP_dtt = ds.dLA_dtt
+    dtP_dtz = ds.dLA_dtz
+    dtP_dzz = ds.dLA_dzz
+    # dzP_dtt = 0
+    # dzP_dtz = 0
+    # dzP_dzz = 0
+
+    JacP_Jac = 1 / (dtP_dt)
+    dJac_JacP_dt = dtP_dtt
+    dJac_JacP_dz = dtP_dtz
+    dJacP_Jac_dtP = JacP_Jac**3 * (-dJac_JacP_dt)
+    dJacP_Jac_dzP = JacP_Jac**3 * (dtP_dz * dJac_JacP_dt - dtP_dt * dJac_JacP_dz)
+
+    dt_dtP = JacP_Jac
+    # dz_dtP = 0
+    dt_dzP = -JacP_Jac * dtP_dz
+    dz_dzP = JacP_Jac * dtP_dt
+
+    dt_dttP = dJacP_Jac_dtP
+    dt_dtzP = dJacP_Jac_dzP
+    dt_dzzP = -dtP_dz * dJacP_Jac_dzP - JacP_Jac * (dt_dzP * dtP_dtz + dz_dzP * dtP_dzz)
+    # dz_dttP = 0
+    dz_dtzP = dtP_dt * dJacP_Jac_dtP + JacP_Jac * (dt_dtP * dtP_dtt)
+    dz_dzzP = dtP_dt * dJacP_Jac_dzP + JacP_Jac * (dt_dzP * dtP_dtt + dz_dzP * dtP_dtz)
+
+    ds["k_tt_P"] = dt_dttP * ds.e_theta
+    ds["k_tt_P"] += dt_dtP**2 * ds.k_tt
+
[…]
