+++ docs/conf.py
@@ -140,10 +140,18 @@ mathjax3_config = {
             "Xp": r"\X^{\prime}",
             "Xpp": r"\X^{\prime\prime}",
             "Xppp": r"\X^{\prime\prime\prime}",
+            "Tvec": r"\vec{T}",
+            "Nvec": r"\vec{N}",
+            "Uvec": r"\vec{U}",
+            "rvec": r"\vec{r}",
+            "Uvectilde": r"\vec{\tilde{U}}",
             "ttilde": r"{\vec{\tilde{T}}_q}",
             "modB": r"\abs{\vec{B}}",
             "dPhidr": r"\dv{\Phi}{\rho}",
             "aqty": [r"\langle{#1}\rangle", 1],
+            "Wr": r"\textrm{Wr}",
+            "Tw": r"\textrm{Tw}",
+            "Lk": r"\textrm{Lk}",
         },
     },
 }
