### 2025-12-03 edae7b4
extract Biot-Savart routines from free boundary branch
