+++ python/gvec/coils.py
@@ -1,0 +1,618 @@
+import numpy as np
+import xarray as xr
+from numpy.typing import ArrayLike
+from gvec.lib import modgvec_biotsavart as _BS
+from warnings import warn
+from collections.abc import Iterable
+from pathlib import Path
+from scipy.integrate import solve_ivp
+from tqdm import tqdm
+import os
+from gvec.util import logging_setup
+from gvec.core.state import State, CoordinateSpec
+from logging import getLogger
+
+mu_0 = 4 * np.pi * 0.99999999987 * 1e-7
+
+
+def _stack_for_BiotSavart(ds: xr.Dataset | xr.DataArray):
+    if "points" in ds.dims:
+        return ds
+    else:
+        stack_dims = [dim for dim in ds.dims if dim != "xyz"]
+        ds = ds.stack(points=stack_dims)
+        return ds
+
+
+class Coil:
+    def __init__(self, coil_points: ArrayLike, coil_current: float):
+        """Filament coil discretized via straight line segments.
+
+        Parameters
+        ----------
+        coil_points : ArrayLike
+            Discretization points of the coil in Cartesian coordinates. Expected shape is (3,n_points).
+        coil_current : float
+            Current flowing in the coil in [A].
+        """
+        if not isinstance(coil_points, xr.DataArray):
+            coil_points = np.atleast_2d(coil_points)
+            self.coil_points = xr.DataArray(
+                coil_points, dims=("xyz", "points"), coords={"xyz": ("xyz", ["x", "y", "z"])}
+            )
+        else:
+            self.coil_points = coil_points
+        self.coil_current = coil_current
+        self.n_points = self.coil_points.shape[1]
+        self.e_vec = self.coil_points[:, 1:] - self.coil_points[:, :-1]
+        self.L = self._mod_vector(self.e_vec)
+        self.e_hat_vec = self.e_vec / self.L
+        self.prefactor = self.coil_current * mu_0 / (4 * np.pi)
+
+    def __call__(self, pos: ArrayLike):
+        return self.eval_B(pos)
+
+    def __repr__(self):
+        return f"Coil(points={self.n_points}, current={self.coil_current:.2e} A)"
+
+    def _mod_vector(self, xyz):
+        return np.sqrt(xr.dot(xyz, xyz, dim="xyz"))
+
+    def eval_B(self, pos: ArrayLike):
+        """Evaluate the magnetic field due to currents flowing in the coil at given positions.
+
+        Parameters
+        ----------
+        pos : ArrayLike
+            Cartesian coordinates of the evaluation points with shape (3,n_positions).
+
+        Returns
+        -------
+        xr.Dataset
+            Dataset containing the magnetic field at the evaluation positions.
+        """
+        if not isinstance(pos, xr.DataArray):
+            pos = np.asanyarray(pos)
+            if pos.ndim == 1:
+                pos = pos.reshape(-1, 1)
+            pos = xr.DataArray(
+                pos,
+                dims=("xyz", "points"),
+                coords={"xyz": ("xyz", ["x", "y", "z"])},
+                attrs=dict(long_name="position vector", symbol=r"\mathbf{x}"),
+            )
+        n_positions = pos.shape[1]
+
+        B = np.asfortranarray(np.zeros(pos.shape))
+
+        _BS.biotsavart(
+            n_positions=n_positions,
+            xyz=pos,
+            n_points=self.n_points,
+            coil_points=self.coil_points,
+            prefactor=self.prefactor,
+            b=B,
+        )
+
+        ds = xr.Dataset(
+            data_vars=dict(B=(("xyz", "points"), B)),
[…]
