### 2025-12-04 5b09ae9
Fix missing comma in FORMAT string
Add missing comma between string literal and I4 format descriptor
to comply with ISO Fortran standard and eliminate GCC 15 warning.
