### 2025-12-04 ea78ca5
Use .F90 extension for preprocessed files, remove global -cpp flag
Rename Fortran source files from .f90 to .F90 to enable automatic
preprocessing by the compiler based on file extension. This eliminates
the need for explicit preprocessor flags (-xf95-cpp-input, -fpp,
-Mpreprocess, -F) and avoids the GCC 15 preprocessor warnings about
apostrophes in Fortran comments.

Changes:
- Rename files containing #if/#ifdef/#include directives to .F90
- Update CMakeLists.txt glob patterns to match .F90
- Remove explicit preprocessor flags from all compiler configurations

The .F90 extension is the standa
[…]
