+++ CMakeLists.txt
@@ -157,12 +157,14 @@ ENDIF()
 # COMPILER FLAGS (compiler & machine specific)
 # =========================================================================
 
+SET(GVEC_COMPILE_FLAGS "")
+
 IF(CMAKE_Fortran_COMPILER_ID MATCHES "GNU")
   SET(CMAKE_Fortran_FLAGS "${CMAKE_Fortran_FLAGS} -fdefault-real-8 -fdefault-double-8 -fbackslash -ffree-line-length-0 -DGNU")
   SET(CMAKE_Fortran_FLAGS_RELEASE "${CMAKE_Fortran_FLAGS}     -O3 -march=native -finline-functions") #gives segfault: -fstack-arrays
   SET(CMAKE_Fortran_FLAGS_PROFILE "${CMAKE_Fortran_FLAGS} -pg -O3 -march=native -finline-functions") #gives segfault: -fstack-arrays
   SET(CMAKE_Fortran_FLAGS_DEBUG   "${CMAKE_Fortran_FLAGS} -g -O0 -ggdb3 -ffpe-trap=invalid -fbounds-check -finit-real=snan -fbacktrace -Wall -Wextra")
-  SET(GVEC_COMPILE_FLAGS "-xf95-cpp-input -fPIC")
+  SET(GVEC_COMPILE_FLAGS "-fPIC")
 
 ELSEIF (CMAKE_Fortran_COMPILER_ID MATCHES "Intel")
   # machine specific flags
@@ -183,28 +185,25 @@ ELSEIF (CMAKE_Fortran_COMPILER_ID MATCHES "Intel")
   IF(CMAKE_Fortran_COMPILER_VERSION VERSION_LESS "2021.2")
     SET(CMAKE_Fortran_FLAGS_DEBUG "${CMAKE_Fortran_FLAGS_DEBUG} -check all,uninit") # crash with ifx; possibly related to netCDF
   ENDIF()
-  SET(GVEC_COMPILE_FLAGS "-fpp -assume bscc -fPIC")
+  SET(GVEC_COMPILE_FLAGS "-assume bscc -fPIC")
 
 ELSEIF(CMAKE_Fortran_COMPILER_ID MATCHES "Cray")
   SET(CMAKE_Fortran_FLAGS         "${CMAKE_Fortran_FLAGS} -f free -s real64 -em -DCRAY")
   SET(CMAKE_Fortran_FLAGS_RELEASE "${CMAKE_Fortran_FLAGS} -O2 -hfp3 -p . -rm")
   SET(CMAKE_Fortran_FLAGS_PROFILE "${CMAKE_Fortran_FLAGS} -O2 -hfp3 -h profile_generate -p . -rm")
   SET(CMAKE_Fortran_FLAGS_DEBUG   "${CMAKE_Fortran_FLAGS} -O0 -eD -rm")
-  SET(GVEC_COMPILE_FLAGS "-F")
 
 ELSEIF(CMAKE_Fortran_COMPILER_ID MATCHES "PGI")
   SET(CMAKE_Fortran_FLAGS         "${CMAKE_Fortran_FLAGS} -r8 -fPIC")
   SET(CMAKE_Fortran_FLAGS_RELEASE "${CMAKE_Fortran_FLAGS} -O3")
   SET(CMAKE_Fortran_FLAGS_PROFILE "${CMAKE_Fortran_FLAGS} -O3")
   SET(CMAKE_Fortran_FLAGS_DEBUG   "${CMAKE_Fortran_FLAGS} -O0 -g")
-  SET(GVEC_COMPILE_FLAGS "-Mpreprocess")
 
[…]
