### 2025-12-04 834758c
Added NCSX example for coils and fieldline tracing - check if first and last point of coil are the same - added dependencies for examples
