+++ pyproject.toml
@@ -60,7 +60,7 @@ Issues = "https://gitlab.mpcdf.mpg.de/gvec-group/gvec/-/issues"
 
 [project.optional-dependencies]
 dev = ["pytest"]
-examples = ["matplotlib", "jupyter", "ipywidgets", "ipympl", "simsopt"]
+examples = ["matplotlib", "jupyter", "ipywidgets", "ipympl", "simsopt", "plotly", "nbformat", "joblib"]
 
 [project.scripts]
 pygvec = "gvec.scripts.main:main"
