### 2025-12-05 7d1e82c
adapted coils example: - added further logging messages - fixed progressbar for parallel fieldline tracing - tuned coils example
