### 2025-12-05 c061736
add description of the double integral algorithm
