### 2025-12-05 ad700c9
Updated plots:\n - Updated example\n - Fixed issues with post_processing
