### 2025-12-08 6914522
adapted coils example:  - added further utility functions  - added function for field-line tracing for quasr cases  - added quasr example for fieldline tracing and comparing to equilibrium
