### 2025-12-08 1203063
Added vector potential methods to  and _Methods need testing_
