+++ python/gvec/coils.py
@@ -1,17 +1,20 @@
+import os
+from collections.abc import Iterable
+from logging import getLogger
+from pathlib import Path
+from typing import Literal
+from warnings import warn
+
 import numpy as np
 import xarray as xr
 from numpy.typing import ArrayLike
-from gvec.lib import modgvec_biotsavart as _BS
-from warnings import warn
-from collections.abc import Iterable
-from pathlib import Path
+from scipy.constants import mu_0
 from scipy.integrate import solve_ivp
 from tqdm import tqdm
-import os
+
+from gvec.core.state import CoordinateSpec, State
+from gvec.lib import modgvec_biotsavart as _BS
 from gvec.util import logging_setup
-from gvec.core.state import State, CoordinateSpec
-from logging import getLogger
-from typing import Literal
 
 try:
     from joblib import Parallel, delayed, parallel_config
@@ -37,8 +40,6 @@ try:
 
 except ImportError:
     has_joblib = False
-
-mu_0 = 4 * np.pi * 0.99999999987 * 1e-7
 
 
 def _stack_for_BiotSavart(ds: xr.Dataset | xr.DataArray):
@@ -134,6 +135,53 @@ class Coil:
         ds.B.attrs = dict(long_name="magnetic coil field", symbol=r"\mathbf{B}_C")
         return ds
 
+    def eval_A(self, pos: ArrayLike):
+        """Evaluate the vector potential due to currents flowing in the coil at given positions.
+
+        Parameters
+        ----------
+        pos : ArrayLike
+            Cartesian coordinates of the evaluation points with shape (3,n_positions).
+
+        Returns
+        -------
+        xr.Dataset
+            Dataset containing the vector potential at the evaluation positions.
+        """
+        if not isinstance(pos, xr.DataArray):
+            pos = np.asanyarray(pos)
+            if pos.ndim == 1:
+                pos = pos.reshape(-1, 1)
+            pos = xr.DataArray(
+                pos,
+                dims=("xyz", "points"),
+                coords={"xyz": ("xyz", ["x", "y", "z"])},
+                attrs=dict(long_name="position vector", symbol=r"\mathbf{x}"),
+            )
+        n_positions = pos.shape[1]
+
+        A = np.asfortranarray(np.zeros(pos.shape))
+
+        _BS.biotsavart_vectorpotential(
+            n_positions=n_positions,
+            xyz=pos,
+            n_points=self.n_points,
+            coil_points=self.coil_points,
+            ehat=self.e_hat_vec,
+            l=self.L,
+            prefactor=self.prefactor,
+            a=A,
+        )
+
+        ds = xr.Dataset(
+            data_vars=dict(A=(("xyz", "points"), A)),
+            coords={"xyz": ("xyz", ["x", "y", "z"])},
+        )
+        ds["pos"] = pos
+
+        ds.A.attrs = dict(long_name="magnetic coil vector potential", symbol=r"\mathbf{A}_C")
+        return ds
+
     def eval_mod_B(self, pos: ArrayLike):
         """Evaluate the modulus of the magnetic field due to currents flowing in the coil at given positions.
 
@@ -149,6 +197,23 @@ class Coil:
         """
         ds = self.eval_B(pos)
         ds["mod_B"] = self._mod_vector(ds.B)
+        return ds
+
+    def eval_mod_A(self, pos: ArrayLike):
+        """Evaluate the modulus of the vector potential due to currents flowing in the coil at given positions.
+
+        Parameters
+        ----------
+        pos : ArrayLike
+            Cartesian coordinates of the evaluation points with shape (3,n_positions).
+
+        Returns
+        -------
+        ds : xr.Dataset
[…]
