### 2025-12-08 d946d3a
added test for generating quasr fieldlines/case
