### 2025-12-08 9de8e12
extended char length of boundary and gframe file
