### 2025-12-09 651325e
python implementation of linking number and writhe.
* add one function to compute the solid angle between to linear segments,
* add linking number of two curves. As long as discretized polygon curves do not intersect, linking number is independent of resolution(!)
* added test for linking number of to linked/unlinked circles with different resolutions
* using numpy arrays to span a matrix of segment pairs and then sum the solid angle, seems to be reasonably fast (<1sec if curves have ~500 points)
* writhe of a single curve implemented, with segment pairs j>i+1 only. Careful here: accuracy depe
[…]
