+++ python/gvec/util.py
@@ -1205,3 +1205,113 @@ def boundary_generator(case: str, X1_00=1.0, a0=0.5, ellipticity=0.4, helix_r=0.
             raise ValueError(f"request boundary '{case}', does not exist!")
 
     return params
+
+
+def linking_number(curve_a: np.ndarray, curve_b: np.ndarray, tol=1e-15):
+    """
+    Compute the linking number of two non-intersecting curves $C_a(\zeta_a),C_b(\zeta_b)$, solving the (non-singular) double integral over two curves
+
+    $$\textrm{Lk} = \Lk = \frac{1}{4\pi}\int_0^{2\pi} \int_0^{2\pi} \frac{ r_b - r_a}{\abs{r_b - r_a}^3} \cdot \left(\frac{\partial r_a}{\zeta_a}  \times \frac{\partial r_b}{\zeta_b}} d{\zeta_a} d{\zeta_b}$$
+
+    Here, the double integral is approximated by representing the curves as polygons, and computing the linking number as the sum of solid angles between all linear segments of polygon curves. The solid angle of two linear segments can be computed exactly. See paper by  [Klenin and Langowski](https://onlinelibrary.wiley.com/doi/10.1002/1097-0282(20001015)54:5%3C307::AID-BIP20%3E3.0.CO;2-Y)
+    See GVEC documentation for more details.
+    Args:
+        curve_a (2D np.ndarray): the x,y,z point positions of the first curve. First and last point must coincide. shape must be [npoints_a,3]
+        curve_b (2D np.ndarray): the x,y,z point positions of the first curve. First and last point must coincide. shape must be [npoints_b,3]
+        tol (float): tolerance to consider points coincident (default: 1e-15)
+    Returns:
+        Lk (float): the linking number of the two curves
+    """
+    assert curve_a.ndim == 2, "curve_a must be a 2D array"
+    assert curve_b.ndim == 2, "curve_b must be a 2D array"
+    assert curve_a.shape[1] == 3, "second dimension of curve_a must be of size 3"
+    assert curve_b.shape[1] == 3, "second dimension of curve_b must be of size 3"
+    assert np.allclose(curve_a[0, :], curve_a[-1, :]), (
+        "first and last point of curve_a must coincide (closed curve)"
+    )
+    assert np.allclose(curve_b[0, :], curve_b[-1, :]), (
+        "first and last point of curve_b must coincide (closed curve)"
+    )
+    # nseg_a = curve_a.shape[0]-1
+    # nseg_b = curve_b.shape[0]-1
+    # Lk = 0.0
+    # for i in range(nseg_a):
+    #    for j in range(nseg_b):
+    #        r1 = curve_a[i,:]
+    #        r2 = curve_a[i+1,:]
+    #        r3 = curve_b[j,:]
+    #        r4 = curve_b[j+1,:]
+    #        Lk += solid_angle_between_segments(r1,r2,r3,r4)
+    # Lk /= (4.0*np.pi)
+    return np.sum(
+        solid_angle_between_segments(
+            curve_a[0:-1, None, :],
+            curve_a[1:, None, :],
+            curve_b[None, 0:-1, :],
+            curve_b[None, 1:, :],
+            tol=tol,
+        )
+    ) / (4 * np.pi)
+
+
+def writhe(curve: np.ndarray):
+    """
+    Compute the writhe of a curve $C(\zeta)$, solving the (singular!) double integral over the curve
+
[…]
