### 2025-12-09 112c3ef
adapted plotting import and quasr fieldline test: - included coils tutorials in getting started, still need adaptation to run in documentation - kept matplotlib still as an optional dependency
