+++ docs/tutorials/getting-started.md
@@ -44,6 +44,8 @@ The available tutorials are:
    * [](/tutorials/notebooks/030_stages)
    * [](/tutorials/notebooks/040_current)
    * [](/tutorials/notebooks/050_pygvec)
+   * [](/tutorials/notebooks/060_coils)
+   * [](/tutorials/notebooks/061_coils)
 
 ### Running Jupyter notebooks
 
