### 2025-12-09 96ac443
updated coils tutorial: - included plotly requirements in conf.py - increases allowed cell-runtime in tutorial notebooks - included plotly<6.0 as requirement for docs
