+++ docs/requirements.txt
@@ -11,3 +11,4 @@ ford
 graphviz
 
 matplotlib
+plotly<6.0 # plotly 6.0 has issues with myst-nb and won't show plots in notebook
