+++ CMakeLists.txt
@@ -418,6 +418,9 @@ IF(COMPILE_PYGVEC)
     src/pygvec/state.f90
     src/pygvec/run.f90
     src/transform_sfl/sfl_boozer.f90
+    src/profiles/c_rprofile.f90
+    src/profiles/rprofile_bspline.f90
+    src/profiles/rprofile_polynomial.f90
   )
 
   FOREACH(file ${pygvec_src_files})
@@ -434,12 +437,11 @@ IF(COMPILE_PYGVEC)
     COMMAND "${PYTHON_EXECUTABLE}"
             -m f90wrap
             ${pygvec_src_files}
+            --verbose
             --mod-name lib  # -> lib.py - autogenerated python wrappers
             --f90-mod-name gvec._libpygvec  # -> _libpygvec.cpython*.so - compiled python module
             --kind-map "${CMAKE_CURRENT_SOURCE_DIR}/python/kind_map.py"
             --class-names "${CMAKE_CURRENT_SOURCE_DIR}/python/class_names.py"
-            --constructors
-            --destructors sfl_boozer_free
     DEPENDS ${pygvec_src_files}
     OUTPUT ${pygvec_wrap_files}
     COMMENT "pyGVEC: generating the Fortran layer with f90wrap"
