+++ pyproject.toml
@@ -111,6 +111,7 @@ version_file = "python/gvec/_version.py"
 
 # ============================================================================ #
 [tool.pytest.ini_options]
+testpaths = ["test-CI"]
 addopts = [
     "--import-mode=importlib",
 ]
