### 2025-12-11 da801b7
Fixed coil vector potential calculation: - added unit tests for coil A - small adaptations in tutorial - removed setting OMP threads in fieldline tracing
