### 2025-12-11 1a38677
Updated plotting in response to MR comments
- Added `usetext` to the `__init__.py` and removed it from other files. This line needs to be run once on import.
- Swapped to `import numpy as np` instead of importing individual functions
- Added method for checking if `\\text` is in the symbol name
- Switched `subplot_grid` to default `None` as input and auto-generate grid.
- Fixed some small issues with labels
- Reordered inputs so that `quantities` is first and `rho`, `theta`, `zeta` are the same order as `state.evaluate`
- Fixed small issues with docstrings
- Removed `_get_coord_range` in favou
[…]
