+++ python/gvec/plotting/plots3d.py
@@ -1,39 +1,37 @@
-from plotly import graph_objects
-import plotly.offline as plotly_offline
+# Copyright (c) 2025 GVEC Contributors, Max Planck Institute for Plasma Physics
+# License: MIT
+
+import numpy as np
 
 from gvec.core.state import State
-
-from numpy import meshgrid
-
-from .utils import _get_coord_range
 
 
 def plot_3d_surface(
     state: State,
-    nzeta: int,
-    ntheta: int,
     rho: float,
-    equilibrium_quantity: str = "mod_B",
+    ntheta: int,
+    nzeta: int,
+    quantity: str = "mod_B",
     full_surface=False,
-    offline: bool = False,
+    to_file: str | None = None,
 ):
     """
     Generate a 3D surface plot with the quantity provided by `equilibrium_quantity` on it at a given `rho` position.
 
     Parameters
     ----------
-    state: GVEC state file
-    nzeta: int
-        Toroidal resolution
-    ntheta: int
-        Poloidal resolution
-    rho: float
+    state : GVEC state file
+    rho : float
         Radial position of the surface
-    equilibrium_quantity: String, optional
+    ntheta : int
+        Poloidal resolution
+    nzeta : int
+        Toroidal resolution
+    equilibrium_quantity : String, optional
         default is "mod_B"
-    offline: bool
-        If true, will automatically save the plot to a file in the current working directory.
-        default is False
+    to_file : str
+        If a string, will automatically save the plot to a file with the given input in the current working directory. Recommended to use this if the plots don't display.
+        default is None
 
     Returns
     -------
@@ -42,15 +40,16 @@ def plot_3d_surface(
 
     plot a 3D surface at a given rho. By default plots the exterior boundary
     """
+    import plotly.offline as plotly_offline
+    from plotly import graph_objects
 
     nfp = state.nfp
     if full_surface:
         nfp = 1.0
 
-    theta = _get_coord_range("theta", nfp, ntheta)
-    zeta = _get_coord_range("zeta", nfp, nzeta)
+    zeta = np.linspace(0.0, 2 * np.pi / nfp, nzeta)
 
-    ev = state.evaluate(equilibrium_quantity, rho=[rho], theta=theta, zeta=zeta).sel(rho=rho)
+    ev = state.evaluate(quantity, rho=[rho], theta=ntheta, zeta=zeta).sel(rho=rho)
 
     # if existing_plot is None:
     plt = graph_objects.Figure()
@@ -61,35 +60,29 @@ def plot_3d_surface(
             x=ev.pos.sel(xyz="x"),
             y=ev.pos.sel(xyz="y"),
             z=ev.pos.sel(xyz="z"),
-            surfacecolor=ev[equilibrium_quantity],
-            colorbar_title_text=equilibrium_quantity,
+            surfacecolor=ev[quantity],
+            colorbar_title_text=quantity,
         )
     )
 
-    if offline:  # TODO See why plotly is not showing
-        plotly_offline.plot(plt)
-        print(
-            "Plotly figure saved to working directory. To avoid saving call with `offline=False`."
-        )
+    if isinstance(to_file, str):  # TODO See why plotly is not showing
+        plotly_offline.plot(plt, filename=to_file)
 
     return plt
 
 
-def plot_boundary(state, nzeta, ntheta, equilibrium_quantity="mod_B", full_surface=True):
+def plot_boundary(state, nzeta, ntheta, quantity="mod_B", full_surface=True, to_file=None):
     """
     Plot the boundary of a GVEC solve.
-
-    Parameters
-    ----------
-    state
-    nzeta
-    ntheta
-    equilibrium_property
-
-    TODO: Docstring
 
[…]
