+++ python/gvec/plotting/plots2d.py
@@ -1,109 +1,106 @@
+# Copyright (c) 2025 GVEC Contributors, Max Planck Institute for Plasma Physics
+# License: MIT
+from typing import Literal
 from warnings import warn
 
-import matplotlib.pyplot as pyplot
-from numpy import amax, amin, any, isnan, linspace, max, min, ndarray, pi, prod
+import numpy as np
+from numpy import ndarray  # type checking
 
 from gvec.core.state import State
-
-from gvec.plotting.utils import _design_subgrid, _get_coord_range, _subplots
-
-pyplot.rcParams.update({"text.usetex": True})
+from gvec.plotting.utils import _design_subgrid, _subplots, _symbol_check
 
 
 def plot_poloidal_plane(
     state: State,
+    quantity: str = "mod_B",
+    /,
     nrho: int = 21,
     ntheta: int = 51,
     zeta: int | float | ndarray = 9,
-    quantities: str = "mod_B",
-    subplot_grid: list = [3, 3],
+    subplot_grid: list | None = None,
     share_axis: bool = False,
     st_contours: list = [0, 0],
     theta_contour_style: str = "theta",
 ):
     """
-    plot_poloidal_slice
+    plot_poloidal_plane
 
-    Plot a poloidal slice with some equilibrium quantity on it. Defaults to plotting $\|\mathbf{B}\|$
+    Plot a poloidal plane with some equilibrium quantity on it. Defaults to plotting $\|\mathbf{B}\|$
 
     Parameters
     ----------
-    state: GVEC state file
-    nrho: int, optional
+    state : GVEC state file
+    quantity : str, optional
+        The quantity to plot. Default is "mod_B"
+    nrho : int, optional
         The radial resolution of the slices. Default is 51
-    ntheta: int, optional
+    ntheta : int, optional
         The poloidal resolution of the slices. Default is 51
-    zeta: int, float, ndarray, optional
+    zeta : int, float, ndarray, optional
         Poloidal slices to plot. Default is 9.
         - int: Number of equally spaced slices.
         - float: The specific slice.
         - ndarray: The specific slices.
-    quantities: str
-        The quantity to plot. Default is "mod_B"
-    subplot_grid: list, conditionally optional
-        The grid shape for the subplots. Default is `[3,3]`. Required if `zeta>1`.
-    share_axis: bool
+    subplot_grid : list, None, optional
+        The grid shape for the subplots. If `None`, grid will be automatically determined. Default is `None`.
+    share_axis : bool
         If true, all subplots will share their `X1` and `X2` axis positions.
-    st_contours: list
+    st_contours : list
         Number of ``$\theta$`` and ``$\zeta$`` contours to plot.
-    theta_contour_style: str
+    theta_contour_style : str
         Either `"theta"` or `"theta_star"`.
 
     Returns
     -------
-    `matplotlib.pyplot.figure` object
+    `matplotlib.pyplot.figure` object and `numpy.ndarray` of `matplotlib.axis._axis.Axes` object(s).
     """
 
     if theta_contour_style not in ["theta", "theta_star"]:
         raise ValueError("theta_contour_type must be 'theta' or 'theta_star'")
 
-    rho = _get_coord_range("rho", state.nfp, nrho)
-    theta = _get_coord_range("theta", state.nfp, ntheta)
-    zeta_eval = _get_coord_range("zeta", state.nfp, zeta)
+    evaluations = state.evaluate(quantity, "X1", "X2", "LA", rho=nrho, theta=ntheta, zeta=zeta)
 
-    ev = state.evaluate(*[quantities, "X1", "X2", "LA"], rho=rho, theta=theta, zeta=zeta_eval)
+    zeta_eval = evaluations.zeta.data
 
-    if any(isnan(ev[quantities])):
[…]

[…]
