### 2025-12-11 85ce6f2
Added `usetext` to the `__init__.py` and removed it from other files. This line needs to be run once on import.
Missed in previous commit.
