### 2025-12-11 41489db
Updated behaviour of labels in `_plot_on_flux_surface`
