### 2025-12-11 ccfb31c
Added tests for circular coil B-field off axis
