### 2025-12-11 1c95ce1
Updated some labels on a few plots and updated notebook as per MR.
