### 2025-12-11 738c745
Changed how colourbars are handled and added plotting functions to State.
