### 2025-12-11 7599d57
Added license to `plotting/__init__.py` and added a `try except` statement for setting the `usetex`
