### 2025-12-11 949d18d
Added way of writing a symbol if one is not present using the `quantity`.
