### 2025-12-11 e58f4ea
Updated typing in plotting inputs and fixed some docstring issues.
