### 2025-12-11 3346a90
more tutorial cleanup
