+++ docs/requirements.txt
@@ -12,3 +12,5 @@ graphviz
 
 matplotlib
 plotly<6.0 # plotly 6.0 has issues with myst-nb and won't show plots in notebook
+simsopt # for running the quasr case
+joblib # for parallel fieldline tracing
