### 2025-12-12 64e0b9f
addressed some review remarks: - simplified tutorials to reduce run-time - extended tests with fieldline tracing and fixtures - included asfortranarray for biot-savart calls - added dimension checks for coils routines - removed joblib from the tutorial requirements - removed extended run-time limit for tutorial notebooks
