+++ test-CI/unit-pygvec/test_biotsavart.py
@@ -3,14 +3,60 @@ import pytest
 try:
     import numpy as np
     from typing import Literal
-    from gvec.coils import Coil, CoilSet
+    from gvec.coils import Coil, CoilSet, trace_fieldlines
     import tempfile
     from scipy.constants import mu_0
     from scipy.special import ellipk, ellipe
 
     rng = np.random.default_rng(0)
 except ImportError:
-    pytest.skip("Import Error", allow_module_level=True)
+    pass
+
+
+@pytest.fixture
+def coil_parameters():
+    params = dict(n_coilpoints=50000, n_evals=1000, coil_radius=1, coil_current=1, n_coils=10)
+    return params
+
+
+@pytest.fixture
+def off_axis_eval_pos(coil_parameters):
+    pos = rng.uniform(-0.8, 0.8, [3, coil_parameters["n_evals"]])
+    return pos
+
+
+@pytest.fixture
+def on_axis_eval_pos(coil_parameters):
+    z = np.linspace(-1, 1, coil_parameters["n_evals"])
+    pos = np.zeros((3, len(z)))
+    pos[2, :] = z
+    return (pos, z)
+
+
+@pytest.fixture
+def circular_coil(coil_parameters):
+    params = coil_parameters
+    circ_coil = create_circular_coil(
+        params["coil_radius"], params["coil_current"], params["n_coilpoints"]
+    )
+    return circ_coil
+
+
+@pytest.fixture
+def circular_coil_set(coil_parameters):
+    params = coil_parameters
+    circ_coils = []
+    shifts = np.arange(params["n_coils"])
+    for shift in shifts:
+        circ_coils.append(
+            create_circular_coil(
+                params["coil_radius"],
+                params["coil_current"],
+                params["n_coilpoints"],
+                shift=shift,
+            )
+        )
+    return CoilSet(circ_coils), shifts
 
 
 def circular_coil_axis_mod_B(z, R, coil_current):
@@ -116,41 +162,28 @@ def create_circular_coil(radius, current, n_coilpoints, shift=0):
     return circ_coil
 
 
-def test_BiotSavart_B_axis_circular_coil():
-    n_coilpoints = 10000
-    n_evals = 1000
-    coil_radius = 1
-    coil_current = 1
-    circ_coil = create_circular_coil(coil_radius, coil_current, n_coilpoints)
-    z = np.linspace(-1, 1, n_evals)
+def test_BiotSavart_B_axis_circular_coil(coil_parameters, on_axis_eval_pos, circular_coil):
+    params = coil_parameters
+    pos, z = on_axis_eval_pos
 
-    pos = np.zeros((3, len(z)))
-    pos[2, :] = z
-
-    mod_B_analytic = circular_coil_axis_mod_B(z, coil_radius, coil_current)
-    ds = circ_coil.eval_mod_B(pos)
+    mod_B_analytic = circular_coil_axis_mod_B(z, params["coil_radius"], params["coil_current"])
+    ds = circular_coil.eval_mod_B(pos)
 
     assert np.all(abs(ds.B.sel(xyz=["x", "y"])) <= 1e-12)
     np.testing.assert_allclose(ds.mod_B, mod_B_analytic)
 
 
-def test_BiotSavart_B_axis_circular_coil_set():
-    n_coilpoints = 10000
-    n_evals = 1000
-    coil_radius = 1
-    coil_current = 1
-    n_coils = 10
-    z = np.linspace(-1, 1, n_evals)
-    circ_coils = []
+def test_BiotSavart_B_axis_circular_coil_set(
[…]
