+++ python/gvec/coils.py
@@ -51,6 +51,61 @@ def _stack_for_BiotSavart(ds: xr.Dataset | xr.DataArray):
         return ds
 
 
+def _pos2dataarray(pos: ArrayLike):
+    """Check if pos has the proper size and stack it if it originates from an ev dataset.
+
+    Parameters
+    ----------
+    pos : ArrayLike
+        Position array in xyz.
+
+    Returns
+    -------
+    pos: xr.DataArray, stacked: bool
+        pos-array and flag to indicate that it has been stacked
+
+    Raises
+    ------
+    ValueError
+        If the shape of the array is not as expected.
+    """
+    stacked = False
+    if not isinstance(pos, xr.DataArray):
+        pos = _pos_atleast_2D(pos)
+        pos = xr.DataArray(
+            pos,
+            dims=("xyz", "points"),
+            coords={"xyz": ("xyz", ["x", "y", "z"])},
+            attrs=dict(long_name="position vector", symbol=r"\mathbf{x}"),
+        )
+    else:
+        pos = _stack_for_BiotSavart(pos)
+        stacked = True
+
+    if (pos.ndim > 2) or (pos.shape[0] != 3):
+        raise ValueError(
+            f"pos.ndim > 2! Expected array of shape (3,n_positions), but got shape {pos.shape}."
+        )
+
+    return pos, stacked
+
+
+def _unstack_BS_ds(ds):
+    ds = ds.unstack("points")
+    for dimname, actual_dimname in zip(["rho", "theta", "zeta"], ["rad", "pol", "tor"]):
+        if dimname in ds.dims:
+            ds = ds.rename_dims({dimname: actual_dimname})
+    return ds
+
+
+def _pos_atleast_2D(pos):
+    """Reshape position vector of shape (3,) to (3,1)"""
+    pos = np.asanyarray(pos)  # similar to np.atleast_2d, transform pos into a numpy array
+    if pos.ndim == 1:
+        pos = pos.reshape(-1, 1)
+    return pos
+
+
 class Coil:
     def __init__(self, coil_points: ArrayLike, coil_current: float):
         """Filament coil discretized via straight line segments.
@@ -63,12 +118,15 @@ class Coil:
             Current flowing in the coil in [A].
         """
         if not isinstance(coil_points, xr.DataArray):
-            coil_points = np.atleast_2d(coil_points)
-            self.coil_points = xr.DataArray(
+            coil_points = np.asanyarray(coil_points)
+            coil_points = xr.DataArray(
                 coil_points, dims=("xyz", "points"), coords={"xyz": ("xyz", ["x", "y", "z"])}
             )
-        else:
-            self.coil_points = coil_points
+        if (coil_points.ndim > 2) or (coil_points.shape[0] != 3):
+            raise ValueError(
+                f"Expected  coil_points of shape (3, n_points) but got {coil_points.shape}!"
+            )
+        self.coil_points = coil_points
         self.coil_current = coil_current
         self.n_points = self.coil_points.shape[1]
         self.e_vec = self.coil_points[:, 1:] - self.coil_points[:, :-1]
@@ -103,25 +161,17 @@ class Coil:
         xr.Dataset
             Dataset containing the magnetic field at the evaluation positions.
         """
-        if not isinstance(pos, xr.DataArray):
-            pos = np.asanyarray(pos)
-            if pos.ndim == 1:
-                pos = pos.reshape(-1, 1)
-            pos = xr.DataArray(
-                pos,
-                dims=("xyz", "points"),
-                coords={"xyz": ("xyz", ["x", "y", "z"])},
-                attrs=dict(long_name="position vector", symbol=r"\mathbf{x}"),
-            )
+        pos, stacked = _pos2dataarray(pos)
+
[…]
