### 2025-12-12 87f2af0
addressed some review remarks: - added copyright to plotting and coils - adapted notebooks
