+++ python/gvec/plotting.py
@@ -1,3 +1,5 @@
+# Copyright (c) 2025 GVEC Contributors, Max Planck Institute for Plasma Physics
+# License: MIT
 from gvec.core.state import State
 import logging
 import numpy as np
