### 2025-12-12 5440ef6
Updated plotting in response to MR comments
- Added docstring to `_symbol_check`
- Added imports to `gvec/__init__.py`
- Fixed some typos and changed plotting `__init__.py`
- Rearranged arguments to be consistent with `state.evaluate`
- Removed `plot_boundary` and set default `rho=1.0` on `plot_3d_surface`
- Removed old `.py` example file
- Updated inputs in `State` class
