+++ python/gvec/plotting/plots2d.py
@@ -18,8 +18,9 @@ def plot_poloidal_plane(
     zeta: int | float | np.ndarray = 9,
     subplot_grid: list[int] | None = None,
     share_axis: bool = False,
-    st_contours: list[int] = [0, 0],
-    theta_contour_style: Literal["theta", "theta_star"] = "theta",
+    rho_contours: int = 4,
+    theta_contours: int = 8,
+    sfl: Literal["pest"] | None = "pest",
 ):
     """
     plot_poloidal_plane
@@ -46,22 +47,41 @@ def plot_poloidal_plane(
     share_axis : bool
         If `True`, all subplots will share their `X1` and `X2` axis positions.
         Default `False`
-    st_contours : list[int]
-        Number of ``$\theta$`` and ``$\zeta$`` contours to plot.
-        Default `[0,0]`
-    theta_contour_style : str
-        Contours of `theta` should be either `"theta"` or `"theta_star"`.
-        Default `"theta"`.
+    rho_contours : int, optional
+        The number of ``$\rho$`` contours to plot.
+        Default `4`
+    theta_contours : int, optional
+        The number of ``$\theta$`` contours to plot.
+        Default `8`
+    sfl : `"pest"` or `None`, optional
+        Plot the `theta` contours or `pest` contours (``$\theta^\star$``).
+        Default `"pest"`.
 
     Returns
     -------
     `matplotlib.pyplot.figure` object and `numpy.ndarray` of `matplotlib.axis._axis.Axes` object(s).
     """
 
-    if theta_contour_style not in ["theta", "theta_star"]:
-        raise ValueError("theta_contour_type must be 'theta' or 'theta_star'")
-
     evaluations = state.evaluate(quantity, "X1", "X2", "LA", rho=nrho, theta=ntheta, zeta=zeta)
+
+    if rho_contours:
+        ev_rho_contours = state.evaluate(
+            "X1", "X2", rho=rho_contours, theta=np.linspace(0, 2 * np.pi, ntheta), zeta=zeta
+        )
+    if theta_contours:
+        if sfl == "pest":
+            ev_theta_contours = state.evaluate_sfl(
+                "X1",
+                "X2",
+                rho=np.linspace(0, 1, nrho),
+                theta=theta_contours,
+                zeta=zeta,
+                sfl="pest",
+            )
+        else:
+            ev_theta_contours = state.evaluate(
+                "X1", "X2", rho=np.linspace(0, 1, nrho), theta=theta_contours, zeta=zeta
+            )
 
     zeta_eval = evaluations.zeta.data
 
@@ -110,32 +130,16 @@ def plot_poloidal_plane(
         )
         if share_axis:
             ax.label_outer()  # Removes any axis labels on subplots on the interior of the grid
-        if st_contours[0] > 0:
+        if rho_contours:
             # We should plot the rho
-            rho_contour = evaluations.X1[:, :, i] * 0.0 + evaluations.rho  # TODO: expand dims?
-            rho_levels_vis = np.linspace(0, 1 - 1e-10, st_contours[0])
-            f_ax = ax.contour(
-                evaluations.X1.data[:, :, i],
-                evaluations.X2.data[:, :, i],
-                rho_contour,
-                rho_levels_vis,
-                colors="black",
+            ax.plot(
+                ev_rho_contours.X1[:, :, i].T, ev_rho_contours.X2[:, :, i].T, "w", linewidth=1.0
             )
 
-        if st_contours[1] > 0:
+        if theta_contours:
             # \theta or \theta^\star contours
-            if theta_contour_style == "theta":
-                prefactor = 0.0
-            else:
-                prefactor = 1.0
[…]
