+++ python/gvec/plotting/__init__.py
@@ -3,10 +3,10 @@
 try:
     import matplotlib.pyplot as plt
 
-    plt.rcParams.update({"text.usetex": True})
+    plt.rcParams["text.usetex"] = True
 except ImportError:
     pass
 
-from .plots1d import plot_on_axis, plot_radial_profile
-from .plots2d import plot_on_flux_surface, plot_poloidal_plane
-from .plots3d import plot_3d_surface, plot_boundary
+from gvec.plotting.plots1d import plot_on_axis, plot_radial_profile
+from gvec.plotting.plots2d import plot_on_flux_surface, plot_poloidal_plane
+from gvec.plotting.plots3d import plot_3d_surface

[…]
