### 2025-12-12 8d6ddc6
Changed label behavior in `plot_on_flux_surface.
