### 2025-12-12 0d93ebc
Updated documentation
- Updated `conf.py` with the plotly `js`
- Updated `getting_started.md` with new tutorial file
- Updated `index.md` with new tutorial
- Updated `pyproject.toml` dependencies
- Removed old `examples` folder
