+++ docs/user/index.md
@@ -20,6 +20,7 @@ coordinate-conventions
 /tutorials/notebooks/030_stages
 /tutorials/notebooks/040_current
 /tutorials/notebooks/050_pygvec
+/tutorials/notebooks/051_plotting
 ```
 
 ```{toctree}
