### 2025-12-15 c510c9e
enable testing of f90wrap pre-release
