+++ CMakeLists.txt
@@ -417,6 +417,7 @@ IF(COMPILE_PYGVEC)
     src/pygvec/binding.f90
     src/pygvec/state.f90
     src/pygvec/run.f90
+    src/globals/newton.f90
     src/transform_sfl/sfl_boozer.f90
     src/profiles/c_rprofile.f90
     src/profiles/rprofile_bspline.f90
@@ -437,7 +438,7 @@ IF(COMPILE_PYGVEC)
     COMMAND "${PYTHON_EXECUTABLE}"
             -m f90wrap
             ${pygvec_src_files}
-            --verbose
+            -vv
             --mod-name lib  # -> lib.py - autogenerated python wrappers
             --f90-mod-name gvec._libpygvec  # -> _libpygvec.cpython*.so - compiled python module
             --kind-map "${CMAKE_CURRENT_SOURCE_DIR}/python/kind_map.py"
