### 2025-12-16 7f2bfaa
add twist computation. fix derivations
* twist computed from curve + linear independent direction. Using dft for first and second derivatives.
* test for linking number from boundary test case added
* test for writhe from two curves of boundary test case, comparison possible of Wr=Lk-Tr, to approximate writhe from double integral (200 points per field period, within accuracy limit of 1e-3).
