+++ python/gvec/util.py
@@ -983,7 +983,9 @@ def compute_FD(f: np.ndarray, pos, coefs, axis=0):
         axis  : axis along which the FD is computed, default is 0
     Returns:
         df    : Finite-Difference result, same shape as f (see warning above!)
+
     Examples:
+
     - examples for first derivative of f:
         - 1st order forward FD: `pos=[1,0]; coefs=[-1,1]/(dx)`
         - 2nd order central FD: `pos=[-1,1]; coefs=[-1,1]/(2*dx)`
@@ -1207,31 +1209,53 @@ def boundary_generator(case: str, X1_00=1.0, a0=0.5, ellipticity=0.4, helix_r=0.
     return params
 
 
-def linking_number(curve_a: np.ndarray, curve_b: np.ndarray, tol=1e-15):
-    """
+def linking_number(curve_a: np.ndarray, curve_b: np.ndarray, tol=1e-15, endpoint=True):
+    r"""
     Compute the linking number of two non-intersecting curves $C_a(\zeta_a),C_b(\zeta_b)$, solving the (non-singular) double integral over two curves
 
-    $$\textrm{Lk} = \Lk = \frac{1}{4\pi}\int_0^{2\pi} \int_0^{2\pi} \frac{ r_b - r_a}{\abs{r_b - r_a}^3} \cdot \left(\frac{\partial r_a}{\zeta_a}  \times \frac{\partial r_b}{\zeta_b}} d{\zeta_a} d{\zeta_b}$$
+    $$
+    \text{Lk} = \frac{1}{4\pi}\int_0^{2\pi} \int_0^{2\pi} \frac{ r_b - r_a}{\abs{r_b - r_a}^3} \cdot \left(\frac{\partial r_a}{\partial\zeta_a}  \times \frac{\partial r_b}{\partial\zeta_b}\right) d{\zeta_a} d{\zeta_b}
+    $$
 
     Here, the double integral is approximated by representing the curves as polygons, and computing the linking number as the sum of solid angles between all linear segments of polygon curves. The solid angle of two linear segments can be computed exactly. See paper by  [Klenin and Langowski](https://onlinelibrary.wiley.com/doi/10.1002/1097-0282(20001015)54:5%3C307::AID-BIP20%3E3.0.CO;2-Y)
     See GVEC documentation for more details.
-    Args:
-        curve_a (2D np.ndarray): the x,y,z point positions of the first curve. First and last point must coincide. shape must be [npoints_a,3]
-        curve_b (2D np.ndarray): the x,y,z point positions of the first curve. First and last point must coincide. shape must be [npoints_b,3]
-        tol (float): tolerance to consider points coincident (default: 1e-15)
-    Returns:
-        Lk (float): the linking number of the two curves
+
+    Parameters
+    ----------
+    curve_a : np.ndarray
+        the x,y,z point positions of the first curve. First and last point must coincide, if endpoint=True. shape must be [npoints_a,3]
+    curve_b : np.ndarray
+        the x,y,z point positions of the first curve. First and last point must coincide, if endpoint=True. shape must be [npoints_b,3]
+    tol : float
+        tolerance to consider points coincident (default: 1e-15)
+    endpoint : bool
+        `True`:  the first and last point of each curve coincide (default), else the last point is connected to the first point to close the curve.
+
+    Returns
+    -------
+    Lk : float
+        the linking number of the two curves
     """
     assert curve_a.ndim == 2, "curve_a must be a 2D array"
     assert curve_b.ndim == 2, "curve_b must be a 2D array"
[…]
