+++ test-CI/unit-pygvec/test_util.py
@@ -218,3 +218,32 @@ def test_linking_number(npoints):
     assert np.allclose(util.linking_number(curve_a, curve_b), 0.0), (
         "linking number of two unlinked circles should be 0.0"
     )
+
+
+@pytest.mark.parametrize(
+    "case, Lk_expected",
+    [
+        ("ellip_cyl_helix", 0),
+        ("ellip_cyl_helix_rot", 1),
+        ("ellip_cyl_helix_rot2", 0),
+    ],
+    ids=["no_link", "link", "no_link2"],
+)
+def test_linking_number_boundary(case, Lk_expected, endpoint):
+    params = gvec.util.boundary_generator(case)
+    nfp = 3
+    params["nfp"] = nfp
+    theta = np.linspace(0, 2 * np.pi, 21, endpoint=False)
+    for phi_dir in [-1, 1]:
+        zeta = np.linspace(0, 2 * np.pi, params["nfp"] * 81, endpoint=endpoint)
+        X1, X2 = gvec.util.evaluate_boundary(theta, zeta, params)
+        xyz_surf = np.zeros((len(zeta), len(theta), 3))
+        xyz_surf[:, :, 0] = X1.T * np.cos(phi_dir * zeta[:, None])
+        xyz_surf[:, :, 1] = X1.T * np.sin(phi_dir * zeta[:, None])
+        xyz_surf[:, :, 2] = X2.T
+        Lk = gvec.util.linking_number(
+            xyz_surf[:, 0, :], xyz_surf[:, len(theta) // 2, :], endpoint=endpoint
+        )
+        assert np.abs(Lk - phi_dir * Lk_expected * nfp) < 1e-8, (
+            f"Linking number of the surface is not as expected! Got {Lk}, expected {phi_dir * Lk_expected * nfp}, for phi direction {-phi_dir}"
+        )
