+++ docs/api/scripts-gist.rst
@@ -1,0 +1,9 @@
+.. _api-scripts-gist:
+
+gvec.scripts.gist
+==================
+
+.. automodule:: gvec.scripts.gist
+  :members:
+  :undoc-members:
+  :show-inheritance:
