+++ docs/api/gframe.rst
@@ -1,0 +1,9 @@
+.. _api-gframe:
+
+gvec.gframe
+=========
+
+.. automodule:: gvec.gframe
+  :members:
+  :undoc-members:
+  :show-inheritance:

[…]
