### 2025-12-16 9eda2bd
ifft1d added. all fft1d can be applied along axis of multi-d array.
