+++ python/gvec/fourier.py
@@ -16,15 +16,18 @@ import numpy as np
 # === Transform functions === #
 
 
-def fft1d(x: Iterable, angle0=0.0):
+def fft1d(y: Iterable, x0=0.0, axis=None):
     """
-    Compute the Fourier transform of a 1D array.
+    Compute the Fourier transform in 1d from data values at equidistant points $y(x_i)$, along a given axis of an array.
 
     Parameters
     ----------
-    x
-        Input array to transform, assumed to be sampled on `angle=angle0+np.linspace(0,2*np.pi,len(x),endpoint=False)`.
-    angle0: starting value of angle, where x was sampled, in [0,2pi] , defaults to 0.
+    y
+        Input array to transform $y(x_i)$, assumed to be sampled on `x=x0+np.linspace(0,2*np.pi,len(x),endpoint=False)`.
+        y is allowed to have multiple dimensions, but the fft will be applied only along the given `axis`.
+    x0: starting value of angle, where x was sampled, in [0,2pi] , defaults to 0.
+    axis : int
+        Axis along which to compute the fft. None by default. Becomes mandatory if y is multi-dimensional.
 
     Returns
     -------
@@ -37,16 +40,27 @@ def fft1d(x: Iterable, angle0=0.0):
     -----
     The function uses the real-input fast Fourier transform (rfft) from numpy.
     """
-    x = np.asarray(x)
-    xf = np.fft.rfft(x, norm="forward")
-    if angle0 != 0.0:
-        ks = np.arange(xf.shape[0])
-        xf *= np.exp(-1j * angle0 * ks)
+    y = np.asarray(y)
+    if y.ndim == 1:
+        _axis = 0
+    else:
+        if axis is None:
+            raise ValueError("axis must be specified if x is multi-dimensional")
+        _axis = axis
 
-    c = xf.real
-    c[1:] *= 2
-    s = -2 * xf.imag
-    s[0] = 0
+    cs = np.fft.rfft(y, norm="forward", axis=_axis)
+    if x0 != 0.0:
+        ks = np.expand_dims(
+            np.arange(cs.shape[_axis]), axis=tuple(i for i in np.arange(y.ndim) if i != _axis)
+        )
+        cs *= np.exp(-1j * x0 * ks)
+    # select 1: along given _axis with a slice
+    c = cs.real
+    sl = tuple(slice(0, None) if i != _axis else slice(1, None) for i in np.arange(y.ndim))
+    c[sl] *= 2
+    s = -2 * cs.imag
+    sl = tuple(slice(0, None) if i != _axis else slice(0, 1) for i in np.arange(y.ndim))
+    s[sl] = 0
     return c, s
 
 
@@ -77,6 +91,60 @@ def shift_1d(y: np.ndarray, x0, axis, newshape=None):
     cshft = c * np.exp(-1j * ks * x0)
     yshft = np.fft.irfft(cshft, newshape, norm="forward", axis=axis)
     return yshft
+
+
+def ifft1d(
+    c: np.ndarray,
+    s: np.ndarray,
+    npoints: int = None,
+    deriv: str | None = None,
+    nfp: int = 1,
+    axis=None,
+) -> np.ndarray:
+    r"""
+    Compute the inverse Fast-Fourier-Transform of a 1D Fourier series.
+
+    Parameters
+    ----------
+    c : numpy.ndarray
+        Cosine coefficients of the Fourier series.
+    s : numpy.ndarray
+        Sine coefficients of the Fourier series.
+    npoints : int
+        Number of points at which to evaluate the series. If None, n=2*m_max+1 is used
+    deriv : int, optional
+        Derivative to evaluate, by default None.
+    nfp : int, optional
[…]
