+++ test-CI/unit-pygvec/test_fourier.py
@@ -61,14 +61,66 @@ def ev(state):
 )
 @pytest.mark.parametrize("points", [10, 11], ids=["even", "odd"])
 @pytest.mark.parametrize("shift", [-0.2, 0, 0.5])
-def test_fft1d(c, s, points: tuple[int, int], shift):
+@pytest.mark.parametrize("axis", [None, 0, 1, 2], ids=["1d", "2dx", "2dy", "3dz"])
+def test_fft1d(c, s, points: tuple[int, int], shift, axis):
     t = np.linspace(0, 2 * np.pi, points, endpoint=False)
     if shift != 0.0:
         t = t + shift * t[1]
     x = sum([ci * np.cos(i * t) for i, ci in enumerate(c)])
     x += sum([si * np.sin(i * t) for i, si in enumerate(s)])
+    # test 1d array and 2d array in two directions
+    if axis is None:
+        xc, xs = fourier.fft1d(x, x0=t[0])
+    elif axis == 0:
+        x2d = np.vstack([x, x]).T
+        xc2d, xs2d = fourier.fft1d(x2d, x0=t[0], axis=0)
+        assert xc2d.ndim == x2d.ndim, (
+            "Problem with 1d fft of 2d array along axis 0, not same number of dimensions "
+        )
+        assert xc2d.shape == xs2d.shape, (
+            "Problem with 1d fft of 2d array along axis 0, not same shape "
+        )
+        assert np.allclose(xc2d[:, 0], xc2d[:, 1]), (
+            "Problem with 1d fft of 2d array along axis 0 "
+        )
+        assert np.allclose(xs2d[:, 0], xs2d[:, 1]), (
+            "Problem with 1d fft of 2d array along axis 0 "
+        )
+        xc, xs = xc2d[:, 0], xs2d[:, 0]
+    elif axis == 1:
+        x2d = np.vstack([x, x])
+        xc2d, xs2d = fourier.fft1d(x2d, x0=t[0], axis=1)
+        assert xc2d.ndim == x2d.ndim, (
+            "Problem with 1d fft of 2d array along axis 1, not same number of dimensions "
+        )
+        assert xc2d.shape == xs2d.shape, (
+            "Problem with 1d fft of 2d array along axis 1, not same shape "
+        )
+        assert np.allclose(xc2d[0, :], xc2d[1, :]), (
+            "Problem with 1d fft of 2d array along axis 1 "
+        )
+        assert np.allclose(xs2d[0, :], xs2d[1, :]), (
+            "Problem with 1d fft of 2d array along axis 1 "
+        )
+        xc, xs = xc2d[0, :], xs2d[0, :]
+    elif axis == 2:
+        x2d = np.vstack([x, x, x])
+        x3d = np.stack([x2d, x2d])
+        xc3d, xs3d = fourier.fft1d(x3d, x0=t[0], axis=2)
+        assert xc3d.ndim == x3d.ndim, (
+            "Problem with 1d fft of 3d array along axis 2, not same number of dimensions "
+        )
+        assert xc3d.shape == xs3d.shape, (
+            "Problem with 1d fft of 3d array along axis 2, not same shape "
+        )
+        assert np.allclose(xc3d[0, 0, :], np.average(xc3d, axis=(0, 1))), (
+            "Problem with 1d fft of 3d array along axis 2 "
+        )
+        assert np.allclose(xs3d[0, 0, :], np.average(xs3d, axis=(0, 1))), (
+            "Problem with 1d fft of 3d array along axis 2 "
+        )
[…]
