### 2025-12-16 aede877
replace dft with fft for twist computation.
* set `endpoint=False` as default for the line segment methods (linking number and write) and fixed the tests.
