### 2025-12-17 a73af85
a bit faster algorithm for line segment method.
