### 2025-12-19 #9670 @editorialbot
[REVIEW]: GVEC: A flexible 3D MHD equilibrium solver
**Submitting author:** @fhindenlang (<a href="http://orcid.org/0000-0002-0439-249X">Florian Hindenlang</a>)
**Repository:** https://github.com/gvec-group/gvec
**Branch with paper.md** (empty if default branch): joss_paper_2025
**Version:** v1.2.0
