### 2026-01-09 3d8e87c
remove module load of anaconda for mpcdf images.
- anaconda is not available anymore on newest images. replaced by python-waterboa.
- for older compilers, which are still checked for in the CI, images are from <=2024, and only have anaconda.
  added new setup scripts marked `mpcdfcirunner_setup_2024_*` to be used by those environements, still using anaconda.
