### 2026-01-10 96ff084
update mpcdf shared runner tags.
