+++ .gitlab-ci.yml
@@ -1330,9 +1330,8 @@ include:
 # template for all mpcdfci jobs
 .mpcdfci:
   tags:
-    - cloud
-    - docker
-    - shared
+    - mpcdf-shared
+    - cpu
   variables:
     CMAKE_HOSTNAME: "mpcdfcirunner"
     CI_HOSTNAME: "mpcdfci"
@@ -1588,9 +1587,8 @@ include:
 
 .ubuntu-22:
   tags:
-    - cloud
-    - docker
-    - shared
+    - mpcdf-shared
+    - cpu
   image: gitlab-registry.mpcdf.mpg.de/gvec-group/gvec/ubuntu-22:latest
   extends:
     - .vars_mpiOFF
@@ -1603,9 +1601,8 @@ include:
 
 .ubuntu-24:
   tags:
-    - cloud
-    - docker
-    - shared
+    - mpcdf-shared
+    - cpu
   image: gitlab-registry.mpcdf.mpg.de/gvec-group/gvec/ubuntu-24:latest
   extends:
     - .vars_mpiOFF
@@ -1618,9 +1615,8 @@ include:
 
 .fedora:
   tags:
-    - cloud
-    - docker
-    - shared
+    - mpcdf-shared
+    - cpu
   image: gitlab-registry.mpcdf.mpg.de/gvec-group/gvec/fedora:latest
   extends:
     - .vars_mpiOFF
