### 2026-01-13 390aa71
fix in docstring example code.
