### 2026-01-14 835c9ca
extend CI scripts, latest compilers allowed to fail.
* change artifacts for python build from the virtual environment folder to the folder of a generated wheel.
  then, in the job that needs gvec, create a fresh environment and install from the wheel, which works since its done on the same architecture.
* minimal pipeline is now DEFAULT (MINIMAL_PIPELINE="true").
  only gfortran-14 is part of the minimal pipeline and ifx-2024 on raven and viper.
* main tests are done with gfortran-14 and ifx-2024. remove old ifort-2023
* all jobs with "latest" compilers are allowed to fail
* separate CI setup 
[…]
