+++ .gitlab-ci.yml
@@ -44,9 +44,9 @@ variables:
     description: "The reference branch/tag to compare against"
   CI_HOSTNAME: ${CMAKE_HOSTNAME}
   MINIMAL_PIPELINE:
-    value: ""
-    options: ["", "true", "false"]
-    description: "''true'' to run only minimal tests, ''false'' to run all tests, '''' to run only minimal tests for ''.*gitlab[-_]ci.*'' branches"
+    value: "true"
+    options: ["true", "false"]
+    description: "''true'' to run only minimal tests, ''false'' to run all tests"
   NIGHTLY_PIPELINE:
     value: "false"
     options: ["true", "false"]
@@ -125,51 +125,47 @@ include:
 
   - local: CI_templates/env.yml
     inputs:
-      env: .mpcdfci_ifort-2023
-      rules: .rules_minimal
-
-  - local: CI_templates/env.yml
-    inputs:
       env: .mpcdfci_ifx-2024
 
   - local: CI_templates/env.yml
     inputs:
       env: .mpcdfci_ifx-latest
-      rules: .rules_minimal
+
 
   # GNU
-  - local: CI_templates/env.yml
-    inputs:
-      env: .mpcdfci_gfortran-12
-
   - local: CI_templates/env.yml
     inputs:
       env: .mpcdfci_gfortran-13
 
   - local: CI_templates/env.yml
     inputs:
-      env: .mpcdfci_gfortran-latest
+      env: .mpcdfci_gfortran-14
       rules: .rules_minimal
 
+
+  - local: CI_templates/env.yml
+    inputs:
+      env: .mpcdfci_gfortran-latest
+
   # Nvidia
+
   - local: CI_templates/env.yml
     inputs:
       env: .mpcdfci_nvfortran-23
 
-  - local: CI_templates/env.yml
-    inputs:
-      env: .mpcdfci_nvfortran-24
-
   # MPI
+
   - local: CI_templates/env.yml
     inputs:
-      env: .mpcdfci_ifort-impi-2023
-      rules: .rules_minimal
+      env: .mpcdfci_ifx-impi-2024
 
   - local: CI_templates/env.yml
     inputs:
       env: .mpcdfci_ifx-impi-latest
-      rules: .rules_minimal
+
+  - local: CI_templates/env.yml
+    inputs:
+      env: .mpcdfci_gfortran-openmpi-14
 
   - local: CI_templates/env.yml
     inputs:
@@ -197,20 +193,13 @@ include:
   - local: CI_templates/env.yml
     inputs:
       env: .raven_ifx-2024
-      rules: .rules_private_runner
+      rules: .rules_private_runner_minimal
 
   # GNU
 
   - local: CI_templates/env.yml
     inputs:
       env: .raven_gfortran-13
-      rules: .rules_private_runner
-
-  # Nvidia
-
-  - local: CI_templates/env.yml
-    inputs:
-      env: .raven_nvfortran-23
       rules: .rules_private_runner
 
   # MPI
@@ -269,17 +258,11 @@ include:
 
   - local: CI_templates/build.yml
     inputs:
-      env: .mpcdfci_ifort-2023
-      rules: .rules_minimal
-
-  - local: CI_templates/build.yml
-    inputs:
       env: .mpcdfci_ifx-2024
 
   - local: CI_templates/build.yml
     inputs:
       env: .mpcdfci_ifx-latest
-      rules: .rules_minimal
 
   - local: CI_templates/build.yml
     inputs:
@@ -299,12 +282,13 @@ include:
 
   - local: CI_templates/build.yml
     inputs:
-      env: .mpcdfci_gfortran-12
+      env: .mpcdfci_gfortran-13
[…]
