### 2026-01-14 8f8bd62
rename defines.h to defines.FPP
.FPP is the standard extension for the fortran preprocessor, while .h is used for C header files
