+++ src/base/bsplines/sll_m_assert.F90
@@ -15,7 +15,7 @@
 !  "http://www.cecill.info".
 !**************************************************************
 #ifndef DOXYGEN_SHOULD_SKIP_THIS
-#include "defines.h"
+#include "defines.FPP"
 
 module sll_m_assert
 !+++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
