+++ src/base/cubic_spline.F90
@@ -10,7 +10,7 @@
 !
 ! You should have received a copy of the GNU General Public License along with GVEC. If not, see <http://www.gnu.org/licenses/>.
 !=================================================================================================================================
-#include "defines.h"
+#include "defines.FPP"
 
 !===================================================================================================================================
 !>
