+++ python/gvec/core/state.py
@@ -858,97 +858,6 @@ class State:
             self, *quantities, rho=rho, theta=theta, zeta=zeta, sfl=sfl, **boozer_kwargs
         )
 
-    def plot_radial_profile(
-        self,
-        quantities: str | list = ["iota", "p", "I_tor", "I_pol"],
-        nrho: int | np.ndarray = 100,
-        subplot_grid: list | None = None,
-        xaxis="rho",
-    ):
-        from gvec.plotting.plots1d import plot_radial_profile
-
-        return plot_radial_profile(self, quantities, nrho, subplot_grid, xaxis)
-
-    def plot_on_axis(
-        self,
-        quantities: str | list = "mod_B",
-        nzeta: int | np.ndarray = 51,
-        subplot_grid: list | None = None,
-    ):
-        from gvec.plotting.plots1d import plot_on_axis
-
-        return plot_on_axis(self, quantities, nzeta, subplot_grid)
-
-    def plot_poloidal_plane(
-        self,
-        quantity: str = "mod_B",
-        /,
-        nrho: int = 21,
-        ntheta: int = 51,
-        zeta: int | float | np.ndarray = 9,
-        subplot_grid: list | None = None,
-        share_axis: bool = False,
-        rho_contours: int = 4,
-        theta_contours: int = 8,
-        theta_contour_style: Literal["theta", "theta_star"] = "theta",
-    ):
-        from gvec.plotting.plots2d import plot_poloidal_plane
-
-        return plot_poloidal_plane(
-            self,
-            quantity,
-            nrho,
-            ntheta,
-            zeta,
-            subplot_grid,
-            share_axis,
-            rho_contours,
-            theta_contours,
-            theta_contour_style,
-        )
-
-    def plot_on_flux_surface(
-        self,
-        quantities: str | list[str] = "mod_B",
-        rho: float | np.ndarray | list = 1.0,
-        ntheta: int = 51,
-        nzeta: int = 51,
-        subplot_grid: list | None = None,
-        share_axis: bool = True,
-        levels: int | np.ndarray = 10,
-        sfl: Literal["pest", "boozer"] | None = "boozer",
-        style: Literal["contour", "filled-contour"] = "contour",
-        **boozer_kwargs,
-    ):
-        from gvec.plotting.plots2d import plot_on_flux_surface
-
-        return plot_on_flux_surface(
-            self,
-            quantities,
-            rho,
-            ntheta,
-            nzeta,
-            subplot_grid,
-            share_axis,
-            levels,
-            sfl,
-            style,
-            **boozer_kwargs,
-        )
-
-    def plot_3d_surface(
-        self,
-        quantity: str = "mod_B",
-        rho: float = 1.0,
-        ntheta: int = 41,
-        nzeta: int = 51,
-        full_surface: bool = False,
-        to_file: str | None = None,
-    ):
-        from gvec.plotting.plots3d import plot_3d_surface
-
-        return plot_3d_surface(self, quantity, rho, ntheta, nzeta, full_surface, to_file)
-
 
 # === Functions === #
 
