### 2026-01-15 59c4fb2
Responding to MR comments: - Removed `try except` from plotting `__init__.py`, we will now require   matplotlib to work - Fixed some small internal handling of axes - Added `plot_kwargs` to all 1D and 2D plotting functions for handling   `subplots(**kwargs)` - Added a decorator to switch `usetex` on for plotting only, this will   now not interfere with a users default setting
