### 2026-01-15 aa97ecc
add plot_fourier_on_surface
