### 2026-01-15 2b7024c
- Fixed some naming issues and plot labels. - Added docstring to `plot_fourier_on_surface`
