### 2026-01-15 ecaeee8
Added visualisation module `visu` for `pygvec`. Called by `pygvec visu`
