+++ python/gvec/scripts/__init__.py
@@ -1,4 +1,4 @@
 # Copyright (c) 2025 GVEC Contributors, Max Planck Institute for Plasma Physics
 # License: MIT
 
-from gvec.scripts import main, run, quasr, cas3d, gist
+from gvec.scripts import cas3d, gist, main, quasr, run, visu
