+++ python/gvec/scripts/main.py
@@ -4,14 +4,14 @@
 
 # === Imports === #
 
-import platform
-from pathlib import Path
-import logging
 import argparse
+import logging
+import platform
 from collections.abc import Sequence
+from pathlib import Path
 
 import gvec
-from gvec.scripts import cas3d, convert, gist, quasr, run
+from gvec.scripts import cas3d, convert, gist, quasr, run, visu
 
 # === Arguments === #
 
@@ -77,6 +77,14 @@ gist_parser = subparsers.add_parser(
     add_help=False,
 )
 
+plot_parser = subparsers.add_parser(
+    "visu",
+    help="Plot some things",
+    description=visu.parser.description,
+    parents=[visu.parser],
+    add_help=False,
+)
+
 
 # === Script === #
 
@@ -104,6 +112,9 @@ def main(args: Sequence[str] | argparse.Namespace | None = None):
 
     elif args.mode == "load-quasr":
         return quasr.main(args)
+
+    elif args.mode == "visu":
+        return visu.main(args)
 
 
 if __name__ == "__main__":
