### 2026-01-15 a279766
wrap fbase & add init / final
