+++ src/base/fbase.f90
@@ -20,7 +20,8 @@ USE MODgvec_Globals                  ,ONLY: TWOPI,wp,Unit_stdOut,abort,MPIRoot
 IMPLICIT NONE
 
 PRIVATE
-PUBLIC t_fbase,fbase_new,sin_cos_map
+PUBLIC t_fBase, sin_cos_map
+PUBLIC fBase_free ! required for f90wrap
 
 TYPE :: t_fBase
   !---------------------------------------------------------------------------------------------------------------------------------
@@ -69,7 +70,7 @@ TYPE :: t_fBase
   CONTAINS
 
   PROCEDURE :: init             => fBase_init
-  PROCEDURE :: free             => fBase_free
+  FINAL ::                         fBase_free
   PROCEDURE :: copy             => fBase_copy
   PROCEDURE :: compare          => fBase_compare
   PROCEDURE :: change_base      => fBase_change_base
@@ -87,6 +88,10 @@ TYPE :: t_fBase
 
 END TYPE t_fBase
 
+INTERFACE t_fBase
+  MODULE PROCEDURE fBase_new
+END INTERFACE t_fBase
+
 CHARACTER(LEN=8)   :: sin_cos_map(3)=(/"_sin_   ", &
                                        "_cos_   ", &
                                        "_sincos_" /)
@@ -101,7 +106,7 @@ CONTAINS
 !> allocate the type fBase
 !!
 !===================================================================================================================================
-SUBROUTINE fBase_new( sf, mn_max_in,mn_nyq_in,nfp_in,sin_cos_in,exclude_mn_zero_in)
+FUNCTION fBase_new(mn_max_in,mn_nyq_in,nfp_in,sin_cos_in,exclude_mn_zero_in) RESULT(sf)
 ! MODULES
 IMPLICIT NONE
 !-----------------------------------------------------------------------------------------------------------------------------------
@@ -113,16 +118,15 @@ IMPLICIT NONE
   LOGICAL         ,INTENT(IN   ) :: exclude_mn_zero_in !! =true: exclude m=n=0 mode in the basis (only important if cos is in basis)
 !-----------------------------------------------------------------------------------------------------------------------------------
 ! OUTPUT VARIABLES
-  TYPE(t_fBase), ALLOCATABLE,INTENT(INOUT)        :: sf !! self
+  TYPE(t_fBase) :: sf !! self
 !-----------------------------------------------------------------------------------------------------------------------------------
 ! LOCAL VARIABLES
 !===================================================================================================================================
-  ALLOCATE(sf)
   __PERFON("fbase_new")
   CALL sf%init(mn_max_in,mn_nyq_in,nfp_in,sin_cos_in,exclude_mn_zero_in)
 
   __PERFOFF("fbase_new")
-END SUBROUTINE fBase_new
+END FUNCTION fBase_new
 
 !===================================================================================================================================
[…]
