### 2026-01-16 0cd4592
Added `pygvec tovtk` to output a vtk file from cli.
