### 2026-01-16 3d34bde
fix t_sfl_boozer finalizer
* set sfl_boozer_free as FINAL subroutine
* set sfl_boozer_new as constructor
* fix unbinding
