+++ pyproject.toml
@@ -3,7 +3,7 @@
 [build-system]
 requires = [
     "scikit-build-core",
-    "f90wrap",
+    "f90wrap>=0.3.0",
     "numpy>=2",  # for f2py
 ]
 build-backend = "scikit_build_core.build"
@@ -14,7 +14,7 @@ name = "gvec"
 description = "Galerkin Variational Equilibrium Code"
 dependencies = [
     "PyYAML",
-    "f90wrap",
+    "f90wrap>=0.3.0",
     "netcdf4",
     "numpy>=2",
     "pyevtk",
