### 2026-01-16 5f702d6
fix sfl_boozer in gvec_to_jorek and state
