### 2026-01-20 07afa05
correct fortran file endings
* in #include must be .F90
* in f90wrap generated functions, its .f90
