### 2026-01-20 17a366d
remove  allocatable from TYPE(t_fbase) variables
* since t_fbase has now been made a "standard" type with the t_fbase function for initialization and a FINALIZE function
  the `type(t_base),allocatable` should not be used anymore, as long as the variable is a scalar.
* removed all these `allocatable` statements, meaning also the deallocate is not needed anymore
* the hmap frenet finalizer, `SDEALLOCATE` has to be used.
