+++ src/readstate/boundaryFromFile.f90
@@ -199,7 +199,7 @@ SUBROUTINE bff_convert_to_modes(sf,x1_fbase_in,x2_fbase_in,X1_b,X2_b,scale_minor
   CLASS(t_boundaryFromFile), INTENT(INOUT) :: sf !! self
   !-----------------------------------------------------------------------------------------------------------------------------------
   ! LOCAL VARIABLES
-  TYPE(t_fBase),ALLOCATABLE        :: X_fbase,Y_fbase
+  TYPE(t_fBase)                     :: X_fbase,Y_fbase
   INTEGER                           :: i,nIP,mIP,mn_max_pts(2)
   REAL(wp)                          :: xn(2,sf%ntheta*sf%nzeta)
   REAL(wp),ALLOCATABLE              :: xydofs(:,:),X12dofs(:,:)
@@ -260,7 +260,6 @@ SUBROUTINE bff_convert_to_modes(sf,x1_fbase_in,x2_fbase_in,X1_b,X2_b,scale_minor
                              -RESHAPE(sf%Y,(/sf%ntheta*sf%nzeta/))))
 
 
-  DEALLOCATE(X_fbase,Y_fbase)
   WRITE(UNIT_stdOut,'(A)')'  ... CONVERT BOUNDARY DONE.'
 END SUBROUTINE bff_convert_to_modes
 
