+++ src/gvec_to_jorek/gvec_to_jorek.f90
@@ -785,9 +785,6 @@ IMPLICIT NONE
   REAL(wp),DIMENSION(1:out_base%f%mn_IP) :: X1_IP,dX1ds_IP,dX1dthet_IP,dX1dzeta_IP,X1_IP_eps,dX1ds_IP_eps,dX1dthet_IP_eps,dX1dzeta_IP_eps
   REAL(wp),DIMENSION(1:out_base%f%mn_IP) :: X2_IP,dX2ds_IP,dX2dthet_IP,dX2dzeta_IP,X2_IP_eps,dX2ds_IP_eps,dX2dthet_IP_eps,dX2dzeta_IP_eps
   REAL(wp),DIMENSION(1:out_base%f%mn_IP) :: LA_IP, LA_IP_eps,field_out_IP
-!  TYPE(t_fbase),ALLOCATABLE          :: X1_fbase_nyq
-!  TYPE(t_fbase),ALLOCATABLE          :: X2_fbase_nyq
-!  TYPE(t_fbase),ALLOCATABLE          :: LA_fbase_nyq
 
   ! Variables used in local interpolations for finite difference calculation of current density
   REAL(wp) :: X1_int,dX1ds,dX1dthet,dX1dzeta
@@ -1261,15 +1258,6 @@ IMPLICIT NONE
     CALL out_base%free()
     DEALLOCATE(out_base)
   END IF
-
-  SDEALLOCATE(X1_fbase_nyq)
-  SDEALLOCATE(X2_fbase_nyq)
-  SDEALLOCATE(LA_fbase_nyq)
-  SDEALLOCATE(fbase_zeta)
-  SDEALLOCATE(X1_fbase_nyq)
-  SDEALLOCATE(X2_fbase_nyq)
-  SDEALLOCATE(LA_fbase_nyq)
-  SDEALLOCATE(fbase_zeta)
 
 
 END SUBROUTINE finalize_gvec_to_jorek
