### 2026-01-20 ba7b07b
- Updated some small things in the plotting to do with labels etc... - Fixed some issues with the `usetex` decorator - Changed `tovtk` to `to-paraview` - Updated some of the outputs in `vtk.py` - Added some documentation for `pygvec` CLI tools in the documentation   website.
