### 2026-01-20 gvec-group/gvec#2 @dpanici
[JOSS review]: Links inside of tutorials do not work when accessing the notebook locally
Hello!

Really nice code setup here, congrats on all of the fine work done so far.

I am going through the [JOSS](https://github.com/openjournals/joss-reviews/issues/9670#issuecomment-3719060210) review and will leave some issues on anything I find relating to install, docs, errors etc.

Something minor which I don't think must be addressed for the JOSS review is that some of the links within the notebook tutorials (which are quite good btw) will not work when used locally, as they are referring to the .md pages of the built docs website, so when a user clicks the link locally it will redirect to a 404 page. Perhaps these could be replaced with the URL of the built docs e.g. using `[here](https://gvec.readthedocs.io/latest/user/gvec-parameter-list.html)` instead of `[here](</user/gvec-parameter-list.md>)`
