### 2026-01-20 gvec-group/gvec#3 @dpanici
[JOSS review]: Improved Documentation for `state.evaluate` and `state.compute`?
https://github.com/openjournals/joss-reviews/issues/9670#issuecomment-3719060210

I really like the way you have setup the `evaluate` and `compute` methods, it makes it quite easy for the user to compute things. Something I noticed though, and correct me if I am wrong, but the methods don't seem to be documented in the [API](https://gvec.readthedocs.io/latest/api/core-state.html) documentation beyond the argument types and order? As currently it seems that they are mainly described inside the post processing tutorial.

Adding some more fleshed-out documentation here would make it easier to use these functions I think e.g. like how [evaluate_profile](https://gvec.readthedocs.io/latest/api/core-state.html#gvec.core.state.State.evaluate_profile) is.
