### 2026-01-20 gvec-group/gvec#5 @dpanici
[JOSS review]: Preferred method of contact for outside contributors
https://github.com/openjournals/joss-reviews/issues/9670#issuecomment-3719060210

As part of the JOSS review there is a line `Community guidelines: Are there clear guidelines for third parties wishing to 1. Contribute to the software 2. Report issues or problems with the software 3. Seek support`

I believe there are very clear guidelines for GVEC in each of these respects in the docs, but I did have a question, what is your preferred method that outside contributors contact you? make an issue on github requesting guest access and listing their email? using the matrix chat channel? Making this clear might help those who do not know any of the maintainers well already make the first step towards contributing.
