### 2026-01-20 gvec-group/gvec#6 @dpanici
[JOSS review]: Tutorial for non-cylindrical G-frame beyond the quasr database?
https://github.com/openjournals/joss-reviews/issues/9670#issuecomment-3719060210

It would be nice to have a tutorial for the G-frame beyond that of the quasr database tutorial (perhaps just using an arclength-based frame for some race-track-like coordinate axis?), for example how one could generate a G-frame based off of a pyQSC or pyQIC configuration?
