### 2026-01-20 #9670 @dpanici
Really great work from the gvec developers and @fhindenlang!

For me my only acceptance blocker of the issues I've linked is (with the second one maybe being something that I am just missing or misunderstanding)
-  https://github.com/gvec-group/gvec/issues/5  (just because the mpcdf account blocks users from making issues on the gitlab, so it is not clear what the developers prefer as the best way of contact to obtain an account and to begin development/contributions)
- https://github.com/gvec-group/gvec/issues/6 the JOSS paper for GVEC includes a result of an equilibrium solved in an axis-aligned frame but I don't believe there currently is a tutorial that shows how to do this. There is a tutorial showing how to obtain the G-frame from the QUASR database, but not how to run the code with the G-frame obtained in such a way. I may be missing something from the tutorials/documentation, and I also am not sure of how much extra work this entails, but I think that it would be beneficial to show a practical example of this as it is showcased as one of the biggest advantages to the new software.
