### 2026-01-21 #9670 @fhindenlang
dear @dpanici, thank you for opening the issues on github and for your suggestions. We will have a look at them in detail and work on them, then give a heads-up here.
