### 2026-01-21 e5588f4
fix gist npol to integer
