### 2026-01-21 73e42ba
fix gist: gvec metadata in comments
GENE does not ignore extra fields in the namelist, therefore the gvec metadata is added as comments.
