### 2026-01-21 86e4945
update python versions in pyproject.toml
