+++ pyproject.toml
@@ -39,7 +39,9 @@ requires-python = ">=3.10"
 classifiers = [
     "Development Status :: 3 - Alpha",
     "Programming Language :: Python :: 3.10",
+    "Programming Language :: Python :: 3.11", # not routinely tested
     "Programming Language :: Python :: 3.12",
+    "Programming Language :: Python :: 3.13",
     "Programming Language :: Fortran",
     "License :: OSI Approved :: MIT License",
     "Intended Audience :: Science/Research",
