### 2026-01-21 17f42c9
add ubuntu 22/24 to minimal CI
with python 3.10 and 3.12 respectively
