### 2026-01-21 cae9c6d
update license metadata
requires scikit-build-core v0.11.0 or newer
