+++ pyproject.toml
@@ -2,7 +2,7 @@
 # ============================================================================ #
 [build-system]
 requires = [
-    "scikit-build-core",
+    "scikit-build-core>=0.11.0",
     "f90wrap>=0.3.0",
     "numpy>=2",  # for f2py
 ]
@@ -34,8 +34,7 @@ maintainers = [
     {name = "Florian Hindenlang", email = "florian.hindenlang@ipp.mpg.de"},
     {name = "Robert Babin", email = "robert.babin@ipp.mpg.de"},
 ]
-requires-python = ">=3.10"
-# match-case, intrinsic typing union
+requires-python = ">=3.10" # match-case, intrinsic typing union
 classifiers = [
     "Development Status :: 3 - Alpha",
     "Programming Language :: Python :: 3.10",
@@ -43,14 +42,14 @@ classifiers = [
     "Programming Language :: Python :: 3.12",
     "Programming Language :: Python :: 3.13",
     "Programming Language :: Fortran",
-    "License :: OSI Approved :: MIT License",
     "Intended Audience :: Science/Research",
     "Topic :: Scientific/Engineering :: Physics",
 ]
 keywords = ["plasma", "mhd", "equilibrium"]
 readme.file = "README.md"
 readme.content-type = "text/markdown"
-license.file = "LICENSE.txt"
+license = "MIT"
+license-files = ["LICENSE.txt"]
 dynamic = ["version"]
 
 [project.urls]
@@ -69,7 +68,7 @@ pygvec = "gvec.scripts.main:main"
 
 # ============================================================================ #
 [tool.scikit-build]
-minimum-version = "0.8"
+minimum-version = "0.11"
 logging.level = "DEBUG"
 metadata.version.provider = "scikit_build_core.metadata.setuptools_scm"
 sdist.include = [
@@ -88,9 +87,9 @@ sdist.exclude = [
 ]
 build-dir = "pybuild"
 editable.rebuild = false
+build.verbose = true
 
 [tool.scikit-build.cmake]
-verbose = true
 version = ">=3.22"
 build-type = "Release"
 
