### 2026-01-22 004183c
checked notebooks for plotting
* hole at the axis was too big in visualization
  --> changed default: minimal rho from `rho[0]=0.1*rho[1]` to `rho[0]=min(1e-4, 0.1*rho[1])` in `compute.py`, which is used in state.evaluate
* added / checked docstrings of plotting functions
* added / checked functionality of the `plot_poloidal_plane`, to be able to plot the rho/theta grid only without a contour, with possibility of a user defined color.
* new plotting notebook cross-checked. Needs "plotly<6.0" to be installed.
  -> Added a requirements.txt for the tutorials
* reformatting notebooks, changed maxi
[…]
