+++ docs/conf.py
@@ -157,7 +157,15 @@ html_js_files = [
 # List of patterns, relative to source directory, that match files and
 # directories to ignore when looking for source files.
 # This pattern also affects html_static_path and html_extra_path.
-exclude_patterns = ["ford/ford.md", "ford/static/index.md", "generators", "_downloads", "build"]
+exclude_patterns = [
+    "ford/ford.md",
+    "ford/static/index.md",
+    "generators",
+    "_downloads",
+    "build",
+    "tutorials/requirements.txt",
+    "tutorials/notebooks/ruff.toml",
+]
 
 intersphinx_mapping = {
     "xarray": ("https://docs.xarray.dev/en/stable/", None),
