+++ python/gvec/plotting/plots2d.py
@@ -13,15 +13,16 @@ from gvec.plotting.utils import _deco_usetex, _design_subgrid, _subplots, _symbo
 @_deco_usetex
 def plot_poloidal_plane(
     state: State,
-    quantity: str = "mod_B",
-    /,
+    quantity: None | str = "mod_B",
     nrho: int = 21,
     ntheta: int = 51,
     zeta: int | float | np.ndarray | list[float] = 9,
     subplot_grid: list[int] | None = None,
     share_axis: bool = True,
     rho_contours: int = 4,
+    rho_contours_color: str = "white",
     theta_contours: int = 8,
+    theta_contours_color: str = "white",
     sfl: Literal["pest"] | None = "pest",
     plot_kwargs: dict = dict(),
 ):
@@ -34,7 +35,7 @@ def plot_poloidal_plane(
     ----------
     state : GVEC state object
     quantity : str, optional
-        The quantity to plot. Default is "mod_B"
+        The quantity to plot. Default is "mod_B". If `None`, no contours are plotted.
     nrho : int, optional
         The radial resolution of the slices.
         Default is `51`
@@ -51,11 +52,17 @@ def plot_poloidal_plane(
         If `True`, all subplots will share their `X1` and `X2` axis positions.
         Default `False`
     rho_contours : int, optional
-        The number of ``$\rho$`` contours to plot.
+        The number of ``$\rho$`` contours to plot. If `0`, no contours are plotted.
         Default `4`
+    rho_contours_color : str, optional
+        The color of the ``$\rho$`` contours.
+        Default `"white"`
     theta_contours : int, optional
-        The number of ``$\theta$`` contours to plot.
+        The number of ``$\theta$`` contours to plot. If `0`, no contours are plotted.
         Default `8`
+    theta_contours_color : str, optional
+        The color of the ``$\theta$`` contours.
+        Default `"white"`
     sfl : `"pest"` or `None`, optional
         Plot the `theta` contours or `pest` contours (``$\theta^\star$``).
         Default `"pest"`.
@@ -67,8 +74,11 @@ def plot_poloidal_plane(
     -------
     `matplotlib.pyplot.figure` object and `numpy.ndarray` of `matplotlib.axis._axis.Axes` object(s).
     """
+    quantities = ["X1", "X2", "LA"]
+    if quantity is not None:
+        quantities.append(quantity)
 
-    evaluations = state.evaluate(quantity, "X1", "X2", "LA", rho=nrho, theta=ntheta, zeta=zeta)
+    evaluations = state.evaluate(*quantities, rho=nrho, theta=ntheta, zeta=zeta)
 
     if rho_contours:
         ev_rho_contours = state.evaluate(
@@ -90,62 +100,70 @@ def plot_poloidal_plane(
             )
 
     zeta_eval = evaluations.zeta.data
-
-    if np.any(np.isnan(evaluations[quantity])):
-        # Sometimes on-axis quantity cannot be computed properly, to avoid erroring out we will just set them to zero
-        #   and warn the user that they should adjust if their plots look weird
-        evaluations[quantity].data[np.isnan(evaluations[quantity].data)] = 0.0
-        warn(
-            "NaNs detected in evaluated dataset, these have been set to 0. It is possible that on-axis quantity cannot be evaluated, a minimum rho value can be set with the min_rho input."
-        )
+    if quantity is not None:
+        if np.any(np.isnan(evaluations[quantity])):
+            # Sometimes on-axis quantity cannot be computed properly, to avoid erroring out we will just set them to zero
+            #   and warn the user that they should adjust if their plots look weird
+            evaluations[quantity].data[np.isnan(evaluations[quantity].data)] = 0.0
+            warn(
[…]
