### 2026-01-22 e7825d2
restructure of the FORTRAN minimize: - introduced minimizer type - removed c_functional.F90 and functional.F90 to reduce unnecessarily complicated structure for one functional type - removed which_functional input variable - removed global U, V, P, F sol_var_mhd3d variables - analyze and visu functions take explicit sol_var_mhd3d variables as input instead of importing them - renamed variables of the style U, V, P, F to dofs, velocity,temp_dofs, force - changed all t_functional types to t_functional_mhd3d - t_functional_mhd3d has now a minimizer object - extended minimize routine from restruct
[…]
