+++ src/functionals/mhd3d/mhd3d_minimize.F90
@@ -1,0 +1,464 @@
+!===================================================================================================================================
+! Copyright (c) 2025 GVEC Contributors, Max Planck Institute for Plasma Physics
+! License: MIT
+!===================================================================================================================================
+#include "defines.FPP"
+
+!===================================================================================================================================
+!>
+!!# Module **MHD3D minimize**
+!!
+!! CONTAINS INITIALIZATION minimizer for MHD3D functional
+!!
+!===================================================================================================================================
+MODULE MODgvec_MHD3D_minimize
+    ! MODULES
+    USE MODgvec_Globals, ONLY:wp,abort,UNIT_stdOut,fmt_sep,MPIRoot,enter_subregion,exit_subregion
+    USE MODgvec_Sol_Var_MHD3D,ONLY: t_sol_var_MHD3D
+    IMPLICIT NONE
+
+    ! PRIVATE
+    ! PUBLIC t_minimizer_mhd3d
+
+    !-----------------------------------------------------------------------------------------------------------------------------------
+    ! TYPES
+    !-----------------------------------------------------------------------------------------------------------------------------------
+    TYPE :: t_minimizer_mhd3d
+        !-------------------------------------------------------------------------------------------------------------------------------
+        LOGICAL   :: initialized, restart_iter, logger_is_initialized
+        INTEGER   :: JacCheck, MinType
+        INTEGER   :: iter,nStepDecreased,nSkip_Jac,nSkip_dw
+        INTEGER   :: lastoutputIter, logiter_ramp, logscreen
+        REAL(wp)  :: dt,deltaW, dW_allowed
+        REAL(wp)  :: t_pseudo,Fnorm(3),Vnorm(3),Fnorm0(3),Fnorm_old(3),W_MHD3D_0
+
+        ! Logging variables
+        REAL(wp) :: min_dt_out,max_dt_out,min_dw_out,max_dw_out,sum_dW_out
+        LOGICAL  :: DoCheckDistance, DoCheckAxis
+        INTEGER  :: outputIter, nlogScreen, logIter, logUnit, StartTimeArray(8)
+
+        TYPE(t_sol_var_MHD3D), ALLOCATABLE :: dofs(:)      !! degrees of freedom at levels (k-1),(k),(k+1)
+        TYPE(t_sol_var_MHD3D), ALLOCATABLE :: force(:)     !! force
+        TYPE(t_sol_var_MHD3D), ALLOCATABLE :: temp_dofs(:) !! temporary for update
+
+        ! acc. gradient descent variables
+        REAL(wp), ALLOCATABLE  :: tau(:)
+        INTEGER   :: ndamp
+        REAL(wp)  :: tau_bar
+        TYPE(t_sol_var_MHD3D), ALLOCATABLE :: velocity(:)       !! 'velocity' in minimizer
+        !-------------------------------------------------------------------------------------------------------------------------------
+        CONTAINS
+        PROCEDURE :: minimize         => MinimizeMHD3D_descent
[…]
