+++ src/pygvec/state.F90
@@ -6,14 +6,14 @@
 
 MODULE MODgvec_py_state
 
-USE MODgvec_c_functional, ONLY: t_functional
+USE MODgvec_MHD3D,        ONLY: t_functional_mhd3d
 USE MODgvec_base,         ONLY: t_base
 USE MODgvec_Globals,      ONLY: enter_subregion,exit_subregion,reset_subregion
 
 IMPLICIT NONE
 PUBLIC
 
-CLASS(t_functional), ALLOCATABLE :: functional
+CLASS(t_functional_mhd3d), ALLOCATABLE :: functional
 LOGICAL :: initialized = .FALSE.
 INTEGER :: nfp = 0
 
@@ -29,13 +29,12 @@ SUBROUTINE Init(parameterfile)
   USE MODgvec_Restart,        ONLY: InitRestart
   USE MODgvec_ReadInTools,    ONLY: FillStrings,GETLOGICAL,GETINT,IgnoredStrings
 !$ USE omp_lib
-  USE MODgvec_Functional,     ONLY: InitFunctional
-  USE MODgvec_MPI    ,ONLY  : par_Init
+  USE MODgvec_MHD3D,  ONLY: InitMHD3D
+  USE MODgvec_MPI    ,ONLY: par_Init
   ! INPUT/OUTPUT VARIABLES ------------------------------------------------------------------------------------------------------!
   CHARACTER(LEN=*) :: parameterfile
   ! LOCAL VARIABLES -------------------------------------------------------------------------------------------------------------!
-  INTEGER :: which_functional
-  !================================================================================================================================!
+  ! CODE ------------------------------------------------------------------------------------------------------------------------!
   CALL reset_subregion()
   CALL enter_subregion("startup")
   CALL par_Init() !USE MPI_COMM_WORLD
@@ -62,8 +61,8 @@ SUBROUTINE Init(parameterfile)
   CALL InitAnalyze()
 
   ! initialize the functional
-  which_functional = GETINT('which_functional',Proposal=1)
-  CALL InitFunctional(functional,which_functional)
+  ALLOCATE(t_functional_mhd3d :: functional)
+  CALL InitMHD3D(functional)
 
   ! print the ignored parameters
   CALL IgnoredStrings()
@@ -77,7 +76,9 @@ END SUBROUTINE Init
 !================================================================================================================================!
 SUBROUTINE InitSolution()
   ! CODE ------------------------------------------------------------------------------------------------------------------------!
+  CALL enter_subregion("initialize solution")
   CALL functional%InitSolution()
+  CALL exit_subregion("initialize solution")
 END SUBROUTINE InitSolution
 
 !================================================================================================================================!
@@ -86,14 +87,60 @@ SUBROUTINE ReadState(statefile)
   USE MODgvec_Output_Vars,    ONLY: outputLevel,ProjectName
   USE MODgvec_ReadState_Vars, ONLY: outputLevel_r
   USE MODgvec_Restart,        ONLY: RestartFromState
-  USE MODgvec_MHD3D_Vars,     ONLY: U
   ! INPUT/OUTPUT VARIABLES ------------------------------------------------------------------------------------------------------!
   CHARACTER(LEN=255) :: statefile
   ! CODE ------------------------------------------------------------------------------------------------------------------------!
[…]
