### 2026-01-22 6986bbc
fixed missing allocation of functional in rungvec
