### 2026-01-22 aca11c8
fix iota extrapolation for pandas 3.0
pandas 3.0 (for python 3.12+) breaks gvec as modifying an index variable is no longer allowed:
"assignment destingation is read-only"
"IndexVariable values cannot be modified"

To workaround we now copy and reassign the coordinate / index.
