+++ python/gvec/core/run.py
@@ -424,7 +424,12 @@ class Run:
             r2 = ev.isel(rad=1)
             r3 = ev.isel(rad=2)
             ev = ev.isel(rad=slice(2, None))
-            ev.rho.data[0] = 0.0  # = self.rho[0]
+
+            # workaround to modify xarray coordinate & index (with pandas >=3.0)
+            rho = ev.rho.data.copy()
+            rho[0] = 0.0
+            ev = ev.assign_coords(rho=("rad", rho)).set_xindex("rho")
+
             for var in ev.data_vars:
                 ev[var].data[0] = 3 * (r1[var].data - r2[var].data) + r3[var].data
 
