### 2026-01-22 21671f8
fix SyntaxWarning from LaTeX formulas
