### 2026-01-22 744c6af
- Fixed issue in adding rationals - Fixed `axis_equal=False` axis limits in  `plot_poloidal_plane`
