### 2026-01-22 #9670 @fhindenlang
Dear @piyueh, about the right "git" way to do the review:

Its clear that the answer to the reviewers questions are given in the issues opened by the reviewer.

But what about changes required to the code, to address the reviewer comments (like adding a tutorial in the documentation, which is part of the gvec repository).
Is it best to add them in the branch  `joss_paper_2025` directly, via a Pull request? We can add a tag to mark the current state of the branch, like a "joss_paper_submission" tag?

And another question:
what about fixes found in the meantime, like for newer python version aso. , or updates to our CI pipeline? Can we simply cherry-pick them and make PR to add them to the branch for the review?
