### 2026-01-22 67804ec
Fixed zeroth iteration log
