### 2026-01-22 f6d0778
- Fixed some typos in the tutorials - Fixed an issue with plotting the rationals in iota plots - Changed the behaviour of the rho and theta contour plots in   `plot_poloidal_plane`
