### 2026-01-22 31b6253
config pypi job to run with ubuntu-22 & more debug output
