### 2026-01-22 #9670 @piyueh
Hi @dpanici and @AnkitBarik, which branch are you reviewing for the source code (`main`, `v1.2.0`, or `joss_paper_2025`)?

Hi @fhindenlang, changes should be applied to the same branch that the reviewers actually reviewed (via PRs or direct commits, both are fine). This applies separately to _**source code**_ and to the _**paper files**_ (`paper.md` / `paper.bib`).

- **Paper files** (`paper.md`, `paper.bib`, figures, etc.):
  Paper edits should go to `joss_paper_2025`, as indicated in your submission information.

- **Source code:**
  If reviewers reviewed `v1.2.0` (the ideal case), apply all requested code changes on top of `v1.2.0`. In practice, some reviewers may overlook the version tag and review the source code in the `main` branch or `joss_paper_2025`. If that happens, apply code changes to the branch they reviewed.

**Bottom line:**
After acceptance, you will need to make a new release and archive it (e.g., on Zenodo or Figshare). That archived version must contain _**only**_ code that has been reviewed. Your published JOSS paper will only cover this specific version of software. The paper files do not need to be included (which is why they are maintained in a separate branch). Cherry-pick extra fixes (CI, new Python versions, etc.) into the review branch only if you want them included in the final archived release. If you do so, please notify reviewers that you have added changes beyond what they requested.

Hope this helps!
