### 2026-01-22 #9670 @dpanici
Ah, very sorry I think I was reviewing the most recent version of the `main` branch. But checking `joss_paper_2025` and looking back at my acceptance-blocking issues that I made, nothing would change if I had looked at that one instead
