### 2026-01-22 c55f008
fix auto-documentation of api with f90wrap>=0.3.0
* Somehow, the header style in Fortran using `!---` and `!===` leads to failure of the autodoc for the api/lib, which is a lib.py file generated by f90wrap
* Solution is simple, but a bit of work: change ALL headers to the "compatible style", like in pygvec/run.F90.
* adapted header style in all Fortran files that are wrapped with f90wrap (list of the files is defined in `CMakeList.txt` file). Also adapted `template.F90`
