+++ src/profiles/c_rprofile.F90
@@ -62,7 +62,7 @@ MODULE MODgvec_rProfile_base
     INTEGER, INTENT(IN) :: D,deriv
     INTEGER :: i
     REAL(wp) :: prefactor
-    !===================================================================================================================================
+    ! CODE --------------------------------------------------------------------------------------------------------------------------!
     prefactor = 1.0_wp
     DO i=D-deriv+1,D
         prefactor = prefactor*i
@@ -74,15 +74,12 @@ MODULE MODgvec_rProfile_base
   !!
   !===================================================================================================================================
   PURE FUNCTION rho2_derivative(rho,deriv) RESULT(rho2_prime)
-  ! MODULES
-  !-----------------------------------------------------------------------------------------------------------------------------------
-  ! INPUT VARIABLES
+    ! INPUT VARIABLES -------------------------!
     REAL(wp), INTENT(IN) :: rho                !! rho position rho ~sqrt(magnetic flux)
     INTEGER, INTENT(IN)  :: deriv  !! derivative in rho
-  !-----------------------------------------------------------------------------------------------------------------------------------
-  ! OUTPUT VARIABLES
+    ! OUTPUT VARIABLES -------------------------!
     REAL(wp) :: rho2_prime     !!n-th derivative of rho2 with respect to rho ~sqrt(magnetic flux).
-  !===================================================================================================================================
+    ! CODE --------------------------------------------------------------------------------------------------------------------------!
     IF (deriv>2) THEN
       rho2_prime = 0.0_wp
   ELSE
@@ -95,18 +92,14 @@ MODULE MODgvec_rProfile_base
   !!
   !===================================================================================================================================
   FUNCTION rProfile_drho2(sf, rho) RESULT(derivative)
-  ! MODULES
-  !-----------------------------------------------------------------------------------------------------------------------------------
-  ! INPUT VARIABLES
+    ! INPUT VARIABLES -------------------------!
     CLASS(c_rProfile)  :: sf !! self
     REAL(wp), INTENT(IN) :: rho                !! rho position rho ~sqrt(magnetic flux)
-  !-----------------------------------------------------------------------------------------------------------------------------------
-  ! OUTPUT VARIABLES
+    ! OUTPUT VARIABLES -------------------------!
     REAL(wp) :: derivative   !! 2nd derivative of a radial profile with respect to rho ~sqrt(magnetic flux).
-  !-----------------------------------------------------------------------------------------------------------------------------------
-  ! LOCAL VARIABLES
+    ! LOCAL VARIABLES -------------------------!
     REAL(wp) :: rho2
-  !===================================================================================================================================
+    ! CODE --------------------------------------------------------------------------------------------------------------------------!
     rho2 = rho2_derivative(rho,deriv=0)
     ! d^2/dx^2 f(g(x)) = [g'(x)]^2 * f''(g(x))+g''(x) * f'(g(x))
     derivative =  rho2_derivative(rho,deriv=1)**2*sf%eval_at_rho2(rho2, deriv=2) &
@@ -118,18 +111,14 @@ MODULE MODgvec_rProfile_base
   !!
   !===================================================================================================================================
   FUNCTION rProfile_drho3(sf, rho) RESULT(derivative)
-  ! MODULES
-  !-----------------------------------------------------------------------------------------------------------------------------------
-  ! INPUT VARIABLES
+    ! INPUT VARIABLES -------------------------!
     CLASS(c_rProfile)  :: sf !! self
     REAL(wp), INTENT(IN) :: rho                !! rho position rho ~sqrt(magnetic flux)
-  !-----------------------------------------------------------------------------------------------------------------------------------
[…]
