+++ src/globals/newton.F90
@@ -160,21 +160,18 @@ CONTAINS
 FUNCTION NewtonMin1D(tol,a,b,maxstep,x,fobj) RESULT (fmin)
   ! MODULES
   IMPLICIT NONE
-  !---------------------------------------------------------------------------------------------------------------------------------
-  ! INPUT VARIABLES
+  ! INPUT VARIABLES -------------------------!
   REAL(wp),INTENT(IN)               :: tol  !! abort tolerance
   REAL(wp),INTENT(IN)               :: a,b  !! search interval
   REAL(wp),INTENT(IN)               :: maxstep  !! max|dx| allowed
   CLASS(c_newton_Min1D),INTENT(IN)  :: fobj !! functional to minimize with f(x), d/dx f(x), d^2/dx^2 f(x)
-  !---------------------------------------------------------------------------------------------------------------------------------
-  ! OUTPUT VARIABLES
+  ! OUTPUT VARIABLES -------------------------!
   REAL(wp),INTENT(INOUT) :: x    !! initial guess on input, result on output
   REAL(wp)               :: fmin !! on output =f(x)
-  !---------------------------------------------------------------------------------------------------------------------------------
-  ! LOCAL VARIABLES
+  ! LOCAL VARIABLES --------------------------!
   REAL(wp)                                      :: x0
   TYPE(t_newton_Root1D_wrap_Min1D), ALLOCATABLE :: fobj_wrap
-  !=================================================================================================================================
+  ! CODE --------------------------------------------------------------------------------------------------------------------------!
   fobj_wrap = t_newton_Root1D_wrap_Min1D(fobj)
   x0=x
   x=NewtonRoot1D(tol,a,b,maxstep,x0,0.0_wp,fobj_wrap)
@@ -190,12 +187,11 @@ END FUNCTION NewtonMin1D
 FUNCTION newton_Root1D_wrap_Min1D_new(parent) RESULT(sf)
   ! MODULES
   IMPLICIT NONE
-  !---------------------------------------------------------------------------------------------------------------------------------
-  ! INPUT VARIABLES
+  ! INPUT VARIABLES -------------------------!
   CLASS(c_newton_Min1D), INTENT(IN) :: parent
-  ! OUTPUT VARIABLES
+  ! OUTPUT VARIABLES ------------------------!
   TYPE(t_newton_Root1D_wrap_Min1D) :: sf
-  !=================================================================================================================================
+  ! CODE --------------------------------------------------------------------------------------------------------------------------!
   sf%parent = parent
 END FUNCTION
 
@@ -207,13 +203,12 @@ END FUNCTION
 FUNCTION newton_Root1D_wrap_Min1D_FR(sf, x) RESULT(y1x1)
   ! MODULES
   IMPLICIT NONE
-  !---------------------------------------------------------------------------------------------------------------------------------
-  ! INPUT VARIABLES
+  ! INPUT VARIABLES -------------------------!
   CLASS(t_newton_Root1D_wrap_Min1D), INTENT(IN) :: sf
   REAL(wp), INTENT(IN) :: x
-  ! OUTPUT VARIABLES
[…]
