### 2026-01-23 fa999a9
update docs and notebooks regarding links
* add the api autodoc for plotting and gframe python files.
* added plotly<6.0 requirement for the building of the documentation
* add/check links to other pages of the documentation and the api autodoc for commands in the notebook
* double links: one for the online documentation, linking directly within the documentation.
* the second link is shown after the first link with a (globesymbol), and links to the latest documentation on readthedocs.
  This is for those who run the notebook locally, and still want to follow the link.
