+++ docs/api/plotting-plots1d.rst
@@ -1,0 +1,9 @@
+.. _api-plotting-plots1d:
+
+gvec.plotting.plots1d
+=====================
+
+.. automodule:: gvec.plotting.plots1d
+  :members:
+  :undoc-members:
+  :show-inheritance:
