+++ docs/api/plotting-plots3d.rst
@@ -1,0 +1,9 @@
+.. _api-plotting-plots3d:
+
+gvec.plotting.plots3d
+=====================
+
+.. automodule:: gvec.plotting.plots3d
+  :members:
+  :undoc-members:
+  :show-inheritance:
