### 2026-01-23 f76feeb
added load_quasr function.
