### 2026-01-23 #9670 @fhindenlang
Hi @piyueh,
 Thanks for the heads up. However, back in November last year, in the [pre review](https://github.com/openjournals/joss-reviews/issues/8909#issuecomment-3553605678), you agreed to review version `1.3.0`,  and not `1.2.0`.

Only now I see that in the top information of this review,  the version is still `1.2.0`.
@piyueh I really hope it is still possible to make use `v1.3.0`.

@dpanici: you looked at the latest `main` branch of GVEC, which is on-par with tag `v1.3.0`.  And the `joss_paper_2025` is 5 commits ahead, which are only related to the actual paper, not to anything else in the code or documentation.
