### 2026-01-23 dc7622b
format python function docstring to rst
* started this for two routines in plotting
* for correct display in documentation
* we need to use ``abc`` for verbatim, and not use raw string for the docstring
* thus in math mode $abc$, special characters like backslash need to be escaped with an additional backslash
