### 2026-01-23 7f2b298
- Updated plotting docstrings to render correctly on the html. - Changed `full_surface` input in `plot_3d_surface` to `period`, the   input now looks the same as `pygvec to-paraview`
