+++ python/gvec/plotting/plots2d.py
@@ -26,53 +26,51 @@ def plot_poloidal_plane(
     sfl: Literal["pest"] | None = "pest",
     plot_kwargs: dict = dict(),
 ):
-    r"""
-    plot_poloidal_plane
-
-    Plot a poloidal plane with some equilibrium quantity on it. Defaults to plotting $\|\mathbf{B}\|$
+    """
+    Plot a poloidal plane with some equilibrium quantity on it. Defaults to plotting $|B|$
 
     Parameters
     ----------
     state : GVEC state object
     quantity : str, optional
-        The quantity to plot. Default is "mod_B". If `None`, no contours are plotted.
+        The quantity to plot. Default is ``"mod_B"``. If ``None``, no contours are plotted.
     nrho : int, optional
         The radial resolution of the slices.
-        Default is `51`
+        Default is ``51``
     ntheta : int, optional
         The poloidal resolution of the slices.
-        Default is `51`
+        Default is ``51``
     zeta : int, float, ndarray, optional
-        The number of equally spaced slices (int), the specific `zeta` value (float) or values (np.ndarray).
-        Default is `9`.
+        The number of equally spaced slices (``int``), the specific ``zeta`` value (``float``) or values (``np.ndarray``).
+        Default is ``9``.
     subplot_grid : list[int], None, optional
-        The grid shape for the subplots. If `None`, grid will be automatically determined.
-        Default is `None`.
+        The grid shape for the subplots. If ``None``, grid will be automatically determined.
+        Default is ``None``.
     share_axis : bool
-        If `True`, all subplots will share their `X1` and `X2` axis positions.
-        Default `False`
+        If ``True``, all subplots will share their ``X1`` and ``X2`` axis positions.
+        Default ``False``
     rho_contours : int, optional
-        The number of ``$\rho$`` contours to plot. If `0`, no contours are plotted.
-        Default `4`
+        The number of ``rho`` contours to plot. If ``0``, no contours are plotted.
+        Default ``4``
     rho_contours_color : str, optional
-        The color of the ``$\rho$`` contours.
-        Default `"white"`
+        The color of the ``rho`` contours.
+        Default ``"white"``
     theta_contours : int, optional
-        The number of ``$\theta$`` contours to plot. If `0`, no contours are plotted.
-        Default `8`
+        The number of ``theta`` contours to plot. If ``0``, no contours are plotted.
+        Default ``8``
     theta_contours_color : str, optional
-        The color of the ``$\theta$`` contours.
-        Default `"white"`
-    sfl : `"pest"` or `None`, optional
-        Plot the `theta` contours or `pest` contours (``$\theta^\star$``).
-        Default `"pest"`.
+        The color of the ``theta`` contours.
+        Default ``"white"``
+    sfl : ``"pest"`` or ``None``, optional
+        Plot the ``theta`` contours or ``pest`` contours ($\\theta^\\star$).
+        Default ``"pest"``.
     plot_kwargs: dict, optional
-        Any `**kwargs` to send to the `plt.figure()` function.
-        For example `plot_kwargs={'figsize': (8,8)}`. See the [matplotlib documentation](https://matplotlib.org/stable/api/_as_gen/matplotlib.pyplot.figure.html) for a list of kwargs.
+        Any ``**kwargs`` to send to the ``plt.figure()`` function.
+        For example ``plot_kwargs={'figsize': (8,8)}``. See the `matplotlib documentation <https://matplotlib.org/stable/api/_as_gen/matplotlib.pyplot.figure.html>`_ for a list of kwargs.
 
[…]
