### 2026-01-23 7069b9a
removed some unused variables
