### 2026-01-23 9b63e2f
extend 3d plots and plotting notebook
