+++ python/gvec/plotting/plots3d.py
@@ -9,11 +9,12 @@ from gvec.core.state import State
 def plot_3d_surface(
     state: State,
     quantity: str = "mod_B",
-    rho: float = 1.0,
+    rho: float | np.ndarray | list = 1.0,
     ntheta: int = 41,
     nzeta: int = 51,
     period: str = "single",
     to_file: str | None = None,
+    surface_kwargs: dict = dict(),
 ):
     """
     Generate a 3D surface plot with the quantity provided by ``quantity`` on it at a given ``rho`` position and
@@ -37,6 +38,8 @@ def plot_3d_surface(
     period : str
         Plot the ``"full"`` surface, a ``"single"`` field period or ``"half"`` a field period.
         Default ``"single"``
+    surface_kwargs : dict, optional
+        Keyword arguments for the surface plot. i.e. `dict(opacity=0.5)`
     to_file : str
         If a string, will automatically save the plot to a file with the given input in the current working directory. Recommended to use this if the plots don't display.
         Default is ``None``
@@ -62,20 +65,33 @@ def plot_3d_surface(
     zeta = np.linspace(0.0, zeta_end, nzeta)
     theta = np.linspace(0.0, 2 * np.pi, ntheta)
 
-    evaluation = state.evaluate(quantity, rho=[rho], theta=theta, zeta=zeta).sel(rho=rho)
-
-    fig = graph_objects.Figure()
-
-    fig.add_trace(
-        graph_objects.Surface(
-            x=evaluation.pos.sel(xyz="x"),
-            y=evaluation.pos.sel(xyz="y"),
-            z=evaluation.pos.sel(xyz="z"),
-            surfacecolor=evaluation[quantity],
-            colorbar_title_text=quantity,
-        )
+    evaluations = state.evaluate(
+        "pos", quantity, rho=np.asarray(rho).tolist(), theta=theta, zeta=zeta
     )
 
+    fig = graph_objects.Figure()
+    min_value, max_value = (
+        np.amin(evaluations[quantity].data),
+        np.amax(evaluations[quantity].data),
+    )
+
+    for irho, rhoval in enumerate(evaluations.rho.data):
+        evaluation = evaluations.isel(rad=irho)
+        fig.add_trace(
+            graph_objects.Surface(
+                x=evaluation.pos.sel(xyz="x"),
+                y=evaluation.pos.sel(xyz="y"),
+                z=evaluation.pos.sel(xyz="z"),
+                surfacecolor=evaluation[quantity].broadcast_like(evaluation.pos.sel(xyz="x")),
+                # colorbar_title_text=f"${evaluation[quantity].attrs['symbol']}$",
+                # colorbar_title_text=f"${evaluation[quantity].attrs.get('symbol', quantity)}$",
+                colorbar_title_text=quantity,
+                showscale=(irho == 0),
+                cmin=min_value,
+                cmax=max_value,
+                **surface_kwargs,
+            )
+        )
     # Ensure the figure correctly scales the axis to the data
     fig["layout"]["scene"]["aspectmode"] = "data"
 
