### 2026-01-23 #9670 @piyueh
> Hi [@piyueh](https://github.com/piyueh), Thanks for the heads up. However, back in November last year, in the [pre review](https://github.com/openjournals/joss-reviews/issues/8909#issuecomment-3553605678), you agreed to review version `1.3.0`, and not `1.2.0`.
>
> Only now I see that in the top information of this review, the version is still `1.2.0`. [@piyueh](https://github.com/piyueh) I really hope it is still possible to make use `v1.3.0`.
>
> [@dpanici](https://github.com/dpanici): you looked at the latest `main` branch of GVEC, which is on-par with tag `v1.3.0`. And the `joss_paper_2025` is 5 commits ahead, which are only related to the actual paper, not to anything else in the code or documentation.

@fhindenlang sorry I thought our bot would automatically change the version when we moved from the [pre-review](https://github.com/openjournals/joss-reviews/issues/8909) thread to the this thread. I have updated the version tag to `v1.3.0`.

@dpanici @AnkitBarik For the source code, please review `v1.3.0`. `v1.3.0` is the same snapshot as the current `main` branch (at commit [`3a93faa`](https://github.com/gvec-group/gvec/commit/3a93faa57df4190edb646b5380965c07d8f156d7)). So nothing changes if you have been reviewing the `main` branch. Thanks a lot!
