### 2026-01-26 f1dbfa3
overhaul docstrings for evaluate & compute
* overhaul docstring of core.compute.compute & core.state.compute
* overhaul docstring of core.compute.evaluate & evaluate_sfl
* fix docstring of core.compute.EvaluationsBoozer
* fix deprecation warning of core.compute.EvaluationsBoozerCustom
