+++ python/gvec/core/compute.py
@@ -157,23 +157,34 @@ def table_of_quantities(markdown: bool = False, registry: Mapping = QUANTITIES):
 def compute(
     ev: xr.Dataset,
     *quantities: str,
-    state: State = None,
+    state: State | None = None,
     registry: Mapping = QUANTITIES,
 ) -> xr.Dataset:
-    """Compute the target equilibrium quantity and add it to the given evaluation dataset.
+    """
+    Compute the target quantity and add it to the given evaluation dataset.
 
-    This method will compute required parameters recursively and add them to the dataset.
+    This method will recursively determine prerequisites, compute them and add them to the dataset as needed.
 
     Parameters
     ----------
     ev : xr.Dataset
-        The evaluation dataset with the target coordinates (rho, theta, zeta) and possibly some precomputed quantities.
-    quantities : str
+        The evaluation dataset with the target grid ``(rad, pol, tor)``, coordinates ``(rho, theta, zeta)`` and possibly some precomputed quantities.
+    *quantities : str
         One or more names of the quantities to compute. See `table_of_quantities` for a list of available quantities.
-    state : State, optional
-        A gvec.State object that is used to compute the quantities. Not necessary if the desired quantities only depend on already computed quantities.
-    registry : Mapping, optional
+
+        See the `default table of available quantities </tutorials/notebooks/050_pygvec.html#available-quantities-for-evaluation>`_
+        or call ``table_of_quantities`` to see all options.
+    state : gvec.State | None, optional
+        A gvec.State object that can be used by quantities which require it. Not necessary if the desired quantities only depend on already computed quantities.
+    registry : Mapping | None, optional
         The registry of computable quantites to use.
+        Default: the ``gvec.core.compute.QUANTITIES`` registry used to evaluate a gvec.State object.
+
+    See Also
+    --------
+    evaluate: create a new grid in logical coordinates and compute target quantities.
+    evaluate_sfl: create a new grid in straight-fieldline coordinates and compute target quantities.
+    gvec.core.state.State.evaluate: this function as a method of gvec.State.
     """
     for quantity in quantities:
         # --- get the compute function --- #
@@ -472,10 +483,16 @@ def EvaluationsBoozer(
     radial_derivative : bool
         Whether to compute the radial derivatives of the `LA` and `NU_B` variables, at fixed GVEC angles
         $(\\vartheta(\\rho_i,\\vartheta_{B,j},\\zeta_{B,k}),\\zeta(\\rho_i,\\vartheta_{B,j},\\zeta_{B,k}))$.
-        Computes boozer transform  at additional radial points `rho- epsilon`, and uses a first order Finite Difference in epsilon (`=1e-8`) for the derivatives.
+        Computes boozer transform at additional radial points $\\rho-\\epsilon$, and uses a first order Finite Difference for the derivatives.
+    epsilon_FD : float
+        The offset in rho used for the Finite Difference computation of the radial derivatives.
     boozer_kwargs : dict
         Additional keyword arguments to pass to the `get_boozer` method of the state object.
         These can be used to specify the Boozer transform parameters, such as the maximum mode numbers via 'MNfactor'.
+
+    Returns
+    -------
+    xr.Dataset
     """
     match rho:
         case str() if rho == "int":
@@ -775,7 +792,7 @@ def EvaluationsPEST(
 def EvaluationsBoozerCustom(rho, theta_B, zeta_B, state, **boozer_kwargs):
     """Create a custom EvaluationsBoozer dataset with Boozer coordinates.
 
-    DEPRECATED: use `EvaluationsBoozer` instead.
+    .. deprecated:: v1.2
     """
[…]
