### 2026-01-26 23dbe15
restructure State.evaluate & evaluate_sfl
State.evaluate and State.evaluate_sfl are defined (injected) in core.compute.
* keeps signature & docstrings consistent
* self replaces state which is correct
* also done with plotting functions in a similar way
* fixes defaults for evaluate_sfl
