+++ python/gvec/core/compute.py
@@ -1011,6 +1011,10 @@ def evaluate_sfl(
     return ev
 
 
+State.evaluate = evaluate
+State.evaluate_sfl = evaluate_sfl
+
+
 # === Fourier Transform === #
 
 
