### 2026-01-26 2900336
clean register, toq & compute signatures
remove explicit registry=QUANTITIES default from signature, replaced with None and delayed default to generate cleaner documentation
