### 2026-01-26 5c81fda
revise contact & contribution information
