+++ README.md
@@ -55,16 +55,16 @@ For required libraries, other installation methods and more details see the docu
 ## Reporting Bugs & Contributing to GVEC
 
 The GVEC development repository is hosted at the [MPCDF-GitLab](https://gitlab.mpcdf.mpg.de/gvec-group/gvec).
-There you can also find the [Issue tracker](https://gitlab.mpcdf.mpg.de/gvec-group/gvec/-/issues), merge requests and releases.
-Contact the maintainers for a guest account if you want to contribute to GVEC.
+There you can also find the [issue tracker](https://gitlab.mpcdf.mpg.de/gvec-group/gvec/-/issues), [merge requests](https://gitlab.mpcdf.mpg.de/gvec-group/gvec/-/merge_requests) and [release notes](https://gitlab.mpcdf.mpg.de/gvec-group/gvec/-/releases).
 
 A public mirror of GVEC is hosted on [GitHub](https://github.com/gvec-group/gvec).
-There you can also open [issues](https://github.com/gvec-group/gvec/issues) and pull requests.
+There everyone can open [issues](https://github.com/gvec-group/gvec/issues) and [pull requests](https://github.com/gvec-group/gvec/pulls).
+**Contributions are always welcome!**
 
-For questions, discussions and announcements you can also join the [matrix-chat channel](https://matrix.to/#/#gvec:mpg.de).
+For questions, discussions and announcements you can join the [matrix-chat channel](https://matrix.to/#/#gvec:mpg.de).
 
-Contributions are always welcome, best get into contact directly with the maintainers.
-Also see the relevant [documentation](https://gvec.readthedocs.io/latest/dev/index.html).
+More information on the structure of GVEC and some development guidlines can be found in the [developer documentation](https://gvec.readthedocs.io/latest/dev/index.html).
+If you don't know where to start, don't hesitate to contact [the maintainers](https://www.ipp.mpg.de/person/139885/5497858) directly.
 
 ## License
 
