### 2026-01-27 89aade4
Mini-fix to docstring
