### 2026-01-27 5fa81c0
documentation: fix sign in user/theory
