### 2026-01-27 bc8fedb
revise contributing information
