### 2026-01-27 6893099
docs: add page for table of quantities
* adds dedicated documentation page for the table of quantities
* links to that page in the docstrings
