### 2026-01-27 a245ff8
fix docstrings
