### 2026-01-27 gvec-group/gvec#3 @Rykath
Yes, you are right, we had neglected the API documentation so far. I've now updated the docstrings for `compute`, `evaluate` and `evaluate_sfl`.

The proposed changes are tracked in [GitLab!131](https://gitlab.mpcdf.mpg.de/gvec-group/gvec/-/merge_requests/131). You can check out a preview of the documentation at https://gvec.readthedocs.io/fix-gh3-gh5

* updated [`compute`](https://gvec.readthedocs.io/fix-gh3-gh5/api/core-compute.html#gvec.core.compute.compute) and [`State.compute`](https://gvec.readthedocs.io/fix-gh3-gh5/api/core-state.html#gvec.core.state.State.compute)
* updated [`evaluate`](https://gvec.readthedocs.io/fix-gh3-gh5/api/core-compute.html#gvec.core.compute.evaluate), identical to [`State.evaluate`](https://gvec.readthedocs.io/fix-gh3-gh5/api/core-state.html#gvec.core.state.State.evaluate)
* updated [`evaluate_sfl`](https://gvec.readthedocs.io/fix-gh3-gh5/api/core-compute.html#gvec.core.compute.evaluate_sfl), identical to [`State.evaluate_sfl`](https://gvec.readthedocs.io/fix-gh3-gh5/api/core-state.html#gvec.core.state.State.evaluate_sfl)

Let me know if this is detailed enough.
