### 2026-01-27 gvec-group/gvec#5 @Rykath
I think that section might have been a bit misleading. We don't require anyone to make a guest account to contribute to GVEC. Rather we also welcome any contributions on GitHub!

For new contributors the easiest would be to open an issue or pull request on GitHub and then we'll keep the conversation here and mirror the code back. A guest account for the MPCDF GitLab is really only necessary for large contributions where we might want to run longer testing pipelines.

I've now updated the relevant section in the [`README.md`](https://gitlab.mpcdf.mpg.de/gvec-group/gvec/-/blob/fix-gh3-gh5/README.md#reporting-bugs--contributing-to-gvec) and [documentation](https://gvec.readthedocs.io/fix-gh3-gh5/dev/index.html) (tracked in [GitLab!131](https://gitlab.mpcdf.mpg.de/gvec-group/gvec/-/merge_requests/131)).

Is this now better?
