### 2026-01-27 d4d8d7c
add Coil.from_simsopt
