### 2026-01-27 gvec-group/gvec#5 @dpanici
Ah ok, apologies for my misunderstanding, this is clearer now!
