### 2026-01-27 gvec-group/gvec#2 @fhindenlang
Hi Dario,
If you look into the lastest documentation on [readthedocs](https://gvec.readthedocs.io/latest/tutorials/notebooks/020_stellarator.html), we changed the notebooks correspondingly. An additional link is added with a symbol (&#x1F310;) which then points to the readthedocs  page of the latest version.
* the reason to have two links: as we have multiple versions documented, and we display the notebooks after execution in the tutorial section of the documentation, we still need the links to follow the internal structure of each versioned documentation
* however if you download a run the notebook locally, to should still be able follow the links by clicking on  ([&#x1F310;](https://gvec.readthedocs.io/latest/user/gvec-parameter-list.html))

What do you think?
