### 2026-01-27 gvec-group/gvec#2 @dpanici
Looks good!
