### 2026-01-28 70efbc7
Removed intermediate plotting
