### 2026-01-28 5ca2c30
added simsopt to notebook requirements
