### 2026-01-28 4266df4
add Mapping ABC for CoilSet
