### 2026-01-29 4d22b74
introduced a vars type in minimizer
* `%vars` has two possible types, one is gradient descent and the other accelerated gradient descent
* this way, the `IF(MinimizerType.EQ.10)` becomes a `SELECT TYPE(vars)`
* using `ASSOCIATE(vars=>sf%minimizer%vars)` to shorten the variable names
