+++ src/functionals/mhd3d/mhd3d_minimize.F90
@@ -23,14 +23,14 @@ MODULE MODgvec_MHD3D_minimize
     !-----------------------------------------------------------------------------------------------------------------------------------
     ! TYPES
     !-----------------------------------------------------------------------------------------------------------------------------------
-    TYPE :: t_minimizer_mhd3d
-        !-------------------------------------------------------------------------------------------------------------------------------
-        LOGICAL   :: initialized, restart_iter, logger_is_initialized
-        INTEGER   :: JacCheck, MinType
+    TYPE, ABSTRACT :: a_minimizer_vars
+        CHARACTER(LEN=40) :: MinimizerType
+        LOGICAL   :: restart_iter, logger_is_initialized
+        INTEGER   :: JacCheck
         INTEGER   :: iter,nStepDecreased,nSkip_Jac,nSkip_dw
         INTEGER   :: lastoutputIter, logiter_ramp, logscreen
         REAL(wp)  :: dt,deltaW, dW_allowed
-        REAL(wp)  :: t_pseudo,Fnorm(3),Vnorm(3),Fnorm0(3),Fnorm_old(3),W_MHD3D_0
+        REAL(wp)  :: t_pseudo,Fnorm(3),Fnorm0(3),Fnorm_old(3),W_MHD3D_0
 
         ! Logging variables
         REAL(wp) :: min_dt_out,max_dt_out,min_dw_out,max_dw_out,sum_dW_out
@@ -40,12 +40,23 @@ MODULE MODgvec_MHD3D_minimize
         TYPE(t_sol_var_MHD3D), ALLOCATABLE :: dofs(:)      !! degrees of freedom at levels (k-1),(k),(k+1)
         TYPE(t_sol_var_MHD3D), ALLOCATABLE :: force(:)     !! force
         TYPE(t_sol_var_MHD3D), ALLOCATABLE :: temp_dofs(:) !! temporary for update
+    END TYPE a_minimizer_vars
+    TYPE, EXTENDS(a_minimizer_vars) :: t_gradient_descent_vars
+    END TYPE t_gradient_descent_vars
 
-        ! acc. gradient descent variables
+    TYPE, EXTENDS(a_minimizer_vars) :: t_accelerated_gradient_descent_vars
+        REAL(wp) :: Vnorm(3)
         REAL(wp), ALLOCATABLE  :: tau(:)
         INTEGER   :: ndamp
         REAL(wp)  :: tau_bar
         TYPE(t_sol_var_MHD3D), ALLOCATABLE :: velocity(:)       !! 'velocity' in minimizer
+    END TYPE t_accelerated_gradient_descent_vars
+
+    TYPE :: t_minimizer_mhd3d
+        !-------------------------------------------------------------------------------------------------------------------------------
+        LOGICAL   :: initialized
+        INTEGER   :: MinType
+        CLASS(a_minimizer_vars), ALLOCATABLE :: vars  !!! depend on the MinimizerType
         !-------------------------------------------------------------------------------------------------------------------------------
         CONTAINS
         PROCEDURE :: minimize         => MinimizeMHD3D_descent
@@ -82,78 +93,108 @@ MODULE MODgvec_MHD3D_minimize
         ALLOCATE(sf)
 
         sf%MinType = MinType_in
-        sf%dt = dt_initial
[…]
