### 2026-01-29 818a29b
small bugfixes found by compiler
