+++ src/pygvec/state.F90
@@ -8,7 +8,7 @@ MODULE MODgvec_py_state
 
 USE MODgvec_MHD3D,        ONLY: t_functional_mhd3d
 USE MODgvec_base,         ONLY: t_base
-USE MODgvec_Globals,      ONLY: enter_subregion,exit_subregion,reset_subregion
+USE MODgvec_Globals,      ONLY: enter_subregion,exit_subregion,reset_subregion,Unit_stdOut,MPIroot,nRanks,abort,wp
 
 IMPLICIT NONE
 PUBLIC
@@ -22,7 +22,7 @@ CONTAINS
 !================================================================================================================================!
 SUBROUTINE Init(parameterfile)
   ! MODULES
-  USE MODgvec_Globals,        ONLY: Unit_stdOut,MPIroot,nRanks,abort,fmt_sep
+  USE MODgvec_Globals,        ONLY: fmt_sep
   USE MODgvec_MHD3D_vars,     ONLY: X1_base
   USE MODgvec_Analyze,        ONLY: InitAnalyze
   USE MODgvec_Output,         ONLY: InitOutput
@@ -99,7 +99,7 @@ END SUBROUTINE ReadState
 !================================================================================================================================!
 SUBROUTINE minimize(dt_in)
   ! MODULES
-  USE MODgvec_Globals,    ONLY: wp, Unit_stdOut, GetTime
+  USE MODgvec_Globals,    ONLY: GetTime
   USE MODgvec_MHD3D_Vars, ONLY: maxIter
   REAL(wp), OPTIONAL, INTENT(IN) :: dt_in
   ! LOCAL VARIABLES -------------------------------------------------------------------------------------------------------------!
@@ -118,7 +118,6 @@ END SUBROUTINE minimize
 !================================================================================================================================!
 SUBROUTINE select_base_dofs(var, base, dofs)
   ! MODULES
-  USE MODgvec_Globals,        ONLY: abort
   USE MODgvec_MHD3D_vars,     ONLY: X1_base,X2_base,LA_base
   USE MODgvec_base,           ONLY: t_base
   ! INPUT/OUTPUT VARIABLES ------------------------------------------------------------------------------------------------------!
@@ -147,7 +146,6 @@ END SUBROUTINE select_base_dofs
 !================================================================================================================================!
 SUBROUTINE select_base(var, base)
   ! MODULES
-  USE MODgvec_Globals,        ONLY: abort
   USE MODgvec_MHD3D_vars,     ONLY: X1_base,X2_base,LA_base
   USE MODgvec_base,           ONLY: t_base
   ! INPUT/OUTPUT VARIABLES ------------------------------------------------------------------------------------------------------!
@@ -209,7 +207,6 @@ END SUBROUTINE evaluate_base_select
 !================================================================================================================================!
 SUBROUTINE get_integration_points_num(var, n_s, n_t, n_z)
   ! MODULES
-  USE MODgvec_Globals,        ONLY: abort
   USE MODgvec_base,           ONLY: t_base
   ! INPUT/OUTPUT VARIABLES ------------------------------------------------------------------------------------------------------!
   CHARACTER(LEN=2), INTENT(IN) :: var           !! selection string: which variable to evaluate
@@ -521,7 +518,6 @@ END SUBROUTINE evaluate_base_tens_all
 !================================================================================================================================!
[…]
