### 2026-01-29 gvec-group/gvec#5 @Rykath
Merged to `develop` and `v1.3.x`.
