### 2026-01-29 #9670 @Rykath
We have now added the first set of (documentation) changes requested by @dpanici to the `v1.3.x` branch. In particular gvec-group/gvec#3 and gvec-group/gvec#5.

In addition that branch also contains a bugfix made necessary by the `pandas` release of `3.0.0` and some CI updates made necessary by changes in the infrastructure. The functionality remains identical to the reviewed `v1.3.0` version. The updated documentation can be accessed at https://gvec.readthedocs.io/v1.3.x
