### 2026-01-29 4a5352c
added docstrings to the new minimizer routines
