+++ src/functionals/mhd3d/mhd3d_minimize.F90
@@ -8,7 +8,7 @@
 !>
 !!# Module **MHD3D minimize**
 !!
-!! CONTAINS INITIALIZATION minimizer for MHD3D functional
+!! CONTAINS minimizer for MHD3D functional
 !!
 !===================================================================================================================================
 MODULE MODgvec_MHD3D_minimize
@@ -24,9 +24,9 @@ MODULE MODgvec_MHD3D_minimize
     ! TYPES
     !-----------------------------------------------------------------------------------------------------------------------------------
     TYPE, ABSTRACT :: a_minimizer_vars
-        CHARACTER(LEN=40) :: MinimizerType
+        CHARACTER(LEN=40) :: MinimizerType !! defines the minimization algorithm: 0 = Gradient-Descent, 10 = Accelerated Gradient-Descent
         LOGICAL   :: restart_iter, logger_is_initialized
-        INTEGER   :: JacCheck
+        INTEGER   :: JacCheck !! switch for restarts, if detJ<0 JacCheck=-1
         INTEGER   :: iter,nStepDecreased,nSkip_Jac,nSkip_dw
         INTEGER   :: lastoutputIter, logiter_ramp, logscreen
         REAL(wp)  :: dt,deltaW, dW_allowed
@@ -41,21 +41,34 @@ MODULE MODgvec_MHD3D_minimize
         TYPE(t_sol_var_MHD3D), ALLOCATABLE :: force(:)     !! force
         TYPE(t_sol_var_MHD3D), ALLOCATABLE :: temp_dofs(:) !! temporary for update
     END TYPE a_minimizer_vars
+
+    !===================================================================================================================================
+    !> State variables for a Gradient descent minimizer
+    !!
+    !===================================================================================================================================
     TYPE, EXTENDS(a_minimizer_vars) :: t_gradient_descent_vars
     END TYPE t_gradient_descent_vars
 
+    !===================================================================================================================================
+    !> State variables for an Accelerated Gradient-Descent descent minimizer
+    !!
+    !===================================================================================================================================
     TYPE, EXTENDS(a_minimizer_vars) :: t_accelerated_gradient_descent_vars
         REAL(wp) :: Vnorm(3)
         REAL(wp), ALLOCATABLE  :: tau(:)
         INTEGER   :: ndamp
         REAL(wp)  :: tau_bar
-        TYPE(t_sol_var_MHD3D), ALLOCATABLE :: velocity(:)       !! 'velocity' in minimizer
+        TYPE(t_sol_var_MHD3D), ALLOCATABLE :: velocity(:) !! 'velocity' in minimizer
     END TYPE t_accelerated_gradient_descent_vars
 
+    !===================================================================================================================================
+    !> Minimizer object
+    !!
+    !===================================================================================================================================
     TYPE :: t_minimizer_mhd3d
         !-------------------------------------------------------------------------------------------------------------------------------
-        LOGICAL   :: initialized
-        INTEGER   :: MinType
+        LOGICAL   :: initialized !! whether the object is initialized or ot
+        INTEGER   :: MinType !! gradient descent = 0, accelerated gradient descent = 10
         CLASS(a_minimizer_vars), ALLOCATABLE :: vars  !!! depend on the MinimizerType
         !-------------------------------------------------------------------------------------------------------------------------------
         CONTAINS
@@ -66,29 +79,40 @@ MODULE MODgvec_MHD3D_minimize
         PROCEDURE :: free             => Free_minimizer
     END TYPE t_minimizer_mhd3d
 
+    !===================================================================================================================================
 
     CONTAINS
 
+
[…]
