### 2026-01-29 #9670 @fhindenlang
Thanks @piyueh for updating the version number to  `v1.3.0`.

 I am currently working on the tutoridal for the G-Frame, requested in the issue [gvec-group/gvec#6](https://github.com/gvec-group/gvec/issues/6). I will keep you posted on the progress.
