### 2026-01-30 26c04da
Adapt finalization of functional in 'gvec_post.F90'
Co-authored-by:Florian Hindenlang
