### 2026-01-30 gvec-group/gvec#2 @fhindenlang
I cherry-picked the new link style to the `v.1.3.x` branch, they are visible on the v1.3 readthedocs page, i.e. [here](https://gvec.readthedocs.io/v1.3.x/tutorials/notebooks/010_tokamak.html)
