### 2026-02-02 d729696
gframe tutorial, docstrings extended.
* added/extended docstrings of gframe and quasr functions.
* bugfix in to_RZ:
  * phi=-zeta_out for cutting planes.
  * Rotation of zeta direction of the surface (clock-wise or ccw, seen from above) must then set the sign of zeta_start for searching the cross-section.
* increase length of hmap netcdf file name, mostly for CI
