+++ python/gvec/gframe.py
@@ -306,14 +306,42 @@ def get_xyz_cut(zeta_start, origins, normals, xyz_in, dft_dict, nfp):
 
 def cut_surf(xyz, nfp, xyz0_in, N_in, B_in):
     """
-    given xyz(zeta,theta) on the full torus, find intersection point of lines of theta=const with all N-B planes with origin xyz0. then project these points to find x1,x2 coordinates in each N-B cross-section
+    Given a surface $\bm x(\zeta,\vartheta)$ on the full torus, find intersection point of lines of $\vartheta=$const with all cutting planes defined by an origin curve ``xyz0`` and two base vectors ``N_in`` and ``B_in`` (normal of the cutting plane is then $N \times B$). The intersection points are projected on the base vectors, yielding the $X^1,X^2$ coordinates in all cutting planes. Cutting planes must be given on the full torus and must be a multiple of the number of field periods.
+
+    The input surface is assumed to have the same number of field periods, and must be given on the full torus. It is described by an array of cartesian point positions $\bm x(\zeta_i,\vartheta_j)$, assumed to be evaluated at a meshgrid of a toroidal and poloidal parameterization ($\zeta,\vartheta$), excluding the periodic endpoints:
+    $\zeta_i=2\pi \frac{i}{n_\zeta n_{FP}},i=0\dots,(n_\zeta n_{FP})-1,\quad \vartheta_j=2\pi\frac{j}{n_\theta},j=0,\dots,n_\theta-1$
+    The surface can thus be evaluated using a Fourier transform.
+    Note: If the number of cutting planes does not match the number of toroidal points of the surface, the cutting planes are resampled to match the number of toroidal points, before cutting.
+
+    Parameters:
+    ----------
+        xyz :  cartesian coordinates of the surface on the full torus, excluding the periodic endpoints: shape is ``(nzeta,ntheta,3)``
+        nfp : number of field periods
+        xyz0_in : origin of the cutting plane, shape is ``(nzeta_cut*nfp,3)``.
+        N_in : first base vector to span the cutting plane, shape is ``(nzeta_cut*nfp,3)``
+        B_in : second base vector to span the cutting plane, shape is ``(nzeta_cut*nfp,3)``.
+
+    Returns:
+    --------
+        x1_cut,x2_cut:
+            coordinates of the intersection points for cutting planes in one field period.
+            shape is ``(nzeta_surface,ntheta)``.
+
     """
+    assert xyz.shape[2] == 3, "xyz must have shape (nzeta,ntheta,3)"
     nz = xyz.shape[0]
+    assert np.mod(nz, nfp) == 0, (
+        "number of surface points in zeta direction must be divisible by nfp!"
+    )
+
     nz_gframe = xyz0_in.shape[0]
     if not xyz0_in.shape[0] == N_in.shape[0] == B_in.shape[0]:
         raise ValueError(
             "xyz0,N,B must have the same number of points, but they have different lengths!"
         )
+    assert xyz0_in.shape[1] == 3, "second dimension of xyz0 must be 3"
+    assert N_in.shape[1] == 3, "second dimension of N must be 3"
+    assert B_in.shape[1] == 3, "second dimension of B must be 3"
     # cut geometry with new frame (xyz0,N,B)
     zeta1d = np.linspace(0.0, 2 * np.pi, nz, endpoint=False)
     if nz != nz_gframe:
[…]
