### 2026-02-03 5ce886d
Fixed some bugs in Gframe and updated some docstrings (needs a bit more work).
