+++ python/gvec/gframe.py
@@ -5,16 +5,16 @@ r"""pyGVEC G-frame
 This module provides functions for computing the G-Frame from a surface, cutting a surface at a given plane, change from xyz to xyz_hat coordinates and read and write the G-Frame from/to a file.
 """
 
+import logging
 from pathlib import Path
 from typing import Literal
-import logging
 
 import numpy as np
 import xarray as xr
 from scipy.optimize import root_scalar
 
-from gvec.util import logging_setup, write_parameters
 from gvec import fourier
+from gvec.util import logging_setup, write_parameters
 
 
 def check_field_periodicity(xyz: np.ndarray, nfp: int, atol=1e-12):
@@ -974,12 +974,15 @@ def to_RZ(
         Dictionary with:
             zeta : zeta positions on one field period
             theta : theta positions
-            R : R positions on one field period, with shape [0:ntheta_out,0:nzeta_out]
-            Z : Z positions on one field period, with shape [0:ntheta_out,0:nzeta_out]
+            R : R positions on one field period, with shape ``[ntheta_out,nzeta_out]``
+            Z : Z positions on one field period, with shape ``[ntheta_out,nzeta_out]``
             nfp : number of field periods
             lasym : logical for asymmetry, =false if stellarator symmetry is found
             Mmax,Nmax : maximum mode numbers needed for the given tolerance
-            Rc,Rs,Zc,Zs : R and Z cosine and sine Fourier mode coefficients, respecting Mmax,Nmax
+            Rc,Rs,Zc,Zs : R and Z cosine and sine Fourier mode coefficients, shape is ``[Mmax+1,2*Nmax+1]``
+            m_modes: poloidal mode numbers (m) in first dimension of ``Rc, Rs, ...``
+            n_modes: toroidal mode numbers (n) in second dimension of ``Rc, Rs, ...``
+            tolerance: input ``tolerance``
     """
     assert xyz.shape[2] == 3, "xyz must have shape [nzeta*nfp, ntheta, 3]"
     nzetafull_in, ntheta_in = xyz.shape[0], xyz.shape[1]
@@ -1001,11 +1004,18 @@ def to_RZ(
     Ncirc = np.zeros((nzeta, 3))
     Bcirc = np.zeros((nzeta, 3))
     Ncirc[:, 0] = np.cos(zeta_out)
-    Ncirc[:, 1] = np.sin(zeta_out)
+    Ncirc[:, 1] = np.sin(-zeta_out)
     Bcirc[:, 2] = 1.0
+    # try to find the rotation direction of zeta from zeta[0] and zeta[1]:
+    # midpoint of zeta[0]
+    xyz_mid = np.mean(xyz[0, :, :], axis=0)
+    e_zeta = np.mean(xyz[1, :, :], axis=0) - xyz_mid
+    sign_zeta = np.sign(
+        np.cross(xyz_mid, e_zeta)[2]
+    )  # +1 if zeta is counterclockwise from above
 
     xyz_RZcut = get_xyz_cut(
-        zeta_out, origins, np.cross(Ncirc, Bcirc, axis=-1), xyz_t, zdft, nfp
+        (-sign_zeta) * zeta_out, origins, np.cross(Ncirc, Bcirc, axis=-1), xyz_t, zdft, nfp
     )
     R = np.sum((xyz_RZcut - origins[:, None, :]) * Ncirc[:, None, :], axis=-1).T
     Z = np.sum((xyz_RZcut - origins[:, None, :]) * Bcirc[:, None, :], axis=-1).T
[…]
