### 2026-02-03 aee5ea1
add fft2d_to_parameters
