+++ docs/user/index.md
@@ -21,6 +21,7 @@ quantities
 /tutorials/notebooks/030_stages
 /tutorials/notebooks/040_current
 /tutorials/notebooks/050_pygvec
+/tutorials/notebooks/022_gframe
 ```
 
 ```{toctree}
