### 2026-02-05 71eb3f5
fix incorrect mpol, ntor in vmec conversion
1) setting maximum M,N from the maximum m,n of the boundary is not desired, rather the VMEC parameters should be preserved
2) N was set to double the value it should be

only set M,N if not given in the input namelist
