### 2026-02-05 abb4e9d
add gframe to api documentation. Update gframe notebook.
* developer documentation has now also the api of `gframe.py` included.
* Compiled and checked documentation, corrected docstrings in `gframe.py`
* cherry pick from develop: path handling in parameter dictionary, in `gvec/core/run.py`
* rename gframe tutorial notebook to `052_gframe.ipynb`
* Add description of surface data in gframe notebook, and addressed changes raised in merge request
