+++ python/gvec/gframe.py
@@ -305,15 +305,18 @@ def get_xyz_cut(zeta_start, origins, normals, xyz_in, dft_dict, nfp):
 
 
 def cut_surf(xyz, nfp, xyz0_in, N_in, B_in):
-    """
-    Given a surface $\bm x(\zeta,\vartheta)$ on the full torus, find intersection point of lines of $\vartheta=$const with all cutting planes defined by an origin curve ``xyz0`` and two base vectors ``N_in`` and ``B_in`` (normal of the cutting plane is then $N \times B$). The intersection points are projected on the base vectors, yielding the $X^1,X^2$ coordinates in all cutting planes. Cutting planes must be given on the full torus and must be a multiple of the number of field periods.
+    r"""
+    Given a surface :math:`\mathbf{x}(\zeta,\theta)` on the full torus, find intersection point of lines of :math:`\theta=` const with all cutting planes defined by an origin curve ``xyz0`` and two base vectors ``N_in`` and ``B_in`` (normal of the cutting plane is then :math:`N \times B`). The intersection points are projected on the base vectors, yielding the :math:`X^1,X^2` coordinates in all cutting planes. Cutting planes must be given on the full torus and must be a multiple of the number of field periods.
 
-    The input surface is assumed to have the same number of field periods, and must be given on the full torus. It is described by an array of cartesian point positions $\bm x(\zeta_i,\vartheta_j)$, assumed to be evaluated at a meshgrid of a toroidal and poloidal parameterization ($\zeta,\vartheta$), excluding the periodic endpoints:
-    $\zeta_i=2\pi \frac{i}{n_\zeta n_{FP}},i=0\dots,(n_\zeta n_{FP})-1,\quad \vartheta_j=2\pi\frac{j}{n_\theta},j=0,\dots,n_\theta-1$
+    The input surface is assumed to have the same number of field periods, and must be given on the full torus. It is described by an array of cartesian point positions :math:`\mathbf{x}(\zeta_i,\vartheta_j)`, assumed to be evaluated at a meshgrid of a toroidal and poloidal parameterization :math:`(\zeta,\theta)`, excluding the periodic endpoints:
+
+    :math:`\zeta_i=2\pi \frac{i}{n_\zeta n_{FP}},i=0\dots,(n_\zeta n_{FP})-1,\quad \theta_j=2\pi\frac{j}{n_\theta},j=0,\dots,n_\theta-1`
+
     The surface can thus be evaluated using a Fourier transform.
+
     Note: If the number of cutting planes does not match the number of toroidal points of the surface, the cutting planes are resampled to match the number of toroidal points, before cutting.
 
-    Parameters:
+    Parameters
     ----------
         xyz :  cartesian coordinates of the surface on the full torus, excluding the periodic endpoints: shape is ``(nzeta,ntheta,3)``
         nfp : number of field periods
@@ -321,12 +324,11 @@ def cut_surf(xyz, nfp, xyz0_in, N_in, B_in):
         N_in : first base vector to span the cutting plane, shape is ``(nzeta_cut*nfp,3)``
         B_in : second base vector to span the cutting plane, shape is ``(nzeta_cut*nfp,3)``.
 
-    Returns:
-    --------
+    Returns
+    -------
         x1_cut,x2_cut:
[…]
