+++ docs/tutorials/getting-started.md
@@ -44,6 +44,7 @@ The available tutorials are:
    * [](/tutorials/notebooks/030_stages)
    * [](/tutorials/notebooks/040_current)
    * [](/tutorials/notebooks/050_pygvec)
+   * [](/tutorials/notebooks/052_gframe)
 
 ### Running Jupyter notebooks
 
