### 2026-02-05 gvec-group/gvec#6 @fhindenlang
Dear @dpanici ,
as you stated correctly in your review that the G-Frame is a unique feature of GVEC, a tutorial was definitely missing!

I have now a first version of a tutorial about the G-Frame.
I pushed it to the branch `v1.3.x`, so it should run with the gvec which is pip installed from that branch.
There were also some code fixes necessary, without adding new functionality to the version v1.3.

The tutorial should be visible on the documentation page of the version
[v1.3 on readthedocs](https://gvec.readthedocs.io/v1.3.x/tutorials/notebooks/052_gframe.html).

In the tutorial, I tried to explain the idea, the requirements, and different construction methods, starting from very simple to more complex. Every G-Frame that is constructed is then also run in GVEC, and visualized.

I did not add any dependency on `pyQSC` / `pyQIC`, but since they can produce a boundary surface using their near-axis-expansion, this should be enough to construct the G-Frame from that surface. The construction from a surface is described at the end of the tutorial.

Looking forward to your comments!
