### 2026-02-09 3c6d83d
add boundary_coefficients option to load-quasr
* write the boundary data as fourier coefficients in the parameter file instead of points in the netCDF file
* still writes the boundary points into the netCDF file
* respects maximum modenumbers found by `--tol` option
