### 2026-02-10 de29845
- Updated rational iota plot so that negative rationals can be plotted - Switched contour plotting to use `contourf` instead of `tricontourf`.   This fixes an issue with concave shapes having filled in triangles   that weren't really there - Fixed some xarray things in poloidal plane plotting and added an error   message when trying to plot vector fields (which cannot be plotted) - Added a `dpi` argument to `pygvec visu` so resolution can be changed,   set default to 600.
