### 2026-02-10 d96aaf7
- Added value error if `quantities_eval` does not exist - Added some defaults for the boozer transform to reduce disk space when   plots are being generated
