### 2026-02-11 eba1629
Bug fix: - Was possible for an error to be thrown if `quantity` was `None` in   `plot_poloidal_plane`
