### 2026-02-12 gvec-group/gvec#6 @fhindenlang
@dpanici , were you able to look into the new tutorial?
