### 2026-02-12 gvec-group/gvec#6 @dpanici
Apologies! I was busy this week (submitting my thesis!)

I've just taken a look today, it looks really good, thank you for putting it together! I'm happy to resolve the issue for JOSS (and excited to try out the G-frame as well)
