### 2026-02-13 #9670 @fhindenlang
FYI, @AnkitBarik , the review is for version 1.3, so all requested code changes and notebooks have been put into the branch  `v1.3.x`, and the corresponding documentation page is [https://gvec.readthedocs.io/v1.3.x/index.html](https://gvec.readthedocs.io/v1.3.x/index.html)
If you want to run the notebooks locally, you would download them from the [getting started page](https://gvec.readthedocs.io/v1.3.x/tutorials/getting-started.html), and you would need to install gvec locally from the `v1.3.x` branch.
