### 2026-02-16 45419e5
reformat quasr docstrings
