+++ python/gvec/scripts/quasr.py
@@ -2,47 +2,60 @@
 # License: MIT
 r"""The pyGVEC script for converting a QUASR configuration to a G-Frame for use with GVEC.
 
-QUASR is the A QUAsi-symmetric Stellarator Repository: https://quasr.flatironinstitute.org/
+QUASR is the QUAsi-symmetric Stellarator Repository: https://quasr.flatironinstitute.org/.
+The algorithm used to construct a G-Frame is described in [1]_ and is as follows:
 
-The algorithm is described in the paper Hindenlang et al. DOI: 10.1088/1361-6587/adba11 and is as follows:
+1) Evaluate surface in cartesian space
 
-### STEP 1: Evaluate surface in cartesian space:
-
-1. using a json file from QUASR (from `https://quasr.flatironinstitute.org/`) with the simsopt interface
-2. Evaluate the surface cartesian position $(x,y,z)(\vartheta_i,\zeta_j)$ at a meshgrid  on the full torus:
+   First use the simsopt interface to load a json file from QUASR, which can also be retrieved automatically with a given ID.
+   Then evaluate the surface cartesian position $(x,y,z)(\vartheta_i,\zeta_j)$ at a meshgrid  on the full torus:
 
    $\vartheta_i=2\pi \frac{i}{n_t},i=0\dots,n_t-1,\quad \zeta_j=2\pi\frac{j}{n_z},j=0,\dots,n_z-1$
 
    where the angles $\vartheta,\zeta$ are just the parametrization of the given surface. (For the quasr surfaces, its a boozer angle parameterization!).
-
    The number $n_z$ is chosen as a multiple of the number of field periods $n_{FP}$, to be able to reduce the discrete dataset exactly to one field period.
 
-### STEP 2: Project to surface with elliptical cross-sections
+2) Project to surface with elliptical cross-sections
 
-Project ${\bm x}_m(\zeta)=\frac{1}{2\pi}\int_{\vartheta=0}^{2\pi}{\bm x}(\vartheta,\zeta)\sigma_m(\vartheta)d\vartheta$ with $\sigma_0(\vartheta)=1,\sigma_{s}(\vartheta)=2\sin(\vartheta),\sigma_{c}(\vartheta)=2\cos(\vartheta)$, leading to a surface
+   Project ${\bm x}_m(\zeta)=\frac{1}{2\pi}\int_{\vartheta=0}^{2\pi}{\bm x}(\vartheta,\zeta)\sigma_m(\vartheta)d\vartheta$
+   with $\sigma_0(\vartheta)=1,\sigma_{s}(\vartheta)=2\sin(\vartheta),\sigma_{c}(\vartheta)=2\cos(\vartheta)$, leading to a surface
 
-${\bm x}_m(\vartheta,\zeta)={\bm x}_0(\zeta) + {\bm x}_s(\zeta)\sin(\vartheta)+ {\bm x}_c(\zeta)\cos(\vartheta)$
+   ${\bm x}_m(\vartheta,\zeta)={\bm x}_0(\zeta) + {\bm x}_s(\zeta)\sin(\vartheta)+ {\bm x}_c(\zeta)\cos(\vartheta)$
 
-where cross-sections of $\zeta=\text{const.}$ are planar elliptical curves.
+   where cross-sections of $\zeta=\text{const.}$ are planar elliptical curves.
 
+3) Compute the plane of the ellipse cross-sections
 
-### STEP 3: Compute the plane of the ellipse cross-sections
+   First choice is to set the first basis unit vector $N$ from the center point of the ellipse to a point on the boundary at $\vartheta=0$ position.
+   Then use a first guess for the second basis unit vector $B$ from the center point to $\theta=\frac{\pi}{2}$ position to span the unit normal of the plane $K=(N \times B)$,
[…]
