### 2026-02-16 6ec5ff7
fix missing colorbar label for plot_on_flux_surface
