+++ python/gvec/plotting/plots2d.py
@@ -354,27 +354,14 @@ def plot_on_flux_surface(
         if isinstance(quantities, list):
             # Specify the quantity type in the top right corner if multiple quantities were requested
             plot_label = f"${evaluations_i[quantity].attrs['symbol']}$"
-            fig.colorbar(f_ax, ax=ax)  # , label=f"${evaluations[quantity].symbol}$")
+            fig.colorbar(f_ax, ax=ax)
         else:
             plot_label = f"$\\rho={evaluations.rho.data[i]}$"
             if not share_contours:
-                fig.colorbar(f_ax, ax=ax)
-
-        # ax.annotate(
-        #     plot_label,
-        #     xy=(1, 1),
-        #     xycoords="axes fraction",
-        #     xytext=(-0.6, -0.6),
-        #     textcoords="offset fontsize",
-        #     verticalalignment="top",
-        #     horizontalalignment="right",
-        #     bbox=dict(facecolor="white", edgecolor="black"),
-        # )
-        ax.set_title(plot_label)
-
-        # zeta_angle = zeta_eval[i] / np.pi
+                fig.colorbar(f_ax, ax=ax, label=f"${evaluations[quantity].symbol}$")
 
         ax.set(
+            title=plot_label,
             xlabel=rf"${zeta_vals.attrs['symbol']} / (2\pi)$",
             ylabel=rf"${theta_vals.attrs['symbol']} / (2\pi)$",
         )
