### 2026-02-16 97c85f0
fix plot_fourier_on_surface
