### 2026-02-16 2a14c70
fix SyntaxWarning, autodocumentation & LaTeX macros
