### 2026-02-17 d75c0aa
more docstring cleanup for proper autodoc output.
