+++ python/gvec/util.py
@@ -73,7 +73,7 @@ class CaseInsensitiveDict(MutableMapping):
     # Adapted from requests.structures.CaseInsensitiveDict
     # See: https://github.com/psf/requests/blob/main/src/requests/structures.py
     # Original license: Apache License 2.0
-    """A dictionary-like Mutable Mapping where string keys are case-insensitive.
+    r"""A dictionary-like Mutable Mapping where string keys are case-insensitive.
 
     Implements all methods and operations of
     ``MutableMapping`` as well as dict's ``copy``. Also
@@ -84,13 +84,13 @@ class CaseInsensitiveDict(MutableMapping):
     ``iter(instance)``, ``keys()``, ``items()``, ``iterkeys()``, and ``iteritems()``
     will contain case-sensitive keys. However, querying and contains
     testing is case insensitive:
-
+    ::
         cid = CaseInsensitiveDict()
         cid['param'] = 'value'
-        cid['Param'] == 'value'  # True
+        cid['Param'] == 'value' # True
 
     If the constructor, ``.update``, or equality comparison
-    operations are given keys that have equal ``.lower()``s, a ValueError is raised.
+    operations are given keys that have equal ``.lower()``, a ``ValueError`` is raised.
     """
 
     def __init__(self, data=(), /, **kwargs):
@@ -178,7 +178,7 @@ class CaseInsensitiveDict(MutableMapping):
 
 def adapt_parameter_file(source: str | Path, target: str | Path, **kwargs):
     """
-    Copy the `source` file to the `target` file and replace the parameters according to `kwargs`.
+    Copy the `source` file to the `target` file and replace the parameters according to ``kwargs``.
 
     Args:
         source (str or Path): The path to the source parameter file.
@@ -190,17 +190,17 @@ def adapt_parameter_file(source: str | Path, target: str | Path, **kwargs):
         AssertionError: If the number of occurrences for any parameter is not exactly 1.
 
     Notes:
-        - If no parameters are provided in `kwargs`, the function simply copies the `source` file to the `target` file.
-        - The function replaces the parameters in the format `key = value`, where value is either a sequence of characters containing
-          no whitespace or a single pair of parentheses with any content. The value from `kwargs` is inserted using the standard python
-          string conversion. There may be a comment, starting with `!`, after the value.
-        - If a parameter already exists in the `source` file, its value is replaced with the corresponding value from `kwargs`.
-        - If a parameter does not exist in the `source` file, it is added to the `target` file.
-        - If the value of the key starts with "!", the line with the keyword is just uncommented.  (i.e. "!key=2.5" -> "key=2.5")
-          If no line with the keyword is found, the key is added with the value, excluding the leading "!"  (i.e. value is "!0.5" -> "key=0.5" is added)
+        - If no parameters are provided in ``kwargs``, the function simply copies the ``source`` file to the ``target`` file.
+        - The function replaces the parameters in the format ``key = value``, where value is either a sequence of characters containing
+          no whitespace or a single pair of parentheses with any content. The value from ``kwargs`` is inserted using the standard python
[…]
