### 2026-02-18 194ac44
fix boozer_kwargs docstring
