### 2026-02-18 cd76c77
fix docstring formatting
