### 2026-02-18 84c0a7d
logger output fix in gframe to surface script.
* distance in field periodicity error was wrongly reported.
* add some more information to logger output.
* add optional input parameter for tolerance on the field-periodicity check.
