### 2026-02-18 gvec-group/gvec#2 @fhindenlang
This issue has been addressed also in the latest documentation on readthedocs
