### 2026-02-19 ef8884a
add kappa_B & kappa_G
