+++ docs/dev/derivations.md
@@ -2,6 +2,34 @@
 
 This page collects some derivations for computable quantities:
 
+Throughout this document we use $\alpha,\beta$ as generic indices for the coordinates $\rho,\theta,\zeta$ unless otherwise specified.
+
+## Flux aligned coordinates
+
+$$
+\begin{align}
+\vec{e}_\alpha &= \pdv{\vec{x}}{\alpha} \\
+\vec{e}^\alpha &= \grad \alpha \\
+\Jac &= \vec{e}_\rho \cdot (\vec{e}_\theta \times \vec{e}_\zeta) \\
+\Jac &= \qty(\grad \rho \cdot \grad \theta \times \grad \zeta)^{-1} \\
+\end{align}
+$$
+
+$$
+\begin{align}
+\vec{e}_\rho &= \Jac \grad\theta \times \grad\zeta \\
+\vec{e}_\theta &= \Jac \grad\zeta \times \grad\rho \\
+\vec{e}_\zeta &= \Jac \grad\rho \times \grad\theta \\
+\end{align}
+$$
+
+$$
+\begin{align}
+\vec{k}_{\alpha\beta} &:= \pdv[2]{\vec{x}}{\alpha}{\beta} \\
+&= \pdv{\vec{e}_\alpha}{\beta} = \pdv{\vec{e}_\beta}{\alpha} \\
+\end{align}
+$$
+
 ## Magnetic field $\vec{B}$
 
 
@@ -39,24 +67,25 @@ $\Rightarrow$ `dmodB_dr`, `dmodB_dt`, `dmodB_dz`, `grad_modB`
 
 ## Derivatives of $\vec{B}$
 
+As $\vec{B}$ is a vector, its derivative $\grad\vec{B}$ is a matrix, which complicates things a bit, as the matrix product is not commutative. We denote the *outer product* with $\otimes$.
+
 $$
 \begin{align}
-\nabla \vec{B} &:= \pdv{\vec{B}}{\vec{x}}
-= \sum_\alpha \pdv{\vec{B}}{\alpha} \nabla\alpha
-= \pdv{\vec{B}}{\rho} \nabla\rho + \pdv{\vec{B}}{\thet} \nabla\thet + \pdv{\vec{B}}{\zeta} \nabla\zeta \\
-\pdv{\vec{B}}{\alpha} &:= \sum_\beta \pdv{B^\beta}{\alpha} \vec{e}_\beta + B^\beta \vec{k}_{\beta\alpha} \\
+\grad \vec{B} &:= \grad \otimes \vec{B} \\
+&= \sum_\alpha \grad\alpha \otimes \pdv{\vec{B}}{\alpha} \\
+&= \grad\rho \otimes \pdv{\vec{B}}{\rho} + \grad\thet \otimes \pdv{\vec{B}}{\thet} + \grad\zeta \otimes \pdv{\vec{B}}{\zeta} \\
+\pdv{\vec{B}}{\alpha} &:= \sum_\beta \pdv{\qty(B^\beta \vec{e}_\beta)}{\alpha} \\
+&= \sum_\beta \pdv{B^\beta}{\alpha} \vec{e}_\beta + B^\beta \vec{k}_{\beta\alpha} \\
 &= \cancel{\pdv{B^\rho}{\alpha}} \erho + \cancel{B^\rho} \vec{k}_{\rho\alpha}
 + \pdv{B^\thet}{\alpha} \ethet + B^\thet \vec{k}_{\thet\alpha}
 + \pdv{B^\zeta}{\alpha} \ezeta + B^\zeta \vec{k}_{\zeta\alpha} \\
[…]
