### 2026-02-23 0b8f315
remove LaTeX requirement
