### 2026-02-25 f1f72ff
add parameter extraction for vmec wout files
