### 2026-02-25 fa528c5
update version switcher
