+++ python/gvec/scripts/convert_wout.py
@@ -1,0 +1,223 @@
+# Copyright (c) 2026 GVEC Contributors, Max Planck Institute for Plasma Physics
+# License: MIT
+"""convert-wout.py - convert a VMEC wout file to a GVEC parameterfile & state"""
+
+from pathlib import Path
+import logging
+import argparse
+from collections.abc import Sequence, Mapping
+import tempfile
+import shutil
+import re
+
+import numpy as np
+import xarray as xr
+
+import gvec
+from gvec.util import CaseInsensitiveDict
+
+parser = argparse.ArgumentParser(
+    prog="pygvec-convert-wout",
+    formatter_class=argparse.RawDescriptionHelpFormatter,
+    description="Convert a VMEC wout file to a GVEC parameterfile & state.",
+)
+parser.add_argument(
+    "wout_file",
+    type=Path,
+    help="input VMEC wout file",
+)
+parser.add_argument(
+    "rundir",
+    type=Path,
+    default=".",
+    nargs="?",
+    help="optional output directory for GVEC state and parameter files (default: current directory)",
+)
+wout_format = parser.add_argument_group(title="wout format").add_mutually_exclusive_group()
+wout_format.add_argument(
+    "--nemec-ascii",
+    action="store_const",
+    const=1,
+    dest="VMECwoutfile_format",
+    help="input wout file is in nemec-ascii format (default: netCDF)",
+)
+wout_format.add_argument(
+    "--nemec-binary",
+    action="store_const",
+    const=2,
+    dest="VMECwoutfile_format",
+    help="input wout file is in nemec-binary format (default: netCDF)",
+)
+parser.add_argument(
+    "--param",
+    type=str,
+    nargs="*",
+    default=[],
+    help="extra parameter to be used in the conversion, e.g. --param=init_LA=True or --param=sgrid_nElems=20",
+)
+verbosity = parser.add_mutually_exclusive_group()
+verbosity.add_argument(
+    "-v",
+    "--verbose",
+    action="count",
+    default=0,
+    help="verbosity level: -v for info, -vv for debug",
+)
+verbosity.add_argument("-q", "--quiet", action="store_true", help="suppress output")
+
+logger = logging.getLogger(__name__)
+
+
+def convert_vmec_wout(
+    wout_file: Path,
+    rundir: Path | None = None,
+    *,
+    VMECwoutfile_format: int = 0,
+    extra_parameters: Mapping = {},
+    convdir: Path | None = None,
+) -> CaseInsensitiveDict:
+    """
+    Convert a VMEC wout file to a GVEC parameter & statefile.
+
+    The boundary and profiles are extracted, to generate a GVEC parameter file corresponding to the same configuration,
+    as well as a GVEC parameter and statefile describing the (interpolated) VMEC state.
+
+    Parameters
+    ----------
+    wout_file
+        Path to the VMEC wout file in netCDF format.
+    rundir
+        Optional path to a directory in which the GVEC state should be generated.
+    VMECwoutfile_format
+        The format of the VMEC wout file (0 for netCDF, 1 for nemec-ascii, 2 for nemec-binary).
+    extra_parameters
+        Extra parameters to be used in the conversion, e.g. ``init_LA`` or ``sgrid_nElems``.
+    """
+    if not Path(wout_file).exists():
+        raise FileNotFoundError(f"wout file '{wout_file}' not found")
+    wout_ds = xr.open_dataset(wout_file)
+
+    M = int(wout_ds.xm.max())
+    N = int(wout_ds.xn.max() / wout_ds.nfp)
+    nfp = int(wout_ds.nfp)
+    lasym = bool(wout_ds.lasym__logical__)  # stellarator asymmetric
[…]
