### 2026-03-03 c41bbab
fix gframe.to_RZ docstring
