### 2026-03-03 #9670 @piyueh
Hi @AnkitBarik , yes, the software documentation (e.g., README, website, or any end-user facing documents, etc.) should have a statement of need.

Reference policy: https://joss.readthedocs.io/en/latest/review_criteria.html#documentation
