### 2026-03-03 #9670 @AnkitBarik
@fhindenlang Also apologies for my delay. I don't see a statement of need on the README or [documentation](https://gvec.readthedocs.io/) (as needed by the latest review criteria). Could you add a few sentences to both? (or point me to the section if I missed it?)
