### 2026-03-04 78e87bb
add keys argument to table_of_quantities
