### 2026-03-04 baca777
two functions for writhe. test gframe files. update docstrings.
* renamed `utils.writhe` with `utils.writhe_from_polygon`, since its a very approximate way to compute writhe due to the polygon approximation, but needs only the curve itself.
* added `gvec.gframe.writhe`, which accurately computes the writhe of the curve by attaching a ribbon, and computing Wr=(Lk-Tw) from the difference of the linking number of the ribbon and its twist.
  If only the curve is given, we build the ribbon via a centroid frame. Validity of the ribbon is checked in the twist computation.
* add tests for all gframe f
[…]
