+++ python/gvec/gframe.py
@@ -19,8 +19,8 @@ from gvec import fourier
 
 def check_field_periodicity(xyz: np.ndarray, nfp: int, atol=1e-12):
     """
-    checks if all  xyz positions of the surface on a full turn, xyz[0:nz*nfp,0:nt,0:2], have the field periodicity with nfp.
-    returns the sign of the rotation +2pi/nfp or -2pi/nfp
+    checks if all  xyz positions of the surface on a full turn, ``xyz[0:nz*nfp,0:nt,0:2]``, have the field periodicity with ``nfp``.
+    returns the sign of the rotation ``+2pi/nfp`` or ``-2pi/nfp``
     """
     assert xyz.shape[-1] == 3, (
         "last dimension must be the cartesian components surface positions!"
@@ -822,10 +822,9 @@ def construct_gframe_from_surface(
 
     xyz0, N, B = get_X0_N_B(xyz_gframe)
 
-    Lk = linking_number(xyz0, xyz0 + 1e-3 * N)
-    Tw = twist_of_ribbon(xyz0, 1e-3 * N)
+    Wr, Lk, Tw = writhe(xyz0, N)
     logger.info(
-        f"  - G-frame linking number = {Lk:.0f}, twist = {Tw:.3f}, writhe = (Lk-Tw)= {Lk - Tw:.3f}"
+        f"  - G-frame linking number = {Lk:.0f}, twist = {Tw:.3f}, writhe = (Lk-Tw)= {Wr:.3f}"
     )
 
     logger.info(". Cutting the surface")
@@ -987,6 +986,49 @@ def minimal_modes(X, Y, Z=None, tolerance=1e-8):
     mask &= error == error[mask].min()
 
     return mcan[mask].item(), ncan[mask].item()
+
+
+def to_axis(dict_in: dict, nzeta: int = 81):
+    """
+    Convert the "axis" of a gframe file to a different resolution.
+
+    Parameters
+    ----------
+    dict_in : dict
+        dictionary of the Gframe netcdf file, from `gvec.gframe.read_Gframe_ncfile(filename)`
+    nzeta : int, optional
+        number of zeta positions for the output, to sample on one field period (default: 81)
+
+    Returns
+    -------
+    dict_axis : dict
+        dictionary containing
+
+        - ``nfp``: number of field periods
+        - ``axis``: dictionary for the G-Frame data:
+
+            - ``nzeta``: number of zeta positions for one field-period
+            - ``zetafull``: equidistant positions along the curve parameter zeta, without endpoint, length is ``nfp*nzeta``
+            - ``xyz``: cartesian positions along the curve, shape is ``(3,nfp*nzeta)``
+            - ``Nxyz``: cartesian components of the 'normal' vector, shape is ``(3,nfp*nzeta)``
+            - ``Bxyz``: cartesian components of the 'bi-normal' vector, shape is ``(3,nfp*nzeta)``
+    """
+    nfp = dict_in["nfp"]
+    zetafull_out = np.linspace(0, 2 * np.pi, nzeta * nfp, endpoint=False)
+    zdft = fourier.real_dft_mat(dict_in["axis"]["zetafull"], zetafull_out, nfp=1)
+    xyz = zdft["BF"] @ dict_in["axis"]["xyz"].T  # [0:nz*nfp,0:2]
+    Nxyz = zdft["BF"] @ dict_in["axis"]["Nxyz"].T
+    Bxyz = zdft["BF"] @ dict_in["axis"]["Bxyz"].T
+    return {
+        "nfp": nfp,
+        "axis": {
+            "nzeta": nzeta,
+            "zetafull": zetafull_out,
+            "xyz": xyz.T,
+            "Nxyz": Nxyz.T,
+            "Bxyz": Bxyz.T,
+        },
+    }
 
 
[…]
