+++ test-CI/unit-pygvec/test_gframe.py
@@ -1,4 +1,5 @@
 import pytest
+from pathlib import Path
 
 try:
     import numpy as np
@@ -9,6 +10,8 @@ except ImportError:
     # therefore we skip the whole module
     pytest.skip("Skipping test_gframe.py: gvec package not available", allow_module_level=True)
 
+
+DATA = Path(__file__).parent / "../data"
 
 # === Fixtures === #
 
@@ -120,7 +123,120 @@ def test_writhe_boundary(case, nfp):
     # "exact" writhe from linking - twist
     Tw = gvec.gframe.twist_of_ribbon(C_a, C_b - C_a)
     # approximate writhe
-    Wr = gvec.util.writhe(C_a, endpoint=False)
-    assert np.abs(Wr - (Lk - Tw)) < 1e-3, (
-        f"Compare approx. writhe {Wr:5.2e} with (Lk - Tw). Lk = {Lk:.0f} Tw={Tw:5.2e} ... Difference: {Wr - (Lk - Tw):5.2e}"
+    Wr_polygon = gvec.util.writhe_from_polygon(C_a, endpoint=False)
+    assert np.abs(Wr_polygon - (Lk - Tw)) < 1e-3, (
+        f"Compare writhe from polygon {Wr_polygon:5.2e} with writhe from riboon, (Lk - Tw). Lk = {Lk:.0f} Tw={Tw:5.2e} ... Difference: {Wr_polygon - (Lk - Tw):5.2e}"
+    )
+
+
+@pytest.mark.parametrize(
+    "gframe_ncfile",
+    [
+        "2022_QH_nfp7.nc",
+        "brezellator-Gframe.nc",
+        "brezellator-stellsym-scaled-Gframe.nc",
+        "N2.12-v3.1-hi.nc",
+        "N2.12-v3.1-lo.nc",
+        "N2-63.nc",
+        "N3.471-v3.1-lo.nc",
+        "N3.Knot-v3.1-lo.nc",
+    ],
+)
+def test_gframe_files_writhe(gframe_ncfile):
+    gframe_ncfile = DATA / gframe_ncfile
+    dict_gframe = gvec.gframe.read_Gframe_ncfile(gframe_ncfile)
+    assert "axis" in dict_gframe and "boundary" in dict_gframe, (
+        f"Expected 'axis' and 'boundary' in gframe dict, but not found. Keys: {dict_gframe.keys()}"
+    )
+    dict_highres = gvec.gframe.to_axis(dict_gframe, nzeta=201)
+    xyz = dict_highres["axis"]["xyz"].T
+    Nxyz = dict_highres["axis"]["Nxyz"].T
+    Wr_polygon = gvec.util.writhe_from_polygon(xyz)
+    Wr_gframe = gvec.gframe.writhe(xyz, N=Nxyz)[0]
+    assert np.isclose(Wr_polygon, Wr_gframe, atol=1e-3), (
+        f"Writhe from polygon {Wr_polygon} and from gframe {Wr_gframe} differ by more than 1e-3 for file {gframe_ncfile}"
+    )
+
+
+def trefoil(t, scale=3.0, zscale=1.0):
+    x = np.sin(t) + 2 * np.sin(2 * t)
+    y = np.cos(t) - 2 * np.cos(2 * t)
+    z = -np.sin(3 * t) * zscale
+    X0 = np.stack((x, y, z), axis=1)
+
+    xp = np.cos(t) + 4 * np.cos(2 * t)
+    yp = -np.sin(t) + 4 * np.sin(2 * t)
+    zp = -3 * np.cos(3 * t) * zscale
+    X0p = np.stack((xp, yp, zp), axis=1)
+
[…]
