+++ docs/api/coils.rst
@@ -1,7 +1,7 @@
 .. _api-coils:
 
 gvec.coils
-========
+==========
 
 .. automodule:: gvec.coils
   :members:
