### 2026-03-05 e8b5976
use dft for frenet.
* numpy fft and ifft cannot handle even number arrays, which is necessary to keep if field periodicity is even. So changed fft1d in frenet_frame to the 1d dft from fourier module.
* dft allows to "downsample", so less points as the original curve.
* twist computation still with fft, much faster for high number of integration points.
* twist with centroid frame less accurate, if "normal" vector is normalized to unit length. Avoid normalization for now, then N has the same spectrum as the curve, and the twist is more accurate (tested in unit test).
