+++ python/gvec/gframe.py
@@ -1304,7 +1304,7 @@ def twist_of_ribbon(X0: np.ndarray, N: np.ndarray, nint: int = None) -> float:
     N  : np.ndarray
         linearly independent, non-normalized "normal" vector at curve positions, in cartesian coordinates, same shape as ``X0``
     nint : int, optional
-        number of points for integration, None sets default ``=npoints``
+        number of points for integration, ``None`` sets default ``=npoints``
 
     Returns
     -------
@@ -1366,7 +1366,7 @@ def writhe(X0: np.ndarray, N: np.ndarray = None, width: float = 1e-6, nint: int 
     X0 : np.ndarray
         positions along closed curve, in cartesian coordinates, shape ``(npoints,3)``,  excluding periodic endpoint!
     N  : np.ndarray, optional
-        linearly independent, non-normalized "normal" vector at curve positions, in cartesian coordinates, same shape as ``X0``. If   not provided, `N` is computed from the centroid frame of ``X0``
+        linearly independent, non-normalized "normal" vector at curve positions, in cartesian coordinates, shape ``(npoints,3)``. If   not provided, `N` is computed from the centroid frame of ``X0``
     width : float, optional
         In order to compute the linking number, the width of the ribbon must be specified, generating a second curve `X0+N*width`. Default width is `1e-6`
     nint : int, optional
@@ -1382,18 +1382,21 @@ def writhe(X0: np.ndarray, N: np.ndarray = None, width: float = 1e-6, nint: int 
         twist of the ribbon defined by ``X0`` and ``N``
     """
     if N is None:
-        N = X0 - np.mean(X0, axis=0)  # centroid frame
+        # use centroid frame
+        N = X0 - np.mean(X0, axis=0)
+        # normalize to max length 1 (not unit length, can be represented with same resolution as X0)
+        N /= np.amax(np.sqrt(np.sum(N**2, axis=-1)))
     Tw = twist_of_ribbon(X0, N, nint=nint)
     Lk = linking_number(X0, X0 + N * width)
     Wr = Lk - Tw
     return Wr, Lk, Tw
 
 
-def frenet_frame(X0: np.ndarray) -> dict:
+def frenet_frame(X0: np.ndarray, nzeta: int = None) -> dict:
     r"""
     Compute the Frenet frame of a closed (!) curve defined by $X_0(\zeta)$, by computing the tangent vector $T$, normal vector $N$ and binormal vector $B$.
 
-    Derivatives of $X_0$ are computed via fft, so the curve is assumed to be given on an equispaced grid in ``zeta=zeta0+np.linspace(0,2*np.pi,npoints,endpoint=False)``, excluding the endpoint.
+    Derivatives of $X_0$ are computed via dft, so the curve is assumed to be given on an equispaced grid in ``zeta=zeta0+np.linspace(0,2*np.pi,npoints,endpoint=False)``, excluding the endpoint.
 
     Warnings
     --------
@@ -1405,6 +1408,8 @@ def frenet_frame(X0: np.ndarray) -> dict:
     ----------
     X0 : np.ndarray
         closed curve positions, in cartesian coordinates, shape ``(npoints,3)``,  excluding periodic endpoint!
+    nzeta : int, optional
+        number of points for the output, default is None, which then uses ``npoints``.
 
     Returns
[…]
