### 2026-03-05 0f1f24d
simplify dmod_B_da & db_da
