+++ python/gvec/quantities.py
@@ -808,32 +808,19 @@ def _dmod_B_factory(a):
     @register(
         quantities=[f"dmod_B_d{a}"],
         requirements=[
+            "B",
             "mod_B",
-            "B_contra_t",
-            "B_contra_z",
-            f"dB_contra_t_d{a}",
-            f"dB_contra_z_d{a}",
-            "g_tt",
-            "g_tz",
-            "g_zz",
-        ]
-        + [f"dg_{ij}_d{a}" for ij in ["tt", "tz", "zz"]],
+            f"dB_d{a}",
+        ],
         attrs=dict(
             long_name=derivative_name_smart("modulus of the magnetic field", a),
             symbol=latex_partial(r"\left|\mathbf{B}\right|", a),
         ),
     )
     def _dmod_B(ds: xr.Dataset):
-        dmod_B2_da = (
-            2 * ds[f"dB_contra_t_d{a}"] * ds.B_contra_t * ds.g_tt
-            + ds.B_contra_t**2 * ds[f"dg_tt_d{a}"]
-            + 2 * ds[f"dB_contra_z_d{a}"] * ds.B_contra_t * ds.g_tz
-            + 2 * ds[f"dB_contra_t_d{a}"] * ds.B_contra_z * ds.g_tz
-            + 2 * ds.B_contra_t * ds.B_contra_z * ds[f"dg_tz_d{a}"]
-            + 2 * ds[f"dB_contra_z_d{a}"] * ds.B_contra_z * ds.g_zz
-            + ds.B_contra_z**2 * ds[f"dg_zz_d{a}"]
-        )
-        ds[f"dmod_B_d{a}"] = dmod_B2_da / (2 * ds.mod_B)
+        ds[f"dmod_B_d{a}"] = xr.dot(ds[f"dB_d{a}"], ds.B, dim="xyz") / ds.mod_B
+
+    return _dmod_B
 
 
 # generate functions from factory function
@@ -859,7 +846,6 @@ def _db_factory(a):
         quantities=[f"db_d{a}"],
         requirements=[
             f"dB_d{a}",
-            f"dmod_B_d{a}",
             "mod_B",
         ],
         attrs=dict(
@@ -868,7 +854,8 @@ def _db_factory(a):
         ),
     )
     def _db(ds: xr.Dataset):
-        ds[f"db_d{a}"] = ds[f"dB_d{a}"] / ds.mod_B - ds.B * ds[f"dmod_B_d{a}"] / ds.mod_B**2
+        b = ds.B / ds.mod_B
+        ds[f"db_d{a}"] = (ds[f"dB_d{a}"] - b * xr.dot(ds[f"dB_d{a}"], b, dim="xyz")) / ds.mod_B
 
     return _db
 
