### 2026-03-05 74509ea
cleanup assert statements.
* in functions, replace `assert (x), msg ` on input parameters (like shapes) by `if not (x): raise ValueError(msg)`
* in tests for all close, replace `assert np.allclose(a,b), msg` by `np.testing.assert_allclose(a,b, err_msg=msg)`.
  But the default `rtol,atol` parameters change between both... so setting a dict `default_rtol_atol` that keeps `np.allclose` defaults `rtol=1e-5,atol=1e-8`
