### 2026-03-09 bc92b1e
add statement of need to README & docs
