### 2026-03-09 b210607
fix bibliography
