+++ JOSS_paper/paper.bib
@@ -223,7 +223,7 @@ keywords = {Plasma simulation, Finite element exterior calculus, Particle-in-cel
 %JOREK3D
 @article{nikulsin2022jorek3d,
   title={{JOREK3D}: An extension of the {JOREK} nonlinear {MHD} code to stellarators},
-  author={Nikulsin, Nikita and Ramasamy, Rohan and Hoelzl, Matthias and Hindenlang, Florian and Strumberger, Erika and Lackner, Karl and G{\"u}nter, Sibylle and Jorek Team and others},
+  author={Nikulsin, Nikita and Ramasamy, Rohan and Hoelzl, Matthias and Hindenlang, Florian and Strumberger, Erika and Lackner, Karl and G{\"u}nter, Sibylle and {Jorek Team} and others},
   journal = {Physics of Plasmas},
   volume = {29},
   number = {6},
@@ -265,7 +265,7 @@ keywords = {Plasma simulation, Finite element exterior calculus, Particle-in-cel
 %CASTOR3D
 @article{puchmayr2023helical,
   title = {Helical mode localization and mode locking of ideal MHD instabilities in magnetically perturbed tokamak plasmas},
-  author = {Puchmayr, Jonas and Dunne, Mike and Strumberger, Erika and Willensdorfer, Matthias and Zohm, Hartmut and Hindenlang, Florian and the ASDEX Upgrade Team},
+  author = {Puchmayr, Jonas and Dunne, Mike and Strumberger, Erika and Willensdorfer, Matthias and Zohm, Hartmut and Hindenlang, Florian and {the ASDEX Upgrade Team}},
   journal={Nuclear Fusion},
   year={2023},
   doi = {10.1088/1741-4326/acdd12},
@@ -378,10 +378,9 @@ keywords = {Plasma simulation, Finite element exterior calculus, Particle-in-cel
 #stellopt
 @misc{doecode_12551,
   title = {STELLOPT},
-  author = {Lazerson, Samuel and Schmitt, John and Zhu, Caoxiang and Breslau, Joshua and STELLOPT Developers, All},
+  author = {Lazerson, Samuel and Schmitt, John and Zhu, Caoxiang and Breslau, Joshua and {All STELLOPT Developers}},
   abstractNote = {The STELLOPT code is designed to optimize 3D MHD equilibria to a set of target physics parameters encompassing stellarator design and 3D equilibrium reconstructions.},
   doi = {10.11578/dc.20180627.6},
-  howpublished = {[Computer Software] \url{https://doi.org/10.11578/dc.20180627.6}},
   year = {2020},
   month = {may}
 }
