### 2026-03-09 #9670 @Rykath
> [@fhindenlang](https://github.com/fhindenlang) Also apologies for my delay. I don't see a statement of need on the README or [documentation](https://gvec.readthedocs.io/) (as needed by the latest review criteria). Could you add a few sentences to both? (or point me to the section if I missed it?)

Hi @AnkitBarik, thank you for your comment. I have now added a statement of need section to both the [README](https://github.com/gvec-group/gvec/tree/v1.3.x) and landing page of the [documentation](https://gvec.readthedocs.io/v1.3.x/index.html). Does this meet the requirements?
