### 2026-03-09 #9670 @AnkitBarik
Hi @Rykath , thank you! Yes that works.

@piyueh I have now checked off all the boxes, and I am happy to recommend this paper for publication in JOSS. Let me know if you need anything else from me. Cheers, Ankit
