### 2026-03-16 #9670 @piyueh
@fhindenlang Could you follow the author task list described [here](https://github.com/openjournals/joss-reviews/issues/9670#issuecomment-4068992865) and report the DOI and the final version number? Let me know if you have any questions. Thanks.
