### 2026-03-18 cdfd5fc
set init_average_axis for load_quasr
