### 2026-03-19 #9670 @fhindenlang
Hi @piyueh

- We have now released the version of GVEC that corresponds to the one reviewed here (branch `v1.3.x`).  We added the files for producing the JOSS paper. The version tag is [`v1.3.2`](https://gitlab.mpcdf.mpg.de/gvec-group/gvec/-/releases/v1.3.2)  which is also available on [PyPI](https://pypi.org/project/gvec/1.3.2/).
- We have also uploaded the version to Zenodo, with the same author list in the same order as in the JOSS paper.
  the corresponding DOI: [`10.5281/zenodo.19076836`](https://doi.org/10.5281/zenodo.19076836). We have checked the ORCIDs and affiliations.
- the license in zenodo is the same as in the repository (MIT license)
