### 2026-03-19 4b0dd0d
fix kappa_G
