### 2026-03-19 ab45d97
add dmod_B_da_P
