+++ python/gvec/quantities.py
@@ -1487,6 +1487,66 @@ def dmod_B_dz_B(ds: xr.Dataset):
     ds["dmod_B_dz_B"] = dt_dzB * ds.dmod_B_dt + dz_dzB * ds.dmod_B_dz
 
 
+@register(
+    requirements=(
+        "dLA_dr",
+        "dLA_dt",
+        "dmod_B_dr",
+        "dmod_B_dt",
+    ),
+    attrs=dict(
+        long_name="radial PEST-like derivative of the modulus of the magnetic field",
+        symbol=r"\frac{\partial\left|\mathbf{B}\right|}{\partial \rho_P}",
+    ),
+)
+def dmod_B_dr_P(ds: xr.Dataset):
+    dtP_dr = ds.dLA_dr
+    dtP_dt = 1 + ds.dLA_dt
+    # dr_drP = 1
+    dt_drP = -dtP_dr / dtP_dt
+    # dz_drP = 0
+    ds["dmod_B_dr_P"] = ds.dmod_B_dr + dt_drP * ds.dmod_B_dt
+
+
+@register(
+    requirements=(
+        "dLA_dt",
+        "dmod_B_dt",
+    ),
+    attrs=dict(
+        long_name="poloidal PEST-like derivative of the modulus of the magnetic field",
+        symbol=r"\frac{\partial\left|\mathbf{B}\right|}{\partial \vartheta_P}",
+    ),
+)
+def dmod_B_dt_P(ds: xr.Dataset):
+    dtP_dt = 1 + ds.dLA_dt
+    # dr_dtP = 0
+    # dt_dtP = 1 / dtP_dt
+    # dz_dtP = 0
+    ds["dmod_B_dt_P"] = 1 / dtP_dt * ds.dmod_B_dt
+
+
+@register(
+    requirements=(
+        "dLA_dt",
+        "dLA_dz",
+        "dmod_B_dt",
+        "dmod_B_dz",
+    ),
+    attrs=dict(
+        long_name="toroidal PEST-like derivative of the modulus of the magnetic field",
+        symbol=r"\frac{\partial\left|\mathbf{B}\right|}{\partial \zeta_P}",
+    ),
+)
+def dmod_B_dz_P(ds: xr.Dataset):
+    dtP_dz = ds.dLA_dz
+    dtP_dt = 1 + ds.dLA_dt
+    # dr_dzP = 0
+    dt_dzP = -dtP_dz / dtP_dt
+    # dz_dzP = 1
+    ds["dmod_B_dz_P"] = ds.dmod_B_dz + dt_dzP * ds.dmod_B_dt
+
+
 # === integrals ======================================================================== #
 # --- geometry --- #
 
