### 2026-03-23 #9670 @piyueh
Hi @fhindenlang here are some minor issues need you to address:

### Name convention

Köberl (paper) v.s. Koeberl (Zenodo)

### Grammar issues in the paper

- line 5: Max _**Plank**_ -> Max _**Planck**_
- line 20: high fidelity -> high-fidelity
- line 45: a tensor-product -> a tensor product
- line 51: a **_python_** package -> a **_Python_** package
- line 52: the **_python_** package -> the **_Python_** package
- line 53: **_python_** API -> **_Python_** API
- line 82: We also like to ... -> We **_would_** also like to ...
- line 88: **_Parts_** of this work **_has_** been ... -> **_Parts_** of this work **_have_** been ... or **_Part_** of this work **_has_** been ...
- line 132: python, fortran -> Python, Fortran

These issues are only related to the paper and the Zenodo record, so it is not necessary to make a new software release, and I'm setting the JOSS version to v1.3.2. (If you still make a new release later to reflect these paper changes, please post the new version tag and new Zenodo DOI here.)
