### 2026-03-23 #9670 @fhindenlang
Hi @piyueh
I have made the requested changes to the paper and bib file, and pushed them to the latest branch `joss_paper_2025` and `v1.3.x`. I updated the name to `Robert Köberl` in the zenodo.
Before making a new zenodo entry, would you please check the latest version of the paper?
