### 2026-03-23 #9670 @piyueh
@fhindenlang Thanks! I think the rendered PDF file was compiled from the branch `joss_paper_2025`, based on the compilation log ([here](https://github.com/openjournals/joss-papers/actions/runs/23457964577/job/68252083867#logs), if you're interested). The latest Zenodo record (https://zenodo.org/records/19076836) looks good to me.

I am now moving this submission to the next stage. The Editor-in-Chief will take over from here, make a final pass of the paper, and proceed.
