### 2026-03-23 #9670 @piyueh
oh, sorry I missed the part where you said "Before making a new zenodo entry." Are you planning to make a new release for these paper editing? If so, I can wait until you make it. If not, I'll proceed with v1.3.2
