### 2026-03-23 #9670 @fhindenlang
Hi @piyueh,
yes as you suggested, I would like to make a new zenodo entry, while keeping the code release on version v1.3.2. That new zenodo would only have the latest  changes in the JOSS paper updated, the rest of the code remains the same as tagged release v1.3.2. What do you think?
