### 2026-03-23 #9670 @piyueh
@fhindenlang Will creating a new Zenodo entry generate a new DOI? If not, and it uses the same DOI, then I think it is fine to create a new Zenodo entry now. However, if it does generate a new DOI, I would suggest waiting until the Editor-in-Chief has finished the final paper review, since they sometimes request additional edits.

For your information and future reference, many authors do not include the paper in their source code, and also they only archive the source code on Zenodo. As a result, they do not need to update the Zenodo entry after paper editing. (However, it is fine if you want to keep the paper with the source code and Zenodo. JOSS has no policy on this.)

Let me know if you would still like to create a new Zenodo entry now.
