### 2026-03-23 #9670 @piyueh
@fhindenlang I think I'll proceed to the Editor-in-Chief now. We can always update Zenodo info before publishing. I will let the Editor-in-Chief know you would like to create a new Zenodo for the final version before publishing.
