### 2026-03-23 #9670 @editorialbot
:wave: @openjournals/pe-eics, this paper is ready to be accepted and published.
