### 2026-03-23 #9670 @piyueh
@openjournals/pe-eics The author of this submission would like to create a new Zenodo entry for the final version of the paper, as they also archive `paper.md` (and the `.bib` file) on Zenodo. Could you please notify the authors when `paper.md` is finalized and no further edits are needed, so they can update Zenodo accordingly? Thanks!
