### 2026-03-25 e64592f
fix gvec.run LogError
