### 2026-03-25 3fc26c6
fix minimize_tol & totalIter defaults
