### 2026-03-25 2925cca
allow_failure for viper/raven CI env jobs
