### 2026-03-25 85211ab
fix CI: upgrade pip
