+++ CI_templates/scripts.yml
@@ -103,6 +103,7 @@
     - source venv_${BUILDNAME}/bin/activate
     - which python
     - which pip
+    - pip install --upgrade pip
     - pip list
     - start_details printenv "Environment Variables"
     - printenv
