### 2026-03-25 479bec9
try fix PyPI CI job
by default, a job loads all the artifacts from all previous jobs
