### 2026-03-25 ff633b8
write suppressed tracebacks to temporary file
