### 2026-03-31 #9670 @kyleniemeyer
@fhindenlang I just have a small suggested edit for the paper before finalizing. In the Statement of Need, the nested parentheses around the example codes and citations are not ideal. Could you change to something like
> For example, they provide the initial conditions for linear and nonlinear
MHD solvers like CASTOR3D (Puchmayr et al., 2023), CAS3D (Schwab, 1993), Jorek3D
(Nikulsin et al., 2022), Struphy (Holderied et al., 2021), M3D-C1 (Jardin, 2004); the
magnetic field for particle orbit tracing as in SIMPLE (Albert et al., 2020); and turbulence
simulations such as BOUT++ (Shanahan et al., 2024), GENE (Bañón Navarro et al., 2020).

(or similar)
