### 2026-03-31 #9670 @fhindenlang
Thank you @kyleniemeyer for your suggestion, I changed it similar to what you suggested.
