### 2026-03-31 #9670 @fhindenlang
@kyleniemeyer:
If no further changes are necessary to paper.md, I would put the source code of gvec and the paper into a zenodo archive and then post its DOI here.
