### 2026-04-02 #9670 @fhindenlang
That sounds great @kyleniemeyer, here the DOI of the zenodo archive containing the paper and GVEC version 1.3.2 :
[10.5281/zenodo.19194335](https://doi.org/10.5281/zenodo.19194335)
